% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{ac_win}
\alias{ac_win}
\title{Windowed autocorrelation function}
\usage{
ac_win(
  df,
  win = NROW(df),
  doPlot = FALSE,
  useVarNames = TRUE,
  colOrder = TRUE,
  useTimeVector = NA,
  timeStamp = "01-01-1999"
)
}
\arguments{
\item{df}{A data frame containing multivariate time series data from 1 person. Rows should indicate time, columns should indicate the time series variables. All time series in \code{df} should be on the same scale, an error will be thrown if the range of the time series in\code{df} is not \verb{[scale_min,scale_max]}.}

\item{win}{Size of window in which to calculate Dynamic Complexity. If \code{win < NROW(df)} the window will move along the time series with a stepsize of \code{1} (default = \code{NROW(df)})}

\item{doPlot}{If \code{TRUE} shows a Complexity Resonance Diagram of the Dynamic Complexity and returns an invisible \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object. (default = \code{FALSE})}

\item{useVarNames}{Use the column names of \code{df} as variable names in the Complexity Resonance Diagram (default = \code{TRUE})}

\item{colOrder}{If \code{TRUE}, the order of the columns in \code{df} determines the of variables on the y-axis. Use \code{FALSE} for alphabetic/numeric order. Use \code{NA} to sort by by mean value of Dynamic Complexity (default = \code{TRUE})}

\item{useTimeVector}{Parameter used for plotting. A vector of length \code{NROW(df)}, containing date/time information (default = \code{NA})}

\item{timeStamp}{If \code{useTimeVector} is not \code{NA}, a character string that can be passed to \code{\link[lubridate:stamp]{lubridate::stamp()}} to format the the dates/times passed in \code{useTimeVector} (default = \code{"01-01-1999"})}
}
\value{
Data frame with autocorrelations in requested window size.
}
\description{
Calculate the autocorrelation function in a right-aligned sliding window on (multivariate) time series data.
}
\note{
For different step-sizes or window alignments see \code{\link[=ts_windower]{ts_windower()}}.
}
\examples{


data(ColouredNoise)
ac_win(elascer(ColouredNoise[,c(1,11,21,31,41)],groupwise = TRUE), win = 128, doPlot = TRUE)

}
