% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{add_alpha}
\alias{add_alpha}
\title{Add transparency to a colour}
\usage{
add_alpha(col, alpha = 1)
}
\arguments{
\item{col}{A colour name, hexadecimal string or positive integer \code{i}, such that \code{palette()[i]}}

\item{alpha}{Alpha transparency value}
}
\value{
An rgb colour with transparency
}
\description{
Add transparency to a colour
}
