% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{aggregate_bins}
\alias{aggregate_bins}
\title{aggregate_bins}
\usage{
aggregate_bins(x, by, FUN, moving = FALSE, ...)
}
\arguments{
\item{x}{x}

\item{by}{by}

\item{FUN}{FUNction}

\item{moving}{moving}

\item{...}{additional}
}
\value{
output
}
\description{
Adaptation from discontinued package \code{ifultools} by William Constantine.
}
\keyword{internal}
