% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{aggregateData}
\alias{aggregateData}
\title{aggregateData}
\usage{
aggregateData(x, by, FUN, moving = FALSE, ...)
}
\arguments{
\item{x}{x}

\item{by}{by}

\item{FUN}{FUNction}

\item{moving}{moving}

\item{...}{additional}
}
\value{
output
}
\description{
aggregateData
}
\note{
Based on "aggregateData" from discontinued package ifultools
}
\keyword{internal}
