% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{as.numeric_character}
\alias{as.numeric_character}
\title{Convert character vector to a named numeric vector}
\usage{
as.numeric_character(x, sort.unique = FALSE)
}
\arguments{
\item{x}{A character vector}

\item{sort.unique}{Should the unique character values be sorted? (\code{default = FALSE})}
}
\value{
A named numeric vector
}
\description{
Convert character vector to a named numeric vector
}
