% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{as.numeric_discrete}
\alias{as.numeric_discrete}
\title{Discrete (factor or character) to numeric vector}
\usage{
as.numeric_discrete(x, keepNA = FALSE)
}
\arguments{
\item{x}{A factor with levels that are numeric, or, a character vector representing numbers.}

\item{keepNA}{Keep NA values (\code{TRUE}), or remove them (default = \code{FALSE})}
}
\value{
A numeric vector with factor levels / numeric character values as names.
}
\description{
Converts a factor with numeric levels, or, a character vector with numeric values to a numeric vector using \link{as.numeric_factor}, or, \link{as.numeric_character} respectively. If an unnamed numeric vector is passed, it will be returned as a named numeric vector.
}
\examples{

f <- factor(round(runif(10,0,9)))
as.numeric_factor(f)

# Add NAs
f <- factor(c(round(runif(9,0,9)),NA))
as.numeric_factor(f)
as.numeric_factor(f, keepNA = TRUE)



}
