% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{as.numeric_factor}
\alias{as.numeric_factor}
\title{Convert numeric factor to numeric vector}
\usage{
as.numeric_factor(x)
}
\arguments{
\item{x}{A factor based on numeric values.}
}
\value{
A numeric vector with factor levels as names.
}
\description{
Convert numeric factor to numeric vector
}
