% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{bandReplace}
\alias{bandReplace}
\title{bandReplace}
\usage{
bandReplace(mat, lower, upper, value = NA)
}
\arguments{
\item{mat}{A Matrix}

\item{lower}{Lower diagonal to be included in the band (should be $\leq0$)}

\item{upper}{Upper diagonal to be included in the band (should be $\geq0$)}

\item{value}{A single value to replace all values in the selected band}
}
\value{
A matrix in which the values in the selected diagonals have been replaced
}
\description{
Sets a band of matrix diagonals to any given value
}
\examples{
# Create a 10 by 10 matrix
m <- Matrix(rnorm(10),10,10)

bandReplace(m,-1,1,0)   # Replace diagonal and adjacent bands
recmat_size(m,FALSE,0)  # Do not subtract diagonal
recmat_size(m,NULL,0)   # Matrix is symmetrical, AUTO is set to TRUE
recmat_size(m,NULL,1)   # Subtract a Theiler window of 1 around and including the diagonal
}
