% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet-package.r
\docType{package}
\name{casnet-package}
\alias{casnet}
\alias{casnet-package}
\title{casnet: A Toolbox for Studying Complex Adaptive Systems and NETworks}
\description{
A collection of analytic tools for studying signals recorded from complex adaptive systems or networks:
    - Recurrence Quantification Analyses (CrossRQA, Categorical RQA, Chromatic RQA, Anisotropic RQA).
    - Fluctuation Analyses (DFA varieties, PSD slope, SDA, Multifractal DFA, Wavelet Singularity Spectrum).
    - Coupling Analyses (Cross Conformal Mapping, Detection of Coupling Direction, CRQA).
    - Network based time series analyses (Recurrence Networks, Multifractal Spectrum Networks, Multiplex Networks).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FredHasselman/casnet}
  \item Report bugs at \url{https://github.com/FredHasselman/casnet/issues}
}

}
\author{
\strong{Maintainer}: Fred Hasselman \email{f.hasselman@pwo.ru.nl} (0000-0003-1384-8361)

Other contributors:
\itemize{
  \item Merlijn Olthof \email{m.olthof@pwo.ru.nl} [contributor]
}

}
\keyword{internal}
