% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet-package.r
\docType{package}
\name{casnet}
\alias{casnet}
\alias{casnet-package}
\title{A Toolbox for Studying Complex Adaptive Systems and NETworks}
\description{
A Toolbox for Studying Complex Adaptive Systems and NETworks
}
\author{
Fred Hasselman
}
