% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{center}
\alias{center}
\title{Center a vector on Mean or Median}
\usage{
center(numvec, na.rm = TRUE, type = c("mean", "median")[1])
}
\arguments{
\item{numvec}{A numeric vector}

\item{na.rm}{Set the \code{na.rm} field}

\item{type}{Center on the \code{"mean"} (default) or the \code{"median"} of the vector}
}
\value{
A mean or median centered vector
}
\description{
Center a vector on Mean or Median
}
\author{
Fred Hasselman
}
