% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ColouredNoise}
\alias{ColouredNoise}
\title{Coloured noise data}
\format{
A data frame with 1024 rows and 41 columns. Each column represent a coloured noise series with spectral slopes ranging from -2 to 2.
}
\usage{
ColouredNoise
}
\description{
Coloured noise data
}
\keyword{data}
