% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOps.R
\name{createCorridor}
\alias{createCorridor}
\title{Corridor analysis}
\usage{
createCorridor(mat, lower, upper, value = NA, silent = TRUE)
}
\arguments{
\item{mat}{A Matrix}

\item{lower}{Lower diagonal to be included in the corridor (should be \eqn{\le 0})}

\item{upper}{Upper diagonal to be included in the corridor (should be \eqn{\ge 0})}

\item{value}{A single value to replace all values outside the corridor (default = \code{NA})}

\item{silent}{Operate in silence, only (some) warnings will be shown (default = \code{TRUE})}
}
\value{
A matrix in which the values outside the corridor have been replaced
}
\description{
Create a corridor around the main diagonal. For long time series it may not make sense to evaluate recurrences on the longest time scales.
}
\examples{
# Create a 10 by 10 matrix
library(Matrix)
m <- Matrix(rnorm(10),10,10)

createCorridor(m,-7,7,0)   # Set diagonals 10 9 and 8 to 0.
}
\seealso{
Other Distance matrix operations (recurrence plot): 
\code{\link{bandReplace}()},
\code{\link{mat_di2bi}()},
\code{\link{mat_di2ch}()},
\code{\link{mat_di2we}()},
\code{\link{mat_hamming}()},
\code{\link{rp}()},
\code{\link{rp_lineDist}()},
\code{\link{rp_nzdiags}()},
\code{\link{rp_plot}()},
\code{\link{rp_size}()}
}
\author{
Fred Hasselman
}
\concept{Distance matrix operations (recurrence plot)}
