% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasNET functies Merlijn.R
\name{crit_in-deprecated}
\alias{crit_in-deprecated}
\title{Critical Instability}
\arguments{
\item{df}{Data frame with output fro `dyn_comp()``}

\item{win}{Window used}
}
\value{
Critical instability
}
\description{
Critical Instability
}
\seealso{
\code{\link{casnet-deprecated}}
}
\keyword{internal}
