% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasNET_Merlijn.R
\name{crit_in}
\alias{crit_in}
\title{Critical Instability}
\usage{
crit_in(df, win)
}
\arguments{
\item{df}{A dataframe of dynamic complexity values, e.g. output data from dyn_comp}

\item{win}{Window size over which dynamic complexity was calculated (nescesary for plot)}
}
\value{
A list with a dataframe of binary critical instability indices and a summed critical intability index, a ggplot object containing a critical instability diagram.
}
\description{
Computes significant peaks in the dynamic complexity time series. Example: Schiepek, Tominschek & Heinzel, 2014.
}
\references{
Schiepek, G., & Strunk, G. (2010). The identification of critical fluctuations and phase transitions in short term and coarse-grained time series-a method for the real-time monitoring of human change processes. Biological cybernetics, 102(3), 197-207.

Schiepek, G. (2003). A Dynamic Systems Approach to Clinical Case Formulation. European Journal of Psychological Assessment, 19, 175-184.

Haken, H. & Schiepek, G. (2006, 2. Aufl. 2010). Synergetik in der Psychologie. Selbstorganisation verstehen und gestalten. G?ttingen: Hogrefe.

Schiepek, G. K., Tominschek, I., & Heinzel, S. (2014). Self-organization in psychotherapy: testing the synergetic model of change processes. Frontiers in psychology, 5, 1089.
}
\author{
Merlijn Olthof
}
