% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crp_calc}
\alias{crp_calc}
\title{crp_calc}
\usage{
crp_calc(RM, radius = NULL, DLmin = 2, VLmin = 2, HLmin = 2,
  DLmax = length(diag(RM)) - 1, VLmax = length(diag(RM)) - 1,
  HLmax = length(diag(RM)) - 1, AUTO = FALSE, chromatic = FALSE,
  matrices = FALSE)
}
\arguments{
\item{RM}{RM}

\item{radius}{r}

\item{DLmin}{d}

\item{VLmin}{v}

\item{HLmin}{h}

\item{DLmax}{dd}

\item{VLmax}{vv}

\item{HLmax}{hh}

\item{AUTO}{a}

\item{chromatic}{c}

\item{matrices}{m}
}
\value{
crqa measures
}
\description{
crp_calc
}
\keyword{internal}
