% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_cl_main}
\alias{crqa_cl_main}
\title{crqa_cl_main [internal]}
\usage{
crqa_cl_main(y1, y2 = NULL, eDim = 1, eLag = 1, eRad = 1, DLmin = 2,
  VLmin = 2, theiler = 0, win = min(length(y1), ifelse(is.null(y2),
  (length(y1) + 1), length(y2)), na.rm = TRUE), step = win, JRP = FALSE,
  distNorm = c("EUCLIDEAN", "MAX", "MIN", "OP")[[1]], returnMeasures = TRUE,
  returnRP = TRUE, returnLineDist = FALSE,
  path_to_rp = getOption("casnet.path_to_rp"), saveOut = FALSE,
  path_out = NULL, file_ID = NULL, ...)
}
\arguments{
\item{y1}{y1}

\item{y2}{y1}

\item{eDim}{y1}

\item{eLag}{y1}

\item{eRad}{y1}

\item{DLmin}{y1}

\item{VLmin}{y1}

\item{theiler}{y1}

\item{win}{y1}

\item{step}{y1}

\item{JRP}{y1}

\item{distNorm}{y1}

\item{returnMeasures}{y1}

\item{returnRP}{y1}

\item{returnLineDist}{y1}

\item{path_to_rp}{y1}

\item{saveOut}{y1}

\item{path_out}{y1}

\item{file_ID}{y1}

\item{...}{}
}
\description{
crqa_cl_main [internal]
}
\keyword{internal}
