% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_diagPofile}
\alias{crqa_diagPofile}
\title{Diagonal Recurrence Profile}
\usage{
crqa_diagPofile(RM, diagWin = NULL, xname = "", yname = "",
  DLmin = 2, VLmin = 2, HLmin = 2, DLmax = length(Matrix::diag(RM))
  - 1, VLmax = length(Matrix::diag(RM)) - 1,
  HLmax = length(Matrix::diag(RM)) - 1, AUTO = NULL,
  chromatic = FALSE, matrices = FALSE, doPlot = TRUE)
}
\arguments{
\item{RM}{A binary recurrence matrix}

\item{diagWin}{Window around the line of synchrony}

\item{xname}{xlab}

\item{yname}{ylab}

\item{DLmin}{DLmin}

\item{VLmin}{VLmin}

\item{HLmin}{HLmin}

\item{DLmax}{DLmax}

\item{VLmax}{VLmax}

\item{HLmax}{HLmax}

\item{AUTO}{AUTO}

\item{chromatic}{chromatic}

\item{matrices}{matrices}

\item{doPlot}{Plot}
}
\value{
A plot and/or the data for the plot
}
\description{
Diagonal Recurrence Profile
}
