% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_mat}
\alias{crqa_mat}
\title{crqa_mat}
\usage{
crqa_mat(rmat, radius = NULL, DLmin = 2, VLmin = 2, HLmin = 2,
  DLmax = length(diag(rmat)) - 1, VLmax = length(diag(rmat)) - 1,
  HLmax = length(diag(rmat)) - 1, AUTO = FALSE, doHalf = FALSE,
  chromatic = FALSE, matrices = FALSE, Nboot = NULL, CL = 0.95)
}
\arguments{
\item{AUTO}{Auto-recurrence? (default: FALSE)}

\item{matrices}{}

\item{rp}{Recurrence Matrix}
}
\description{
crqa_mat
}
