% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R, R/nlRtsa_SOURCE.old.R
\name{crqa_mat_measures}
\alias{crqa_mat_measures}
\alias{crqa_mat_measures}
\title{Get (C)RQA measures from Recurrence Matrix}
\usage{
crqa_mat_measures(RM, DLmin = 2, VLmin = 2, HLmin = 2,
  DLmax = length(diag(rmat)) - 1, VLmax = length(diag(rmat)) - 1,
  HLmax = length(diag(rmat)) - 1, AUTO = NULL, chromatic = FALSE,
  matrices = FALSE, doHalf = FALSE, Nboot = NULL, CL = 0.95)

crqa_mat_measures(RM, DLmin = 2, VLmin = 2, HLmin = 2,
  DLmax = length(diag(rmat)) - 1, VLmax = length(diag(rmat)) - 1,
  HLmax = length(diag(rmat)) - 1, AUTO = NULL, chromatic = FALSE,
  matrices = FALSE, doHalf = FALSE, Nboot = NULL, CL = 0.95)
}
\arguments{
\item{RM}{A binary recurrence matrix}

\item{DLmin}{Minimal diagonal line length}

\item{VLmin}{Minimal vertical line length}

\item{HLmin}{Minimal horizontal line length}

\item{DLmax}{Maximal diagonal line length}

\item{VLmax}{Maximal vertical line length}

\item{HLmax}{Maximal horizontal line length}

\item{AUTO}{Is this an AUTO RQA?}

\item{chromatic}{Chromatic RQA?}

\item{matrices}{Return Matrices?}

\item{doHalf}{Analyse half of the matrix?}

\item{Nboot}{How many bootstraps?}

\item{CL}{Confidence Limit for bootstrap results}

\item{RM}{A binary recurrence matrix}

\item{DLmin}{Minimal diagonal line length}

\item{VLmin}{Minimal vertical line length}

\item{HLmin}{Minimal horizontal line length}

\item{DLmax}{Maximal diagonal line length}

\item{VLmax}{Maximal vertical line length}

\item{HLmax}{Maximal horizontal line length}

\item{AUTO}{Is this an AUTO RQA?}

\item{chromatic}{Chromatic RQA?}

\item{matrices}{Return Matrices?}

\item{doHalf}{Analyse half of the matrix?}

\item{Nboot}{How many bootstraps?}

\item{CL}{Confidence Limit for bootstrap results}
}
\description{
Use `crqa_mat`

Use `crqa_mat`
}
