% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_parameters}
\alias{crqa_parameters}
\title{Find optimal (C)RQA parameters}
\usage{
crqa_parameters(y, maxDim = 5, maxLag = round(length(y)/(maxDim + 1)),
  emLag = est_emLag(y, selection.methods = c("first.e.decay", "first.zero",
  "first.minimum"), maxLag = round(length(y)/(maxDim + 1))),
  emLags = est_emLag(y, maxLag = round(length(y)/(maxDim + 1))),
  ami.method = c("first.minimum"), nnSizes = c(0.1, 0.3, 0.6, 0.9),
  nnThres = 0.1, diagPlot = TRUE)
}
\arguments{
\item{y}{Timeseries}

\item{maxDim}{Maximum number of embedding dimensions}

\item{maxLag}{Maximum embedding lag}

\item{emLag}{Optimal embedding lag (delay), e.g., provided by optimising algorithm.}

\item{emLags}{A range of embedding lags to consider (default)}

\item{ami.method}{Method}

\item{nnSizes}{Neighbourhood size}

\item{nnThres}{Threshold}

\item{diagPlot}{Plot the results}
}
\value{
A list object containing the optimal values and iteration history.
}
\description{
A wrapper for various algorithms used to find optimal embedding delay, number of embedding dimensions and radius.
}
