% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_rp_calc}
\alias{crqa_rp_calc}
\title{crqa_rp_calc}
\usage{
crqa_rp_calc(RM, emRad = NULL, DLmin = 2, VLmin = 2, HLmin = 2,
  DLmax = length(Matrix::diag(RM)) - 1,
  VLmax = length(Matrix::diag(RM)) - 1,
  HLmax = length(Matrix::diag(RM)) - 1, theiler = 0, AUTO = NULL,
  chromatic = FALSE, matrices = FALSE)
}
\arguments{
\item{RM}{RM}

\item{emRad}{r}

\item{DLmin}{d}

\item{VLmin}{v}

\item{HLmin}{h}

\item{DLmax}{dd}

\item{VLmax}{vv}

\item{HLmax}{hh}

\item{theiler}{t}

\item{AUTO}{a}

\item{chromatic}{c}

\item{matrices}{m}
}
\value{
crqa measures
}
\description{
crqa_rp_calc
}
\keyword{internal}
