% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_rp_measures}
\alias{crqa_rp_measures}
\title{Get (C)RQA measures from a Recurrence Matrix}
\usage{
crqa_rp_measures(RM, radius = NULL, DLmin = 2, VLmin = 2, HLmin = 2,
  DLmax = length(Matrix::diag(RM)) - 1, VLmax = length(Matrix::diag(RM)) -
  1, HLmax = length(Matrix::diag(RM)) - 1, AUTO = NULL, chromatic = FALSE,
  matrices = FALSE, doHalf = FALSE, Nboot = NULL, CL = 0.95,
  doParallel = FALSE)
}
\arguments{
\item{RM}{A binary recurrence matrix}

\item{radius}{Threshold for distance value that counts as a recurrence}

\item{DLmin}{Minimal diagonal line length}

\item{VLmin}{Minimal vertical line length}

\item{HLmin}{Minimal horizontal line length}

\item{DLmax}{Maximal diagonal line length}

\item{VLmax}{Maximal vertical line length}

\item{HLmax}{Maximal horizontal line length}

\item{AUTO}{Is this an AUTO RQA?}

\item{chromatic}{Chromatic RQA?}

\item{matrices}{Return Matrices?}

\item{doHalf}{Analyse half of the matrix?}

\item{Nboot}{How many bootstraps?}

\item{CL}{Confidence Limit for bootstrap results}

\item{doParallel}{Use parallel?}
}
\description{
Use `crqa_rp`
}
\seealso{
Other Recurrence Quantification Analysis: \code{\link{crqa_cl}},
  \code{\link{crqa_parameters}}, \code{\link{crqa_radius}},
  \code{\link{crqa_rp}}
}
\keyword{internal}
