% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_rp_prep}
\alias{crqa_rp_prep}
\title{Prepare matrix}
\usage{
crqa_rp_prep(RP, radius = NULL, DLmin = 2, VLmin = 2, HLmin = 2,
  DLmax = length(Matrix::diag(RP)) - 1, VLmax = length(Matrix::diag(RP)) -
  1, HLmax = length(Matrix::diag(RP)) - 1, AUTO = FALSE,
  chromatic = FALSE, matrices = FALSE, doHalf = FALSE)
}
\arguments{
\item{RP}{Recurrence plot}

\item{radius}{Radiuc}

\item{DLmin}{Minimal diagonal line length}

\item{VLmin}{Minimal vertical line length}

\item{HLmin}{Minimal horizontal line length}

\item{DLmax}{Maximal diagonal line length}

\item{VLmax}{Maximal vertical line length}

\item{HLmax}{Maximal horizontal line length}

\item{AUTO}{Is this an AUTO RQA?}

\item{chromatic}{Chromatic RQA?}

\item{matrices}{Return Matrices?}

\item{doHalf}{Analyse half of the matrix?}
}
\value{
A prepped matrix
}
\description{
Prepare matrix
}
\keyword{internal}
