% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{dc_ccp}
\alias{dc_ccp}
\title{Cumulative Complexity Peaks (CCP)}
\usage{
dc_ccp(df_win, col_first, col_last, win = NROW(df), scale_min, scale_max,
  alpha_item = 0.05, alpha_time = 0.05, doPlot = FALSE,
  useVarNames = TRUE, colOrder = TRUE, useTimeVector = NA,
  timeStamp = "01-01-1999")
}
\arguments{
\item{df_win}{A data frame containing series of Dynamic Complexity values obtained by running function \code{\link[=dc_win]{dc_win()}}}

\item{col_first}{The first column of the dataframe that should be analyzed.}

\item{col_last}{The last column of the dataframe that should be analyzed.}

\item{win}{Size of window in which to calculate Dynamic Complexity. If \code{win < NROW(df)} the window will move along the time series with a stepsize of \code{1} (default = \code{NROW(df)})}

\item{scale_min}{Theoretical minimum value of thescale.}

\item{scale_max}{Theoretical maximum value of the scale.}

\item{alpha_item}{The significance level of the one-sided Z-test used to determine which peaks are \code{> 0}.}

\item{alpha_time}{The significance level of the one-sided Z-test used to determine if the number of significant peaks (as determined  by \code{alpha_item}) at a specific time stamp are \code{> 0}.}

\item{doPlot}{If \code{TRUE} shows a Complexity Resonance Diagram and returns an invisible \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object. If \code{FALSE}, plot will be drawn and no object will be returned (default = \code{FALSE})}

\item{useVarNames}{Use the column names of \code{df} as variable names in the Complexity Resonance Diagram (default = \code{TRUE})}

\item{colOrder}{If \code{TRUE}, the order of the columns in \code{df} determines the of variables on the y-axis. Use \code{FALSE} for alphabetic/numeric order. Use \code{NA} to sort by by mean value of Dynamic Complexity (default = \code{TRUE})}

\item{useTimeVector}{Parameter used for plotting. Either a variable name, or a vector of length \code{NROW(df)}, containing date/time information (default = \code{NA})}

\item{timeStamp}{If \code{useTimeVector} is not \code{NA}, a character string that can be passed to \code{\link[lubridate:stamp]{lubridate::stamp()}} to format the the dates/times passed in \code{useTimeVector} (default = \code{"01-01-1999"})}
}
\value{
A list with a dataframe of binary complexity peak indices and a cumulative complexity peak index, a CCP diagram.
}
\description{
Computes significant peaks in the dynamic complexity time series. Example: Schiepek, Tominschek & Heinzel, 2014.
}
\references{
Schiepek, G., & Strunk, G. (2010). The identification of critical fluctuations and phase transitions in short term and coarse-grained time series-a method for the real-time monitoring of human change processes. Biological cybernetics, 102(3), 197-207.

Schiepek, G. (2003). A Dynamic Systems Approach to Clinical Case Formulation. European Journal of Psychological Assessment, 19, 175-184.

Haken, H. & Schiepek, G. (2006, 2. Aufl. 2010). Synergetik in der Psychologie. Selbstorganisation verstehen und gestalten. G?ttingen: Hogrefe.

Schiepek, G. K., Tominschek, I., & Heinzel, S. (2014). Self-organization in psychotherapy: testing the synergetic model of change processes. Frontiers in psychology, 5, 1089.
}
\seealso{
Other Dynamic Complexity functions: \code{\link{dc_d}},
  \code{\link{dc_f}}, \code{\link{dc_win}},
  \code{\link{plotDC_ccp}}, \code{\link{plotDC_res}}
}
\author{
Merlijn Olthof

Fred Hasselman
}
\concept{Dynamic Complexity functions}
