% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasNET functies Merlijn.R
\name{DC_D-deprecated}
\alias{DC_D-deprecated}
\title{Distribution Uniformity}
\arguments{
\item{df}{data}

\item{win}{win}

\item{xmin}{xmin}

\item{xmax}{xmax}

\item{col_first}{col_first}

\item{col_last}{col_last}
}
\value{
a dataframe
}
\description{
Distribution Uniformity
}
\seealso{
\code{\link{casnet-deprecated}}
}
\keyword{internal}
