% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasNET_Merlijn.R
\name{DC_D}
\alias{DC_D}
\title{Distribution Uniformity}
\usage{
DC_D(df, win, xmin, xmax, col_first, col_last)
}
\arguments{
\item{win}{win}

\item{xmin}{xmin}

\item{xmax}{xmax}

\item{col_first}{col_first}

\item{col_last}{col_last}

\item{data}{data}
}
\value{
a dataframe
}
\description{
Distribution Uniformity
}
\keyword{internal}
