% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasNET_functies_Merlijn.R
\name{DC_F-deprecated}
\alias{DC_F-deprecated}
\title{Fluctuation Intensity}
\arguments{
\item{df}{df}

\item{win}{win}

\item{xmin}{xmin}

\item{xmax}{xmax}

\item{col_first}{col_first}

\item{col_last}{col_last}
}
\value{
dataframe
}
\description{
Used to calculate dynamic complexity
}
\seealso{
\code{\link{casnet-deprecated}}
}
\keyword{internal}
