% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasNET_Merlijn.R
\name{DC_F}
\alias{DC_F}
\title{Fluctuation Intensity}
\usage{
DC_F(df, win, xmin, xmax, col_first, col_last)
}
\arguments{
\item{df}{df}

\item{win}{win}

\item{xmin}{xmin}

\item{xmax}{xmax}

\item{col_first}{col_first}

\item{col_last}{col_last}
}
\value{
dataframe
}
\description{
Used to calculate dynamic complexity
}
\keyword{internal}
