% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{dc_f2}
\alias{dc_f2}
\title{Title}
\usage{
dc_f2(df, win = NROW(df), scale_min, scale_max, doPlot = FALSE,
  useVarNames = TRUE, colOrder = TRUE, useTimeVector = NA,
  timeStamp = "01-01-1999")
}
\description{
Title
}
