% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{DFA1}
\alias{DFA1}
\title{Detrended Fluctuation Analysis}
\usage{
DFA1(signal, mins = 4, maxs = 12, ressc = 30, m = 1)
}
\arguments{
\item{signal}{An input signal.}

\item{mins}{Minimum scale to consider.}

\item{maxs}{Maximum scale to consider.}

\item{ressc}{ressc}

\item{m}{m}
}
\value{
output
}
\description{
Detrended Fluctuation Analysis
}
