% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOps.R
\name{di2ch}
\alias{di2ch}
\title{Distance 2 chromatic matrix}
\usage{
di2ch(distmat, y, emRad, theiler = 0, convMat = FALSE)
}
\arguments{
\item{distmat}{Distance matrix}

\item{y}{One of the dimensions (as a data frame or matrix) of the RP which must contain unique unordered categorical values}

\item{emRad}{The radius or threshold value}

\item{theiler}{Use a theiler window around the line of identity / synchronisation to remove high auto-correlation at short time-lags (default = \code{0})}

\item{convMat}{convMat Should the matrix be converted from a \code{distmat} object of class \code{\link[Matrix:Matrix]{Matrix::Matrix()}} to \code{\link[base:matrix]{base::matrix()}} (or vice versa)}
}
\value{
A matrix with 0s and the unordered categorical values that are recurring
}
\description{
Distance matrix to chromatic matrix based on unordered categorical series
}
\seealso{
Other Distance matrix operations (recurrence plot): 
\code{\link{bandReplace}()},
\code{\link{di2bi}()},
\code{\link{di2we}()},
\code{\link{dist_hamming}()},
\code{\link{rp_lineDist}()},
\code{\link{rp_nzdiags}()},
\code{\link{rp_plot}()},
\code{\link{rp_size}()},
\code{\link{rp}()}

Other Distance matrix operations (recurrence network): 
\code{\link{di2bi}()},
\code{\link{di2we}()},
\code{\link{rn_phases}()},
\code{\link{rn_plot}()},
\code{\link{rn_recSpec}()},
\code{\link{rn}()}
}
\concept{Distance matrix operations (recurrence network)}
\concept{Distance matrix operations (recurrence plot)}
