% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R, R/nlRtsa_SOURCE.old.R
\name{dist.hamming}
\alias{dist.hamming}
\alias{dist.hamming}
\title{Calculate Hamming distance}
\usage{
dist.hamming(X, Y = NULL, embedded = TRUE)

dist.hamming(X, Y = NULL, embedded = TRUE)
}
\arguments{
\item{X}{A matrix (of coordinates)}

\item{Y}{A matrix (of coordinates)}

\item{embedded}{Do X and/or Y represent surrogate dimensions of an embedded time series?}

\item{X}{A matrix (of coordinates)}

\item{Y}{A matrix (of coordinates)}

\item{embedded}{Do X and/or Y represent surrogate dimensions of an embedded time series?}
}
\value{
A hamming-distance matrix of X, or X and Y. Useful for ordered and unordered categorical data.

A hamming-distance matrix of X, or X and Y. Useful for ordered and unordered categorical data.
}
\description{
Calculate Hamming distance

Calculate Hamming distance
}
