% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{dist_hamming}
\alias{dist_hamming}
\title{Calculate Hamming distance}
\usage{
dist_hamming(X, Y = NULL, embedded = TRUE)
}
\arguments{
\item{X}{A matrix (of coordinates)}

\item{Y}{A matrix (of coordinates)}

\item{embedded}{Do X and/or Y represent surrogate dimensions of an embedded time series?}
}
\value{
A hamming-distance matrix of X, or X and Y. Useful for ordered and unordered categorical data.
}
\description{
Calculate Hamming distance
}
\seealso{
Other Distance matrix operations: \code{\link{bandReplace}},
  \code{\link{di2bi}}, \code{\link{di2we}},
  \code{\link{lineDists}}, \code{\link{nzdiags}},
  \code{\link{recmat_plot}}, \code{\link{recmat_size}},
  \code{\link{recmat}}
}
\author{
Fred Hasselman
}
