% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasNET_functies_Merlijn.R
\name{dyn_comp-deprecated}
\alias{dyn_comp-deprecated}
\title{Dynamic Complexity}
\arguments{
\item{df}{df multivariate time series data - rows=time 1 person  cols = vars}

\item{col_first}{col_first}

\item{col_last}{col_last}

\item{col_time}{if col time use it otherwise assume ....}

\item{win}{win}

\item{fromScaleMin}{dat scale xmin}

\item{fromScaleMax}{data scale xmax}

\item{toScaleMin}{theoretical scale min}

\item{toScaleMax}{theoretical scale min}

\item{doPlot}{Show a plot}
}
\value{
A list object
}
\description{
Calculate DynComp (deprecated, use \link{dc_win} instead)
}
\references{
Guenther
}
\seealso{
\code{\link{casnet-deprecated}}
}
\author{
Merlijn Olthof
}
\keyword{internal}
