% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasNET_Merlijn.R
\name{dyn_comp}
\alias{dyn_comp}
\title{Dynamic Complexity}
\usage{
dyn_comp(df, col_first, col_last, scale_min, scale_max, win = NROW(df),
  doPlot = FALSE)
}
\arguments{
\item{df}{A dataframe containing multivariate time series data from 1 person. Rows should indicate time,  columns should indicate variables. The multivariate time series should be on the same scale. If nescessary, rescale variables.}

\item{col_first}{The first column of the dataframe that should be analyzed.}

\item{col_last}{The last column of the dataframe that should be analyzed.}

\item{scale_min}{Theoretical minimum value of thescale.}

\item{scale_max}{Theoretical maximum value of the scale.}

\item{win}{Window size (default = \code{NROW(df)}}

\item{doPlot}{If \code{TRUE} shows a plot and returns an invisible \code{\link[ggplot2]{ggplot}} object. If \code{FALSE}, plot will be drawn and no object will be returned (default = \code{FALSE})}
}
\value{
If \code{doPlot = TRUE}, a list object containing a dataframe of dynamic complexity values and a \code{ggplot2} object of the dynamic complexity resonance diagram (e.g. Schiepek et al., 2016). If \code{doPlot = FALSE} the dataframe is returned.
}
\description{
Calculates Dynamic Complexity, a complexity index for short and coarse-grained time series (Schiepek & Strunk, 2010; Schiepek, 2003; Haken & Schiepek 2006).
}
\references{
Schiepek, G., & Strunk, G. (2010). The identification of critical fluctuations and phase transitions in short term and coarse-grained time series-a method for the real-time monitoring of human change processes. Biological cybernetics, 102(3), 197-207.

Schiepek, G. (2003). A Dynamic Systems Approach to Clinical Case Formulation. European Journal of Psychological Assessment, 19, 175-184.

Haken, H. & Schiepek, G. (2006, 2. Aufl. 2010). Synergetik in der Psychologie. Selbstorganisation verstehen und gestalten. G?ttingen: Hogrefe.

Schiepek, G. K., St?ger-Schmidinger, B., Aichhorn, W., Sch?ller, H., & Aas, B. (2016). Systemic case formulation, individualized process monitoring, and state dynamics in a case of dissociative identity disorder. Frontiers in psychology, 7, 1545.
}
\author{
Merlijn Olthof
}
