% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{est_eDim}
\alias{est_eDim}
\title{eDim}
\usage{
est_eDim(y, delay = tau(y), maxDim = 20, threshold = 0.95,
  max.relative.change = 0.1, ...)
}
\arguments{
\item{y}{Time series}

\item{delay}{Embeding lag}

\item{maxDim}{Maximum number of embedding dimensions}

\item{...}{}
}
\description{
A wrapper for nonlinearTseries::estimateEmbeddingDim
}
