% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{est_eLag}
\alias{est_eLag}
\title{Estimate embedding lag (tau)}
\usage{
est_eLag(y, selection.methods = c("first.minimum"), maxLag = length(y)/4,
  ...)
}
\arguments{
\item{y}{Time series}

\item{selection.methods}{Selecting an optimal embedding lag (default: Return "first.e.decay", "first.zero", "first.minimum", "first.value", where value is 1/exp(1))}

\item{maxLag}{Maximal lag to consider (default: 1/4 of timeseries length)}

\item{...}{Additional parameters}
}
\value{
The ami function with minima
}
\description{
A wrapper for nonlinearTseries::timeLag
}
