% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{est_emDim}
\alias{est_emDim}
\title{Estimate number of embedding dimensions}
\usage{
est_emDim(y, delay = est_emLag(y), maxDim = 20, threshold = 0.95,
  max.relative.change = 0.1, ...)
}
\arguments{
\item{y}{Time series}

\item{delay}{Embeding lag}

\item{maxDim}{Maximum number of embedding dimensions}

\item{threshold}{See \code{\link[nonlinearTseries]{estimateEmbeddingDim}}}

\item{max.relative.change}{See \code{\link[nonlinearTseries]{estimateEmbeddingDim}}}

\item{...}{Other arguments (not in use)}
}
\value{
Embedding dimensions
}
\description{
A wrapper for nonlinearTseries::estimateEmbeddingDim
}
