% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est_maxPhases}
\alias{est_maxPhases}
\title{Estimate the maximum number of Phases}
\usage{
est_maxPhases(
  RN,
  range = 2:10,
  minStatesinPhase = 1,
  maxStatesinPhase = NROW(RN),
  useDegree = FALSE,
  inverseWeight = TRUE,
  cleanUp = TRUE,
  removeSingularities = TRUE
)
}
\arguments{
\item{RN}{Recurrence matrix}

\item{range}{Two element vector with minimum and maximum \code{c(min,max)} number of phases to check.}

\item{minStatesinPhase}{A parameter applied after the extraction of phases (limited by \code{maxPhases}). If any extracted phases do not have a minimum number of \code{minStatesinPhase} + \code{1} (= the state that was selected based on node strength), the phase will be removed from the result (default = \code{1})}

\item{maxStatesinPhase}{A parameter applied after the extraction of phases (limited by \code{maxPhases}). If any extracted phases exceeds a maximum number of \code{maxStatesinPhase} + \code{1} (= the state that was selected based on node strength), the phase will be removed from the result (default = \code{NROW(RN)})}

\item{useDegree}{By default, node strength will be used to determine the phases. Set to \code{TRUE} to use the node degree (default = \code{FALSE})}

\item{inverseWeight}{Whether to perform the operation \code{1/weight} on the edge weights. The default is \code{TRUE}, if the matrix was weighted by a distance metric (\code{weightedBy = "si"}) edges with smaller distances (recurring coordinates closer to the current coordinate) have greater impact on the node strength calculation used to select the phases. If the matrix was weighted by recurrence time (\code{weightedBy = "rt"}) and \code{inverseWeight = TRUE}, recurrent points with shorter recurrence times will have greater impact on the strength calculation. If \code{weightedBy = "rf"}, lower frequencies will end up having more impact. (default = \code{TRUE})}

\item{cleanUp}{Try to assign states to phases that were not assigned by the algorithm. If \code{FALSE}, these states will be added to category "Other" (default = \code{TRUE})}

\item{removeSingularities}{Will remove states that recur only once (nodes with \code{degree(g) == 1}) (default = \code{FALSE})}
}
\value{
Data frame with maxPhases by detectedPhases
}
\description{
Parameter sweep of function \link{rn_phases} for argument \code{maxPhases}. Use to check at which value of \code{maxPhases} no additional phases will be detected.
}
