% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est_parameters}
\alias{est_parameters}
\title{Estimate RQA parameters}
\usage{
est_parameters(
  y,
  lagMethods = c("first.minimum", "global.minimum", "max.lag"),
  estimateDimensions = "preferSmallestInLargestHood",
  maxDim = 10,
  emLag = NULL,
  maxLag = floor(NROW(y)/(maxDim + 1)),
  minVecLength = 20,
  nnSize = NA,
  nnRadius = sd(y)/10,
  nnThres = 10,
  theiler = 0,
  doPlot = TRUE,
  silent = TRUE,
  ...
)
}
\arguments{
\item{y}{A numeric vector or time series}

\item{lagMethods}{A character vector with one or more of the following strings: \verb{"first.minimum","global.minimum","max.lag"}. If \code{emLag} represents a valid lag this value will be reported as \code{"user.lag"} (default = \code{c("first.minimum","global.minimum","max.lag")})}

\item{estimateDimensions}{Decide on an optimal embedding dimension relative to the values in \code{maxDim} and \code{lagMethods}, according to a number of preferences passed as a character vector. The order in which the preferences appear in the vector affects the selection procedure, with index \code{1} being most important preference. The following options are available:
\itemize{
\item \code{preferNone} - No optimal number will be picked all other preferences will be ignored
\item \code{preferSmallestDim} - Pick smallest number of dimensions associated with a percentage NN below \code{nnThres}
\item \code{preferSmallestNN} - Pick the number of dimensions that is associated with the smallest percentage NN below \code{nnThres}
\item \code{preferSmallestLag} - If the value of \code{nnThres} does not lead to a unique preference for a pair of dimension and lag values, use the pair with the smallest lag
\item \code{preferSmallestInLargestHood} - The default option: If no unique pair can be found, prefer pairs with smallest values for lag, dimensions, percentage NN for the largest NN size
}}

\item{maxDim}{Maximum number of embedding dimensions to consider (default = \code{10})}

\item{emLag}{Optimal embedding lag (delay), e.g., provided by an optimising algorithm. If \code{NULL} the lags based on the mutual information in \code{lagMethods} will be reported. If a numeric value representing a valid lag is passed, this value will be used to estimate the number of dimensions (default = \code{NULL})}

\item{maxLag}{Maximum embedding lag to consider. Default value is: \code{floor(length(y)/(maxDim+1))}}

\item{minVecLength}{The minimum length of state space vectors after delay-embedding. For short time series, this will affect the possible values of \code{maxDim} that can be used to evaluate the drop in nearest neighbours. In general it is not recommended to evaluate high dimensional state spaces, based on a small number of state soace coordinates, the default is an absolute minimum and possibly even lower than that. (default = \code{20})}

\item{nnSize}{Neighbourhood diameter (integer, the \code{number.boxes} parameter of \code{\link[tseriesChaos:false.nearest]{tseriesChaos::false.nearest()}}) used to speed up neighbour search. (default = \code{NA})}

\item{nnRadius}{Points smaller than the radius are considered neighbours (default = \code{sd(y)/10})}

\item{nnThres}{Threshold value (in percentage 0-100) representing the percentage of Nearest Neighbours that would be acceptable when using N surrogate dimensions. The smallest number of surrogate dimensions that yield a value below the threshold will be considered optimal (default = \code{10})}

\item{theiler}{Theiler window on distance matrix (default = \code{0})}

\item{doPlot}{Produce a diagnostic plot the results (default = \code{TRUE})}

\item{silent}{Silent-ish mode}

\item{...}{Other parameters passed to \code{\link[nonlinearTseries:timeLag]{nonlinearTseries::timeLag()}}}
}
\value{
A list object containing the optimal values (as indicated by the user) and iteration history.
}
\description{
Find optimal parameters for constructing a Recurrence Matrix. A wrapper for various algorithms used to find optimal values for the embedding delay and the number of embedding dimensions.
}
\details{
A number of functions are called to determine optimal parameters for delay embedding a time series:
\itemize{
\item Embedding lag (\code{emLag}): The default is to call \code{\link[=est_emLag]{est_emLag()}}, which is a wrapper around \code{\link[nonlinearTseries:timeLag]{nonlinearTseries::timeLag()}} with \code{technique=ami} to get lags based on the mutual information function.
\item Embedding dimension (\code{m}, \code{emDim}): The default is to call \code{\link[=est_emDim]{est_emDim()}}, which is a wrapper around \code{\link[tseriesChaos:false.nearest]{tseriesChaos::false.nearest()}}
}
}
\examples{

set.seed(4321)
est_parameters(y=rnorm(500))

}
\seealso{
Other Estimate Recurrence Parameters: 
\code{\link{est_emDim}()},
\code{\link{est_emLag}()},
\code{\link{est_parameters_roc}()},
\code{\link{est_radius}()}
}
\concept{Estimate Recurrence Parameters}
