% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est_parameters_roc}
\alias{est_parameters_roc}
\title{Estimate ROC radius}
\usage{
est_parameters_roc(
  y,
  emRad,
  emDim = 1,
  emLag = 1,
  noiseLevel = 0.75,
  standardise = c("mean.sd", "median.mad", "none")[3],
  noiseType = c("normal", "uniform")[1]
)
}
\arguments{
\item{y}{y}

\item{emRad}{radius}

\item{emDim}{embedding Dims}

\item{emLag}{embedding Lag}

\item{noiseLevel}{noise Level}

\item{standardise}{Standardise y? Choose from "mean.sd","median.mad","none".}

\item{noiseType}{Use a Normal distribution of uniform distribution for noiselevels}
}
\value{
data frame for ROC
}
\description{
Experimental.
}
\seealso{
Other Estimate Recurrence Parameters: 
\code{\link{est_emDim}()},
\code{\link{est_emLag}()},
\code{\link{est_parameters}()},
\code{\link{est_radius}()}
}
\concept{Estimate Recurrence Parameters}
\keyword{internal}
