% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{factor_obs_exp}
\alias{factor_obs_exp}
\title{Add expected factor labels to observed values}
\usage{
factor_obs_exp(
  observed_Ncat,
  observed_labels,
  expected_Ncat = 0,
  expected_labels = "",
  varname = ""
)
}
\arguments{
\item{observed_Ncat}{obsN}

\item{observed_labels}{obsL}

\item{expected_Ncat}{expN}

\item{expected_labels}{expL}

\item{varname}{varname}
}
\value{
character vector
}
\description{
Add expected factor labels to observed values
}
\seealso{
Other State Space Grid functions: 
\code{\link{ssg_gwf2long}()},
\code{\link{ssg_winnowing}()}
}
\concept{State Space Grid functions}
