% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R, R/nlRtsa_SOURCE.old.R
\name{fd.default}
\alias{fd.default}
\alias{fd.default}
\title{fd.default}
\usage{
\method{fd}{default}(y, ...)

\method{fd}{default}(y, ...)
}
\arguments{
\item{y}{timeseries}

\item{...}{Other arguments}

\item{y}{timeseries}

\item{...}{Other arguments}
}
\description{
fd.default

fd.default
}
\seealso{
Other FD estimators: \code{\link{fd.dfa}},
  \code{\link{fd.psd}}, \code{\link{fd.sda}}

Other FD estimators: \code{\link{fd.dfa}},
  \code{\link{fd.psd}}, \code{\link{fd.sda}}
}
\keyword{internal}
