% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R, R/nlRtsa_SOURCE.old.R
\name{fd.dfa}
\alias{fd.dfa}
\alias{fd.dfa}
\title{Detrended Fluctuation Analysis (DFA)}
\usage{
\method{fd}{dfa}(y, fs = NULL, dtrend = "poly1", normalize = FALSE,
  sum.order = 1, scale.max = trunc(length(y)/4), scale.min = 4,
  elasceratio = 2^(1/4), overlap = 0, plot = FALSE)

\method{fd}{dfa}(y, fs = NULL, dtrend = "poly1", normalize = FALSE,
  sum.order = 1, scale.max = trunc(length(y)/4), scale.min = 4,
  elasceratio = 2^(1/4), overlap = 0, plot = FALSE)
}
\arguments{
\item{y}{A numeric vector or time series object.}

\item{normalize}{Normalize the series (default).}

\item{plot}{Return the log-log spectrum with linear fit (default).}

\item{detrend}{Subtract linear trend from the series (default).}

\item{dmethod}{Method to use for detrending, see \code{\link[fractal]{DFA}}.}

\item{y}{A numeric vector or time series object.}

\item{normalize}{Normalize the series (default).}

\item{detrend}{Subtract linear trend from the series (default).}

\item{dmethod}{Method to use for detrending, see \code{\link[fractal]{DFA}}.}

\item{plot}{Return the log-log spectrum with linear fit (default).}
}
\value{
Estimate of Hurst exponent (slope of \code{log(bin)} vs. \code{log(RMSE))} and an FD estimate based on Hasselman(2013)
A list object containing:
\itemize{
\item A data matrix \code{PLAW} with columns \code{freq.norm}, \code{size} and \code{bulk}.
\item Estimate of scaling exponent \code{sap} based on a fit over the standard range (\code{fullRange}), or on a user defined range \code{fitRange}.
\item Estimate of the the Fractal Dimension (\code{FD}) using conversion formula's reported in Hasselman(2013).
\item Information output by various functions.
}

Estimate of Hurst exponent (slope of \code{log(bin)} vs. \code{log(RMSE))} and an FD estimate based on Hasselman(2013)
A list object containing:
\itemize{
\item A data matrix \code{PLAW} with columns \code{freq.norm}, \code{size} and \code{bulk}.
\item Estimate of scaling exponent \code{sap} based on a fit over the standard range (\code{fullRange}), or on a user defined range \code{fitRange}.
\item Estimate of the the Fractal Dimension (\code{FD}) using conversion formula's reported in Hasselman(2013).
\item Information output by various functions.
}
}
\description{
fd.dfa

fd.dfa
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}

Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}
}
\seealso{
Other FD estimators: \code{\link{fd.default}},
  \code{\link{fd.psd}}, \code{\link{fd.sda}}

Other FD estimators: \code{\link{fd.default}},
  \code{\link{fd.psd}}, \code{\link{fd.sda}}
}
\author{
Fred Hasselman

Fred Hasselman
}
