% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{fd.sda}
\alias{fd.sda}
\title{Standardised Dispersion Analysis (SDA).}
\usage{
\method{fd}{sda}(y, fs = NULL, normalize = TRUE, dtrend = FALSE,
  scales = fractal::dispersion(y)$scale, fitRange = c(scales[1],
  scales[length(scales) - 2]), doPlot = FALSE)
}
\arguments{
\item{y}{A numeric vector or time series object.}

\item{fs}{Sample rate (default = NULL)}

\item{normalize}{Normalize the series (default = TRUE)}

\item{dtrend}{Subtract linear trend from the series (default = TRUE)}

\item{scales}{default = \code{fractal::dispersion(y)$scale}, see \code{\link[fractal]{dispersion}}}

\item{fitRange}{Scale bins (\code{c(min,max)}) to use for fitting the scaling relation}

\item{doPlot}{Return the log-log spectrum with linear fit (default).}
}
\value{
A list object containing:
\itemize{
\item A data matrix \code{PLAW} with columns \code{freq.norm}, \code{size} and \code{bulk}.
\item Estimate of scaling exponent \code{sap} based on a fit over the standard range (\code{fullRange}), or on a user defined range \code{fitRange}.
\item Estimate of the the Fractal Dimension (\code{FD}) using conversion formula's reported in Hasselman(2013).
\item Information output by various functions.
}
}
\description{
fd.sda
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}
}
\seealso{
Other FD estimators: \code{\link{fd.default}},
  \code{\link{fd.dfa}}, \code{\link{fd.psd}}
}
\author{
Fred Hasselman
}
