% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{fd_allan}
\alias{fd_allan}
\title{Allan Variance Analysis}
\usage{
fd_allan(y, fs = stats::tsp(stats::hasTsp(y))[3], doPlot = FALSE,
  useSD = FALSE)
}
\arguments{
\item{y}{A numeric vector or time series object}

\item{fs}{Sample frequency in Hz}

\item{doPlot}{Produce a plot of the Allan variance?}

\item{useSD}{Use the standarddeviation instead of variance?}
}
\value{
A dataframe with the Allan Factor (variance), Alan standard deviation and error due to bin size
}
\description{
Allan Variance Analysis
}
\seealso{
Other Fluctuation Analyses: \code{\link{MFDFA}},
  \code{\link{fd_RR}}, \code{\link{fd_dfa}},
  \code{\link{fd_psd}}, \code{\link{fd_sda}},
  \code{\link{fd_sev}}
}
\concept{Fluctuation Analyses}
