% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{fd_mfdfa}
\alias{fd_mfdfa}
\title{Multi-fractal Detrended Fluctuation Analysis}
\usage{
fd_mfdfa(signal, qq = c(-10, -5:5, 10), mins = 6, maxs = 12,
  ressc = 30, m = 1)
}
\arguments{
\item{signal}{An input signal.}

\item{qq}{A vector containing a range of values for the order of fluctuation \code{q}.}

\item{mins}{Minimum scale to consider.}

\item{maxs}{Maximum scale to consider.}

\item{ressc}{rescc}

\item{m}{m}
}
\value{
output
}
\description{
Multi-fractal Detrended Fluctuation Analysis
}
\seealso{
Other Fluctuation Analyses: \code{\link{fd_RR}},
  \code{\link{fd_allan}}, \code{\link{fd_dfa}},
  \code{\link{fd_psd}}, \code{\link{fd_sda}},
  \code{\link{fd_sev}}
}
\concept{Fluctuation Analyses}
