% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{fd_prepSeries}
\alias{fd_prepSeries}
\title{Prepare time series for fluctuation analysis}
\usage{
fd_prepSeries(
  y,
  fs = NULL,
  removeTrend = c("no", "poly", "adaptive", "bridge")[2],
  polyOrder = 1,
  standardise = c("none", "mean.sd", "median.mad")[2],
  adjustSumOrder = FALSE,
  scaleMin = 4,
  scaleMax = log2(floor(NROW(y)/2)),
  scaleResolution = (scaleMax - scaleMin),
  dataMin = NA,
  scaleS = NA,
  overlap = NA,
  silent = TRUE
)
}
\arguments{
\item{y}{y}

\item{fs}{fs}

\item{removeTrend}{rt}

\item{polyOrder}{po}

\item{standardise}{st}

\item{adjustSumOrder}{ao}

\item{scaleMin}{sm}

\item{scaleMax}{smm}

\item{scaleResolution}{sr}

\item{dataMin}{dm}

\item{scaleS}{sc}

\item{overlap}{ov}

\item{silent}{si}
}
\value{
transformed series
}
\description{
Prepare time series for fluctuation analysis
}
\keyword{internal}
