% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{fd_RR}
\alias{fd_RR}
\title{Relative Roughness}
\usage{
fd_RR(y)
}
\arguments{
\item{y}{A numeric vector.}
}
\value{
The Relative Roughness of y, the values of local and global variance are returned as attributes
}
\description{
Relative Rougness is a ratio of local variance (autocovariance at lag-1) to global variance (autocovariance at lag-0) that can be used to classify different 'noises'.
}
\details{
\deqn{RR = 2 * \left[1 - \frac{\gamma(y)}{Var(y)}\right]}{RR = 2 * [ 1 - autoCov-1(y) / Var(y) ]}
}
\references{
\itemize{
\item{Marmelat, V., Torre, K., & Delignieres, D. (2012). Relative roughness: an index for testing the suitability of the monofractal model. \emph{Frontiers in Physiology, 3}, 208.}}
}
\seealso{
Other Fluctuation Analyses: 
\code{\link{fd_allan}()},
\code{\link{fd_dfa}()},
\code{\link{fd_mfdfa}()},
\code{\link{fd_psd}()},
\code{\link{fd_sda}()},
\code{\link{fd_sev}()}
}
\concept{Fluctuation Analyses}
