% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{fd_sda}
\alias{fd_sda}
\title{Standardised Dispersion Analysis (SDA).}
\usage{
fd_sda(y, fs = NULL, standardise = TRUE, detrend = FALSE,
  scales = 2^(1:(log2(length(y)/2))), fitRange = c(scales[1],
  scales[length(scales) - 2]), doPlot = FALSE, returnPLAW = FALSE,
  returnInfo = FALSE, silent = TRUE)
}
\arguments{
\item{y}{A numeric vector or time series object.}

\item{fs}{Sample rate (default = NULL)}

\item{standardise}{standardise the series (default = TRUE)}

\item{detrend}{Subtract linear trend from the series (default = TRUE)}

\item{scales}{default = \code{fractal::dispersion(y)$scale}, see \code{\link[fractal]{dispersion}}}

\item{fitRange}{Scale bins (\code{c(min,max)}) to use for fitting the scaling relation}

\item{doPlot}{Return the log-log spectrum with linear fit (default = \code{TRUE}).}

\item{returnPLAW}{Return the power law data (default = \code{FALSE})}

\item{returnInfo}{Return all the data used in DFA (default = \code{FALSE})}

\item{silent}{Silent-ish mode}
}
\value{
A list object containing:
\itemize{
\item A data matrix \code{PLAW} with columns \code{freq.norm}, \code{size} and \code{bulk}.
\item Estimate of scaling exponent \code{sap} based on a fit over the standard range (\code{fullRange}), or on a user defined range \code{fitRange}.
\item Estimate of the the Fractal Dimension (\code{FD}) using conversion formula's reported in Hasselman(2013).
\item Information output by various functions.
}
}
\description{
fd_sda
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}
}
\author{
Fred Hasselman
}
