% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{fd_sev}
\alias{fd_sev}
\title{Calculate FD using Sevcik's method}
\usage{
fd_sev(y, detrend = FALSE, adjustSumOrder = FALSE, doPlot = TRUE,
  returnPLAW = FALSE, returnInfo = FALSE, silent = TRUE, ...)
}
\arguments{
\item{y}{A time series or numeric vector}

\item{detrend}{Subtract linear trend from the series (default = \code{TRUE}).}

\item{adjustSumOrder}{Adjust the time series (summation or differencing), based on the global scaling exponent, see e.g. [Ihlen (2012)](https://www.frontiersin.org/files/Articles/23948/fphys-03-00141-r2/image_m/fphys-03-00141-t001.jpg) (default = \code{TRUE})}

\item{doPlot}{Return the log-log scale versus fluctuation plot with linear fit (default = \code{TRUE}).}

\item{returnPLAW}{Return the FD series (default = \code{FALSE})}

\item{returnInfo}{Return all the data used in DFA (default = \code{FALSE})}

\item{silent}{Silent-ish mode}

\item{...}{Other arguments}
}
\value{
An FD estimate
}
\description{
Calculate FD using Sevcik's method
}
\references{
Sevcik, C. (1998). A procedure to Estimate the Fractal Dimension of Waveforms. Paper available at http://arxiv.org/pdf/1003.5266.pdf
}
\author{
Fred Hasselman
}
