% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{find_peaks}
\alias{find_peaks}
\title{Find Peaks or Wells}
\usage{
find_peaks(x, window = 3, includeWells = FALSE,
  minPeakDist = round(window/2), minPeakHeight = 0.2 * diff(range(x, na.rm =
  TRUE)))
}
\arguments{
\item{x}{A time series}

\item{window}{Window in whcih to look for peaks or wells}

\item{includeWells}{Find wells?}

\item{minPeakDist}{Minimum distance between peaks or wells}

\item{minPeakHeight}{Minimum height / depth for a peak / well}
}
\value{
Index with peak or well coordinates
}
\description{
Find Peaks or Wells
}
