% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{flight_Cauchy}
\alias{flight_Cauchy}
\title{Create Cauchy Flight}
\usage{
flight_Cauchy(
  N = 1000,
  ndims = 2,
  alpha = 1,
  beta = 0,
  scale = 1,
  location = 0
)
}
\arguments{
\item{N}{Length of time series (default = \code{1000})}

\item{ndims}{Number of dimensions (default = \code{2})}

\item{alpha}{Index of stability parameter in \verb{(0,2]}}

\item{beta}{Skewness parameter in \verb{[-1,1]}}

\item{scale}{Scale parameterin \verb{(0,Inf)}}

\item{location}{Location (shift) parameter in \verb{[-Inf,Inf]}}
}
\value{
A data frame with \code{ndims} columns and \code{N} rows.
}
\description{
Creates a Cauchy flight by taking increments from the Cauchy distributions
implemented as the stable distribution (\code{\link[stabledist:dist-stable]{stabledist::rstable()}}) with index paramter \code{alpha = 1}
and skewness parameter \code{beta = 0}.
}
\examples{

df <- flight_Cauchy()
plot(density(diff(df$dim1)))
plot(df$dim1, df$dim2, type = "l")

}
