% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{flight_LevyPareto}
\alias{flight_LevyPareto}
\title{Create a Levy-Pareto flight}
\usage{
flight_LevyPareto(
  N = 1000,
  ndims = 2,
  alpha = 1.5,
  beta = 0,
  scale = 1,
  location = 0
)
}
\arguments{
\item{N}{Length of time series (default = \code{1000})}

\item{ndims}{Number of dimensions (default = \code{2})}

\item{alpha}{Index of stability parameter in \verb{(0,2]}}

\item{beta}{Skewness parameter in \verb{[-1,1]}}

\item{scale}{Scale parameterin \verb{(0,Inf)}}

\item{location}{Location (shift) parameter in \verb{[-Inf,Inf]}}
}
\value{
A data frame with \code{ndims} columns and \code{N} rows.
}
\description{
Creates a Rayleigh flight by taking increments from the Normal distributions
implemented as the stable distribution (\code{\link[stabledist:dist-stable]{stabledist::rstable()}}) with index paramter \code{alpha = 1.5}
and skewness parameter \code{beta = 0}.
}
\details{
Note that the increments are not strictly from the distribution called \strong{the} Levy distribution, but rather \strong{a}
a Levy-with-Pareto-tail-type distribution (i.e. when \verb{1 < alpha < 2}). Use \code{alpha = 1/2} and \code{beta = 1} if \strong{the} Levy distribution
is required.
}
\examples{

# Levy-Pareto
df <- flight_LevyPareto()
plot(density(diff(df$dim1)))
plot(df$dim1, df$dim2, type = "l")

# "The" Levy distribution
df <- flight_LevyPareto(alpha = 1/2, beta = 1)
plot(density(diff(df$dim1)))
plot(df$dim1, df$dim2, type = "l")

}
