% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{fnn}
\alias{fnn}
\title{False Nearest Neighbours}
\usage{
fnn(y, emLag = 1, maxDim = 10, radius = sd(y)/10, number.boxes = NULL)
}
\arguments{
\item{y}{A numeric vector or time series}

\item{emLag}{Optimal embedding lag (delay), e.g., provided by an optimising algorithm. If \code{NULL} the lags based on the mutual information in \code{lagMethods} will be reported. If a numeric value representing a valid lag is passed, this value will be used to estimate the number of dimensions (default = \code{NULL})}

\item{maxDim}{Maximum number of embedding dimensions to consider (default = \code{10})}

\item{radius}{Size of the neighbourhood: Every point smaller than the radius will be considered a near neighbour, see \code{\link[nonlinearTseries:findAllNeighbours]{nonlinearTseries::findAllNeighbours()}} (default = \code{sd(y)/10}).}

\item{number.boxes}{Integer representing number of boxes to to speed up neighbour search, if \code{NULL} an optimal number will be chosen \code{\link[nonlinearTseries:findAllNeighbours]{nonlinearTseries::findAllNeighbours()}} (default = \code{NULL}).}
}
\value{
FNN curve
}
\description{
Search for FNN to get an optimal Embedding Dimension using by using \code{\link[nonlinearTseries:findAllNeighbours]{nonlinearTseries::findAllNeighbours()}} in a loop.
}
