% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{get_fluct}
\alias{get_fluct}
\title{Get fluctuations (internal)}
\usage{
get_fluct(y_win, s)
}
\arguments{
\item{y_win}{a time series}

\item{s}{scale range}
}
\value{
fluctuation intensity
}
\description{
Get fluctuations (internal)
}
\keyword{internal}
