% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_command_line_crp.R
\name{get_os}
\alias{get_os}
\title{Which OS is running?}
\usage{
get_os()
}
\value{
A string, "osx", "windows", "linux"
}
\description{
Some systems not tested, but based on the cran page: \href{https://cran.r-project.org/web/checks/check_flavors.html}{check flavors}
}
