% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{getColours}
\alias{getColours}
\title{Get some nice colours}
\usage{
getColours(pal = c("pe", "mm", "le", "an")[1], Ncols = 20)
}
\arguments{
\item{pal}{The colour pallette, one of \code{"pe"},\code{"mm"},\code{"le"``"an"} (default = \code{"pe"})}

\item{Ncols}{Number of colours}
}
\value{
A list of colours
}
\description{
Get some nice colours
}
\examples{

getColours(Ncol=5)

}
