% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getPairs}
\alias{getPairs}
\title{Get all combinations}
\usage{
getPairs(x, y = NULL)
}
\arguments{
\item{x}{A vector with values.}

\item{y}{An optional vector with values}
}
\value{
a data frame with pairs of values
}
\description{
Get all combinations
}
\examples{

getPairs(x=1:6)

getPairs(x=1:3, y=c("A","B","C"))

}
