% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R, R/nlRtsa_SOURCE.old.R
\name{gg.theme}
\alias{gg.theme}
\alias{gg.theme}
\title{gg.theme}
\usage{
gg.theme(type = c("clean", "noax"), useArial = F, afmPATH = "~/Dropbox")

gg.theme(type = c("clean", "noax"), useArial = F, afmPATH = "~/Dropbox")
}
\arguments{
\item{type}{One of \code{"clean"}, or \code{"noax"}}

\item{useArial}{Use the Arial font (requires \code{.afm} font files in the \code{afmPath})}

\item{afmPATH}{Path to Arial \code{.afm} font files.}

\item{type}{One of \code{"clean"}, or \code{"noax"}}

\item{useArial}{Use the Arial font (requires \code{.afm} font files in the \code{afmPath})}

\item{afmPATH}{Path to Arial \code{.afm} font files.}
}
\value{
A theme for \code{ggplot2}.

A theme for \code{ggplot2}.
}
\description{
gg.theme

gg.theme
}
\details{
Will generate a \code{"clean"} ggplot theme, or a theme without any axes (\code{"noax"}).

Some scientific journals explicitly request the Arial font should be used in figures. This can be achieved by using \code{.afm} font format (see, e.g. http://www.pure-mac.com/font.html).

Will generate a \code{"clean"} ggplot theme, or a theme without any axes (\code{"noax"}).

Some scientific journals explicitly request the Arial font should be used in figures. This can be achieved by using \code{.afm} font format (see, e.g. http://www.pure-mac.com/font.html).
}
\examples{
library(ggplot2)
g <- ggplot(data.frame(x = rnorm(n = 100), y = rnorm(n = 100)), aes(x = x, y = y)) + geom_point()
g + gg.theme()
g + gg.theme("noax")
library(ggplot2)
g <- ggplot(data.frame(x = rnorm(n = 100), y = rnorm(n = 100)), aes(x = x, y = y)) + geom_point()
g + gg.theme()
g + gg.theme("noax")
}
