% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{gg_theme}
\alias{gg_theme}
\title{gg_theme}
\usage{
gg_theme(type = c("clean", "noax"))
}
\arguments{
\item{type}{One of \code{"clean"}, or \code{"noax"}}
}
\value{
A theme for \code{ggplot2}.
}
\description{
gg_theme
}
\details{
Will generate a \code{"clean"} ggplot theme, or a theme without any axes (\code{"noax"}).

Some scientific journals explicitly request the Arial font should be used in figures. This can be achieved by using \code{.afm} font format (see, e.g. http://www.pure-mac.com/font.html).
}
\examples{
library(ggplot2)
g <- ggplot(data.frame(x = rnorm(n = 100), y = rnorm(n = 100)), aes(x = x, y = y)) + geom_point()
g + gg_theme()
g + gg_theme("noax")
}
