% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{\%+-\%}
\alias{\%+-\%}
\title{Signed increment}
\usage{
counter \%+-\% increment
}
\arguments{
\item{counter}{If \code{counter} and \code{increment} are both a (signed) integers \code{counter} will change by the value of \code{increment}.}

\item{increment}{An integer value \eqn{\neq 0} to add to \code{counter}}
}
\description{
Increment an integer counter by an arbitrary (signed) interval.
}
\examples{

# Notice the difference between passing an object and a value for counter

# Value
(10 \%+-\% -5)
(10 \%+-\% -5)

# Object
i <- 10
(i \%+-\% -5)
(i \%+-\% -5)

# This means we can use the infix in a while ... statement
while(i > -3) i \%+-\% -5
# WARNING: As is the case for any while ... statement, be careful not to create an infinite loop!

}
\author{
Fred Hasselman
}
