% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{growth_ac}
\alias{growth_ac}
\title{Examples of dynamical growth models (maps)}
\usage{
growth_ac(
  Y0 = 0.01,
  r = 1,
  k = 1,
  N = 100,
  type = c("driving", "damping", "logistic", "vanGeert")[1]
)
}
\arguments{
\item{Y0}{Initial value.}

\item{r}{Growth rate parameter.}

\item{k}{Carrying capacity.}

\item{N}{Length of the time series.}

\item{type}{One of: "driving" (default), "damping", "logistic", "vanGeert1991".}
}
\value{
A timeseries object of length N.
}
\description{
Autocatlytic Growth: Iterating differential equations (maps)
}
\examples{
# The logistic map in the chaotic regime
growth_ac(Y0 = 0.01, r = 4, type = "logistic")
}
\seealso{
Other autocatalytic growth functions: 
\code{\link{growth_ac_cond}()}
}
\author{
Fred Hasselman
}
\concept{autocatalytic growth functions}
