% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Heaviside}
\alias{Heaviside}
\title{Heaviside step function}
\usage{
Heaviside(value)
}
\arguments{
\item{value}{value}
}
\value{
heaviside change
}
\description{
Heaviside step function
}
\keyword{internal}
