% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{HurvichDeo}
\alias{HurvichDeo}
\title{HDEst}
\usage{
HurvichDeo(nr, spec, A = 0.3, delta = 6/7)
}
\arguments{
\item{nr}{NFT}

\item{spec}{sdf}

\item{A}{A}

\item{delta}{delta}
}
\value{
Hurvich-Deo estimate
}
\description{
Latest version from discontinued package \code{fractal}.
}
\keyword{internal}
