% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{il_mi}
\alias{il_mi}
\title{Inter-layer mutual information}
\usage{
il_mi(g0, g1, probTable = FALSE)
}
\arguments{
\item{g0}{An igraph object representing a layer in a multiplex graph}

\item{g1}{An igraph object representing a layer in a multiplex graph}

\item{probTable}{Option to return the table with marginal and joint degree distribution probabilities (default = \code{TRUE})}
}
\value{
The inter-layer mutual information between \code{g1} and \code{g2}. If \code{probTable=TRUE}, a list object with two fields, the inter-layer mutual information and the table with marginal and joint degree distributions
}
\description{
Inter-layer mutual information
}
