% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{inf_MSE}
\alias{inf_MSE}
\title{Multi-Scale Entropy}
\usage{
inf_MSE(
  y,
  summaryFunction = c("mean", "median", "min", "max")[1],
  retainLength = FALSE,
  m = 2,
  r = 0.2,
  D = NA,
  fs = NULL,
  transformBefore = TRUE,
  removeTrend = c("no", "poly", "adaptive", "bridge")[1],
  polyOrder = 1,
  standardise = c("none", "mean.sd", "median.mad")[1],
  adjustSumOrder = FALSE,
  scaleMin = 1,
  scaleMax = floor(NROW(y)/10),
  scaleS = NA,
  overlap = 0,
  minData = 4,
  relativeEntropy = FALSE,
  doPlot = FALSE,
  returnPlot = FALSE,
  returnPLAW = FALSE,
  returnInfo = FALSE,
  silent = FALSE,
  noTitle = FALSE,
  tsName = "y"
)
}
\arguments{
\item{y}{A numeric vector}

\item{summaryFunction}{How should the data be summarized in the bins?}

\item{retainLength}{Return only the bin values (\code{FALSE}), or retain the length of the original series? (default = \code{TRUE})}

\item{m}{The size of the window in which tho evaluate whether a pattern repeats (default = \code{2})}

\item{r}{A factor that will determine the threshold for similarity of values, calculated as r x D (default = \code{0.2})}

\item{D}{Commonly the standard deviation of the time series, the similarity threshold will be calculated as r x D. Note that if the series is detrended and/or standardised and \code{D = NA} the standard deviation will be calculated after the transformations (default = \code{NA})}

\item{fs}{Sample rate}

\item{transformBefore}{Detrend/standardise before coarse graining. If set to \code{FALSE}, each coarsegrained series will be detrended/standardised separately (default = \code{TRUE})}

\item{removeTrend}{Method to use for detrending (default = "poly")}

\item{polyOrder}{Order of polynomial trend to remove if \code{removeTrend = "poly"}}

\item{standardise}{Standardise by the series using \code{\link[=ts_standardise]{ts_standardise()}} with \code{adjustN = FALSE} (default = "mean.sd")}

\item{adjustSumOrder}{Adjust the time series (summation or differencing), based on the global scaling exponent, see e.g. \url{https://www.frontiersin.org/files/Articles/23948/fphys-03-00141-r2/image_m/fphys-03-00141-t001.jpg}{Ihlen (2012)} (default = \code{FALSE})}

\item{scaleMin}{Minimium scale (as a power of 2) to use}

\item{scaleMax}{Maximum scale (as a power of 2) to use}

\item{scaleS}{If not \code{NA}, it should be a numeric vector listing the scales on which to evaluate the detrended fluctuations. Arguments \verb{scaleMax, scaleMin, scaleResolution} will be ignored.}

\item{overlap}{Turn DFA into a sliding window analysis. A number in \verb{[0 ... 1]} representing the amount of 'bin overlap'. If \code{length(y) = 1024} and overlap is \code{.5}, a scale of \code{4} will be considered a sliding window of size \code{4} with stepsize \code{floor(.5 * 4) = 2}. The detrended fluctuation in   For scale \code{128} this will be  (default = \code{0})}

\item{minData}{Minimum number of data points in a bin needed to calculate detrended fluctuation}

\item{relativeEntropy}{The relative entropy, SampEn / (-1 * log(1/length(y))) will be returned (default = \code{FALSE})}

\item{doPlot}{Return the log-log scale versus fluctuation plot with linear fit (default = \code{TRUE}).}

\item{returnPlot}{Return ggplot2 object (default = \code{FALSE})}

\item{returnPLAW}{Return the power law data (default = \code{FALSE})}

\item{returnInfo}{Return all the data used in DFA (default = \code{FALSE})}

\item{silent}{Silent-ish mode}

\item{noTitle}{Do not generate a title (only the subtitle)}

\item{tsName}{Name of y added as a subtitle to the plot}
}
\value{
MSE
}
\description{
Calculate the Multi-Scale (Sample) Entropy of a time series.
}
\note{
The transformation settings (detrending and/or normalisation), \code{transformBefore} will determine the value of the product r x D if \code{D = NA}.
}
\examples{

y <- rnorm(100)

out <- inf_MSE(y, r=.5, scaleMin=1, scaleMax=10, returnInfo = TRUE, doPlot = FALSE)


}
\seealso{
Other Information based complexity measures: 
\code{\link{inf_SampEn}()}
}
\concept{Information based complexity measures}
