% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irn.R
\name{irn_crossClustering}
\alias{irn_crossClustering}
\title{Cross CLustering Coefficient}
\usage{
irn_crossClustering(Ax, Ay, Axy, local = FALSE, global = TRUE)
}
\arguments{
\item{Ax}{Adjacency matrix x}

\item{Ay}{Adjacency matrix y}

\item{Axy}{Adjacency matrix xy}

\item{local}{local CC}

\item{global}{global CC}
}
\value{
list object with local and global CC
}
\description{
Cross CLustering Coefficient
}
