% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irn.R
\name{irn_crossDegree}
\alias{irn_crossDegree}
\title{Cross Degree}
\usage{
irn_crossDegree(Axy)
}
\arguments{
\item{Axy}{Adjacency matrix xy}
}
\value{
list object wih local and global cross-degree
}
\description{
Cross Degree
}
