% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irn.R
\name{irn_crossTriples}
\alias{irn_crossTriples}
\title{Cross Triples}
\usage{
irn_crossTriples(Ax, Ay, Axy)
}
\arguments{
\item{Ax}{Adjacency matrix x}

\item{Ay}{Adjacency matrix y}

\item{Axy}{Adjacency matrix xy}
}
\value{
list object with triasngles and triads
}
\description{
Cross Triples
}
