% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irn.R
\name{irn_directedCrossTriples}
\alias{irn_directedCrossTriples}
\title{Directed Cross Triples}
\usage{
irn_directedCrossTriples(Ax, Ay, Axy)
}
\arguments{
\item{Ax}{Adjacency matrix x}

\item{Ay}{Adjacency matrix y}

\item{Axy}{Adjacency matrix xy}
}
\value{
list object with triangles and triads
}
\description{
Directed Cross Triples
}
