% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irn.R
\name{irn_plot}
\alias{irn_plot}
\title{Inter system recurrence networks}
\usage{
irn_plot(
  RMxy,
  gx = NULL,
  gy = NULL,
  merge_seed = 5,
  curvature = 0,
  angle = 0,
  scaleEdgeSize = 1,
  alphaE = 0.1,
  alphaV = 1,
  colorExy = "grey30",
  colorEyx = "grey70",
  sizeEbyCC = FALSE,
  doPlot = TRUE,
  ggplotReturn = FALSE,
  igraphReturn = FALSE
)
}
\arguments{
\item{RMxy}{A cross recurrence matrix of x and y}

\item{gx}{An igraph object of recurrence network x}

\item{gy}{An igraph object of recurrence network y}

\item{merge_seed}{Seed}

\item{curvature}{The \code{curvature} parameter for edges see \code{\link[=geom_curve]{geom_curve()}} (default = \code{-0.7})}

\item{angle}{The \code{angle} parameter for edges see \code{\link[=geom_curve]{geom_curve()}} (default = \code{90})}

\item{scaleEdgeSize}{Scale the size of the edges by a constant: \code{E(g)$width * scaleEdgeSize} (default = \code{1/5})}

\item{alphaE}{Set transparency for Edges. A single numeric, or a vector of length \code{ecount(g)} (default = \code{0.8})}

\item{alphaV}{Set transparency for Vertices (default = \code{1})}

\item{colorExy}{Edge colour x >> y}

\item{colorEyx}{Edge colour y >> x}

\item{sizeEbyCC}{Node size by Clustering Coefficient}

\item{doPlot}{Produce a plot? (default = \code{TRUE})}

\item{ggplotReturn}{returns the ggplot object (default = \code{FALSE})}

\item{igraphReturn}{returns the intermediate iGraph object. This will not look the same as the final graph, but has most of the attributes, like edge and vertex colors and spiral layout (default = \code{FALSE})}
}
\value{
A plot
}
\description{
Inter system recurrence networks
}
