% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Test for wholenumber}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{x}

\item{tol}{tol}
}
\value{
logical vector
}
\description{
Test for wholenumber
}
\keyword{internal}
