% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{lagEmbed}
\alias{lagEmbed}
\title{lag Embed a time series}
\usage{
lagEmbed(y, emDim, emLag)
}
\arguments{
\item{y}{Time series}

\item{emDim}{Embedding dimension}

\item{emLag}{Embedding lag}
}
\value{
The lag embedded time series
}
\description{
lag Embed a time series
}
