% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{logScale}
\alias{logScale}
\title{logScale}
\usage{
logScale(
  scale.min,
  scale.max,
  scale.ratio = 2,
  scale.res = NULL,
  coerce = NULL
)
}
\arguments{
\item{scale.min}{min}

\item{scale.max}{max}

\item{scale.ratio}{ratio}

\item{scale.res}{reolution}

\item{coerce}{coerce?}
}
\value{
output
}
\description{
logScale
}
\note{
Based on "logScale" from discontinued package ifultools.
}
\keyword{internal}
