% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{lv_Ndim}
\alias{lv_Ndim}
\title{Lotka-Volterra model for N species}
\usage{
lv_Ndim(
  Nsim = 1000,
  Ndim = 4,
  Y0 = rep(0.1, Ndim),
  r = c(1, 0.72, 1.53, 1.27),
  A = matrix(c(1, 1.09, 1.52, 0, 0, 1, 0.44, 1.36, 2.33, 0, 1, 0.47, 1.21, 0.51, 0.35,
    1), nrow = 4, byrow = TRUE),
  K = rep(1, Ndim),
  h = 0.5
)
}
\arguments{
\item{Nsim}{How many data points to simulate (default = \code{1000})}

\item{Ndim}{How many dimensions (coupled L-V equations) to use (default = \code{4})}

\item{Y0}{A vector with \code{Ndim} elements representing the initial value (default = \code{rep(.1, Ndim)})}

\item{r}{A vector with \code{Ndim} elements representing the growth rates for each dimension. The default settings are taken from  Vano et al. (2006)  (default = \code{c(1, 0.72, 1.53, 1.27)})}

\item{A}{The interaction matrix of \verb{Ndim X Ndim} representing the nature of the coupling between each dimension. \code{A[1,2]} will appear as a parameter in the  equation for \code{Y1} setting the magnitude of the interaction with \code{Y2}. Conversely \code{A[2,1]} will appear in the equation for \code{Y2} setting the magnitude of the interaction with \code{Y1}. The diagonal elements have to be \code{1}. The default settings are taken from  Vano et al. (2006)}

\item{K}{A vector with \code{Ndim} elements representing the carrying capacity  for each dimension. (default = \code{rep(1, Ndim)})}

\item{h}{The Euler parameter for RK4 integration.}

\item{returnLongData}{Return the data in long format for easy plotting (default = \code{FALSE})}
}
\value{
A matrix of \verb{Ndim X Nsim} with simulated values.
}
\description{
The default simulates a system of \code{Ndim = 4} coupled competitive Lotka-Volterra equations studied by Vano et al. (2006) using RK4 numerical integration. Vano et al. describe the dynamics of the resulting attractor as chaotic (bounded, quasi-periodic, sensitive dependence on initial conditions), also see \href{https://en.wikipedia.org/wiki/Competitive_Lotka\%E2\%80\%93Volterra_equations#4-dimensional_example}{this Wiki page}.
}
\examples{

library(plot3D)

Y <- lv_Ndim()
df <- as.data.frame(apply(t(Y), 2, ts_standardise))
lines3D(df$Y1, df$Y2, df$Y3, colvar = df$Y4, clab = "Y4", xlab= "Y1", ylab = "Y2", zlab ="Y3", ticktype = "detailed")

}
\references{
Vano, J. A., Wildenberg, J. C., Anderson, M. B., Noel, J. K., & Sprott, J. C. (2006). Chaos in low-dimensional Lotka–Volterra models of competition. \emph{Nonlinearity, 19}(10), 2391.
}
