% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOps.R
\name{mat_coursegrain}
\alias{mat_coursegrain}
\title{Course grain a matrix for plotting}
\usage{
mat_coursegrain(
  RM,
  target_height = round(NROW(RM)/2),
  target_width = round(NCOL(RM)/2),
  summary_func = NA,
  recurrence_threshold = NA,
  categorical = NA,
  output_type = 0,
  n_core = 1,
  silent = FALSE
)
}
\arguments{
\item{RM}{A (recurrence) matrix}

\item{target_height}{How many rows? (default = \code{NROW(RM)/2})}

\item{target_width}{How many columns? (default = \code{NCOL(RM)/2})}

\item{summary_func}{How to summarise the values in subset \code{X} of \code{RM}. If set to \code{NA}, the function will try to pick a summary function based on the cell values: If \code{RM} is a distance matrix,  \code{mean(X, na.rm = TRUE)} will be used; If it is a binary matrix \code{ifelse(mean(X, na.rm = TRUE)>recurrence_threshold,1,0)}, a categorical matrix (\code{categorical = TRUE}, or, matrix attribute \code{chromatic = TRUE}) will pick the most frequent category in the subset \code{attributes(ftable(X))$col.vars$x[[which.max(ftable(X))]]}. (default = \code{NA})}

\item{recurrence_threshold}{For a binary matrix the mean of the cells to be summarised will vary between \code{0} and \code{1}, which essentially represents the recurrence rate for that subset of the matrix. If \code{NA} the threshold will be set to a value that in most cases should return a plot with a similar \code{RR} as the original plot. (default = \code{NA})}

\item{categorical}{If set to \code{TRUE}, will force \code{summary_func} to select the most frequent value. If \code{NA} the matrix attribute \code{chromatic} will be used. If \code{chromatic} is not present, all values in the matrix have to be whole numbers as determined by \code{plyr::is.discrete()}. (default = \code{NA})}

\item{output_type}{The output format for \code{plyr::vapply()}. (default = \code{0.0})}

\item{n_core}{Number of cores for parallel processing. Set to \code{NA} to automatically choose cores. (default = \code{1})}

\item{silent}{Silt-ish mode (default = \code{FALSE})}
}
\value{
A coursegrained matrix of size \code{target_width} by \code{target_height}.
}
\description{
Course grain a matrix for plotting
}
\note{
This code was inspired by code published in a blog post by Guillaume Devailly on 29-04-2020 (https://gdevailly.netlify.app/post/plotting-big-matrices-in-r/)
}
\examples{

# Continuous
RMc1 <- rp(cumsum(rnorm(200)))
rp_plot(RMc1)
RMc2 <- mat_coursegrain(RMc1)
rp_plot(RMc2)

# Binary
RMb1 <- rp(cumsum(rnorm(200)), emRad = NA)
rp_plot(RMb1, plotMeasures = TRUE)
# Reported RQA measures in rp_plot will be based on the full matrix
rp_plot(RMb1, maxSize = 100^2, plotMeasures = TRUE)
# Plotting the coursegrained matrix itself will yield different values
RMb2 <- mat_coursegrain(RMb1)
rp_plot(RMb2, plotMeasures = TRUE)

# Categorical
RMl1 <- rp(y1 = round(runif(100, min = 1, max = 3)), chromatic = TRUE)
rp_plot(RMl1)
RMl2 <- mat_coursegrain(RMl1, categorical = TRUE)
rp_plot(RMl2)

}
