% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOps.R
\name{mat_hamming}
\alias{mat_hamming}
\title{Calculate Hamming distance}
\usage{
mat_hamming(X, Y = NULL, embedded = TRUE)
}
\arguments{
\item{X}{A matrix (of coordinates)}

\item{Y}{A matrix (of coordinates)}

\item{embedded}{Do X and/or Y represent surrogate dimensions of an embedded time series?}
}
\value{
A hamming-distance matrix of X, or X and Y. Useful for ordered and unordered categorical data.
}
\description{
Calculate Hamming distance
}
\seealso{
Other Distance matrix operations (recurrence plot): 
\code{\link{bandReplace}()},
\code{\link{mat_di2bi}()},
\code{\link{mat_di2ch}()},
\code{\link{mat_di2we}()},
\code{\link{rp_lineDist}()},
\code{\link{rp_nzdiags}()},
\code{\link{rp_plot}()},
\code{\link{rp_size}()},
\code{\link{rp}()}
}
\author{
Fred Hasselman
}
\concept{Distance matrix operations (recurrence plot)}
