% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOps.R
\name{mat_ind}
\alias{mat_ind}
\title{Get indices of matrix diagonals, rows, or columns}
\usage{
mat_ind(Xlength, Ylength, index, diagonal = FALSE)
}
\arguments{
\item{Xlength}{X dim}

\item{Ylength}{Y dim}

\item{index}{index of diagonal, row or column}

\item{diagonal}{diagonal}

\item{horizontal}{horizontal}
}
\value{
list
}
\description{
Get indices of matrix diagonals, rows, or columns
}
