% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOps.R
\name{mat_nodeDegree}
\alias{mat_nodeDegree}
\title{Matrix node degree}
\usage{
mat_nodeDegree(adjMAT, mode = c("undirected", "in", "out")[1])
}
\arguments{
\item{adjMAT}{Adjacency matrix}

\item{mode}{Type of degree}
}
\value{
Vector of node degrees
}
\description{
Matrix node degree
}
