% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrn.R
\name{mi_interlayer}
\alias{mi_interlayer}
\title{Inter-layer mutual information}
\usage{
mi_interlayer(g0, g1, probTable = FALSE)
}
\arguments{
\item{g0}{An igraph object representing a layer in a multiplex graph}

\item{g1}{An igraph object representing a layer in a multiplex graph}

\item{probTable}{Option to return the table with marginal and joint degree distribution probabilities (default = \code{TRUE})}
}
\value{
The inter-layer mutual information between \code{g1} and \code{g2}. If \code{probTable=TRUE}, a list object with two fields, the inter-layer mutual information and the table with marginal and joint degree distributions
}
\description{
Inter-layer mutual information
}
\note{
If the networks are weighted the strength distribution will be used instead of the the degree distribution.
}
\seealso{
Other Redundancy measures (mutual information): 
\code{\link{mi_mat}()},
\code{\link{mif}()}
}
\concept{Multiplex Recurrence Networks}
\concept{Redundancy measures (mutual information)}
