% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{mi_mat}
\alias{mi_mat}
\title{Mutual Information variations}
\usage{
mi_mat(y, ID1, ID2)
}
\arguments{
\item{y}{Matrix}

\item{ID1}{ids}

\item{ID2}{ids}
}
\value{
mi
}
\description{
Mutual Information variations
}
