% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrn.R
\name{mrn}
\alias{mrn}
\title{Multiplex Recurrence Network}
\usage{
mrn(
  layers,
  MRNweightedBy = c("InterLayerMI", "InterLayerCor", "EdgeOverlap", "JRP")[1],
  jrp_measure = "RR",
  win = NA,
  step = NA,
  overlap = NA,
  alignment = "r",
  cumulative = FALSE,
  doPlot = FALSE,
  silent = TRUE
)
}
\arguments{
\item{layers}{A list of igraph objects representing the layers of the multiplex network. The layer networks must all have the same number of vertices.}

\item{MRNweightedBy}{The measure to be used to evaluate the average structural similarities between the layers of the network. Valid options are: \code{"InterLayerMI"} (Mutual information based on similarity of the vertex degree across layers), \code{"InterLayerCor"} (pearson correlation between vertex degree or strength of each layer pair),  \code{"EdgeOverlap"} (proportion of vertices sharing the same edges across layers), or \code{"JRP"} which will calculate a joint recurrence plot for each layer pair and add the value passed to \code{jrp_measure} as the measure of association (default = \code{InterLayerMI})}

\item{jrp_measure}{Which JRP measure should be used to asses layer similarity. Use a column name of the data frame output by \link{rp_measures} (default = \code{"RR"})}

\item{win}{The window size passed to \code{\link[=ts_windower]{ts_windower()}} in which to evaluate \code{"InterLayerMI"}, \code{"InterLayerCor"}, \code{"EdgeOvelap"}, or "\code{JRP}". (default = \code{NA}).}

\item{step}{The stepsize for the sliding window (default = \code{NA}).}

\item{overlap}{The window overlap passed to \code{\link[=ts_windower]{ts_windower()}} if \code{MRNweightedBy} is \code{"InterLayerMI"} or \code{"EdgeOvelap"}. The value of \code{step} will be ignored if \code{overlap} is not \code{NA}. (default = \code{NA}).}

\item{alignment}{Whether to right (\code{"r"}), center (\code{"c"}), or left (\code{"l"}) align the window.}

\item{cumulative}{To make the network represent cumulative time, set \code{directed = TRUE} and \code{cumulative = TRUE}. This will set the upper triangle of the recurrence matrix to \code{0} and ensures that the network edges represent recurrent values that have occurred in the \code{past} relative to the current observed value (node). If \code{directed = FALSE} the argument is ignored (default = \code{TRUE}).}

\item{doPlot}{Plot the multiplex recurrence network (default = \code{FALSE}).}

\item{silent}{Silent-ish mode. (default = FALSE).}
}
\value{
A list object with a (windowed) Multiplex Recurrence Matrix based on \code{MRNweightedBy}:
\itemize{
\item \emph{interlayerMI} - One or more matrices with edge weights between layers that represent the interlayer Mutual Information.
\item \emph{interlayerCor} - One or more matrices with edge weights between layers that represent the Pearson correlation between vertex degrees (or strength if layers are weighted graphs) of layers.
\item \emph{edgeOverlap} - One or more matrices with edge weights between layers that represent the overlapping edges between layers.
\item \emph{jointRP} -  One or more matrices with edge weights between layers that represent the value of the joint RP passed in \code{jrp_measure}. The default is the \code{"RR"} of joint Recurrence Plot of the two layers.
\item The field \emph{meanValues} contains one or more matrices that represent the means and SDs of the weights requested in\code{MRNweightedBy}. The measure \code{eo_joint} (returned with \code{"edgeOverlap"}) refers to the number of edges shared among \emph{all} layers of the MRN.
}
}
\description{
This function will create a Multiplex Recurrence Network from a list of \link{igraph} objects that can be considered the layers of a network. The layers must have the same number of nodes. There are two modes of operation: \emph{Layer similarity} (\code{MRNweightedBy} is set to \code{"InterLayerMI"}, \code{"InterLayerCor"}, \code{"EdgeOvelap"}, or "\code{JRP}") and \emph{Layer importance} (not implemented yet). The former generates weighted MRN based on \emph{Interlayer Mutual Information}, \emph{Interlayer Correlation}, \emph{Edge Overlap}, \emph{Joint Recurrence Rate}.
}
\examples{

# Create some layers
library(igraph)

layers <- list(g1 = igraph::sample_smallworld(1, 100, 5, 0.05),
g2 = igraph::sample_smallworld(1, 100, 5, 0.5),
g3 = igraph::sample_smallworld(1, 100, 5, 1))

mrn(layers = layers)


}
