% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{noise_fBm}
\alias{noise_fBm}
\title{Generate fractional Brownian motion}
\usage{
noise_fBm(H = 1.5, N = 512, mu = NULL, sigma = NULL)
}
\arguments{
\item{H}{Hurst exponent}

\item{N}{Length of noise series}

\item{mu}{Mean}

\item{sigma}{SD}
}
\value{
fBm
}
\description{
Generate fractional Brownian motion
}
