% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{noise_fGn}
\alias{noise_fGn}
\title{Generate fractional Gaussian noise}
\usage{
noise_fGn(H = 0.5, N = 512, mu = NULL, sigma = NULL)
}
\arguments{
\item{H}{Hurst exponent}

\item{N}{Length of noise series}

\item{mu}{Mean}

\item{sigma}{SD}
}
\value{
fGn
}
\description{
Generate fractional Gaussian noise
}
