% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{noise_powerlaw}
\alias{noise_powerlaw}
\title{Generate noise series with power law scaling exponent}
\usage{
noise_powerlaw(y = NULL, alpha = -1, N = 100, standardise = FALSE,
  randomPower = FALSE)
}
\arguments{
\item{y}{Time series to use as a 'model'. If specified, \code{N} will be \code{N = length(y)}, and the series will be constructed based on \code{stats::fft(y)}.}

\item{alpha}{The log-log spectral slope, the scaling exponent.}

\item{N}{Length of the time series}

\item{randomPower}{If \code{TRUE} phases will be deterministic, uniformly distributed in \code{[-pi,pi]}. If \code{FALSE}, the spectrum will be stochastic with a Chi-square distribution. If \code{y} is not \code{NULL} this argument will be ignored.}

\item{normalise}{Forces scaling of the output to the range \code{[-1, 1]}, consequently the power law will not necessarily extend right down to \code{0Hz}.}
}
\value{
Time series with a power law of alpha.
}
\description{
Generate noise series with power law scaling exponent
}
\note{
Adapted from a Matlab script called \code{powernoise.m}. The script contained the following commented text:

With no option strings specified, the power spectrum is
}
