% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{pacf_fisherZ}
\alias{pacf_fisherZ}
\title{FisherZ for PACF}
\usage{
pacf_fisherZ(r, n, lag, siglevel = 0.05, sides = 2, type = "")
}
\arguments{
\item{r}{r}

\item{n}{n}

\item{lag}{lag}

\item{siglevel}{siglevel}

\item{sides}{sides}

\item{type}{type}
}
\value{
dataframe
}
\description{
FisherZ for PACF
}
\keyword{internal}
