% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotDC_ccp}
\alias{plotDC_ccp}
\title{Plot Cumulative Complexity Peaks}
\usage{
plotDC_ccp(
  df_ccp,
  win,
  useVarNames = TRUE,
  colOrder = TRUE,
  useTimeVector = NA,
  timeStamp = "31-01-1999",
  doPlot = TRUE,
  title = "Critical Instability Plot",
  subtitle = "",
  xlabel = "Time",
  ylabel = ""
)
}
\arguments{
\item{df_ccp}{A dataframe generated by \code{dc_ccp()}}

\item{win}{Size of window in which to calculate Dynamic Complexity. If \code{win < NROW(df)} the window will move along the time series with a stepsize of \code{1} (default = \code{NROW(df)})}

\item{useVarNames}{Use the column names of \code{df} as variable names in the Complexity Resonance Diagram (default = \code{TRUE})}

\item{colOrder}{If \code{TRUE}, the order of the columns in \code{df} determines the of variables on the y-axis. Use \code{FALSE} for alphabetic/numeric order. Use \code{NA} to sort by by mean value of Dynamic Complexity (default = \code{TRUE})}

\item{useTimeVector}{Parameter used for plotting. A vector of length \code{NROW(df)}, containing date/time information (default = \code{NA})}

\item{timeStamp}{If \code{useTimeVector} is not \code{NA}, a character string that can be passed to \code{\link[lubridate:stamp]{lubridate::stamp()}} to format the the dates/times passed in \code{useTimeVector} (default = \code{"01-01-1999"})}

\item{doPlot}{If \code{TRUE} shows a Complexity Resonance Diagram of the Dynamic Complexity and returns an invisible \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object. (default = \code{FALSE})}

\item{title}{A title for the plot.}

\item{subtitle}{A subtitle for the plot.}

\item{xlabel}{A label for the x-axis.}

\item{ylabel}{A label for the y-axis.}
}
\value{
An invisible ggplot2 object.
}
\description{
Plot Cumulative Complexity Peaks
}
\seealso{
Other Dynamic Complexity functions: 
\code{\link{dc_ccp}()},
\code{\link{dc_d}()},
\code{\link{dc_f}()},
\code{\link{dc_win}()},
\code{\link{plotDC_lvl}()},
\code{\link{plotDC_res}()}
}
\concept{Dynamic Complexity functions}
