% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotDC_helper}
\alias{plotDC_helper}
\title{Create breaks for Resonance plots}
\usage{
plotDC_helper(useTimeVector, timeStamp, win, df_DC, trimFirstWin = TRUE)
}
\arguments{
\item{useTimeVector}{useTimeVector}

\item{df_DC}{df_DC}
}
\value{
breaks
}
\description{
Create breaks for Resonance plots
}
\keyword{internal}
