% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotMRN_win}
\alias{plotMRN_win}
\title{Plot windowed Multiplex Recurrence Network measures}
\usage{
plotMRN_win(
  df_mrn,
  layerMeasures = c("InterLayerMI", "InterLayerCor", "EdgeOverlap", "JRP")[1],
  vertexMeasures = c("none", "degree", "strength", "transitivity", "closeness",
    "betweenness")[1],
  weighted = FALSE,
  directed = FALSE,
  cumulative = FALSE,
  plotSD = FALSE,
  doPlot = TRUE
)
}
\arguments{
\item{df_mrn}{Output from function \code{\link[=mrn]{mrn()}} with arguments set for a windowed analysis.}

\item{layerMeasures}{Character vector indicating which layer similarity measure(s) should be plotted. Valid elements in the vector are: \code{"InterLayerMI"}, \code{"InterLayerCor"}, \code{"EdgeOverlap"}, \code{"JRP"}.}

\item{vertexMeasures}{Character vector indicating which vertex measure(s) should be plotted.  Valid elements in the vector are: \code{"none"}, \code{"degree"},\code{"strength"},\code{"transitivity"} (= clustering),\code{"closeness"},\code{"betweenness"}. See the \link{igraph} The calculations  depend on the value of \code{MRNweightedBy} that was used with function \code{\link[=mrn]{mrn()}}.}

\item{weighted}{If \code{vertexMeasures} is not equal to \code{"none"}, should the graph be considered weighted? (default = \code{FALSE})}

\item{directed}{If \code{vertexMeasures} is not equal to \code{"none"}, should the graph be considered directed? (default = \code{FALSE})}

\item{cumulative}{If \code{vertexMeasures} is not equal to \code{"none"}, should the graph be considered cumulative? This will set the mode to \code{"upper"} (default = \code{FALSE})}

\item{plotSD}{If \code{TRUE}, a ribbon with range mean ± SD will be plotted (default = \code{FALSE})}

\item{doPlot}{Plot the igraph object. If \code{FALSE}, a just the graph object will be returned invisible.}
}
\value{
a ggplot object
}
\description{
Plot windowed Multiplex Recurrence Network measures
}
\seealso{
Other Tools for windowed analyses: 
\code{\link{ts_windower}()}
}
\concept{Tools for plotting multiplex recurrence networks}
\concept{Tools for windowed analyses}
