% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotNET_BA}
\alias{plotNET_BA}
\title{Example of Barabasi scale-free network}
\usage{
plotNET_BA(n = 100, pwr = 1, out.dist = NULL, doPlot = TRUE)
}
\arguments{
\item{n}{Number of vertices}

\item{pwr}{Power of preferential attachment}

\item{out.dist}{Degree distribution}

\item{doPlot}{Plot the igraph object}
}
\value{
A Barabasi scale-free igraph object
}
\description{
A wrapper around \code{\link[igraph:sample_pa]{igraph::sample_pa()}}
}
\seealso{
\code{\link[igraph:sample_pa]{igraph::sample_pa()}}

Other tools for plotting networks: 
\code{\link{plotNET_SW}()},
\code{\link{plotNET_groupColour}()},
\code{\link{plotNET_groupWeight}()},
\code{\link{plotNET_prep}()}
}
\concept{tools for plotting networks}
