% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotNET_groupColour}
\alias{plotNET_groupColour}
\title{Vertex and Edge Group Colours}
\usage{
plotNET_groupColour(
  g,
  groups,
  colourV = TRUE,
  alphaV = 1,
  colourE = FALSE,
  alphaE = 0.8,
  groupColours = NULL,
  defaultEdgeColour = "grey70",
  doPlot = TRUE
)
}
\arguments{
\item{g}{An igraph object}

\item{groups}{A named numeric vector with \code{length(V(g))} integers representing each group, or, a named character vector describing each group. If \code{names(groups)==NULL} then the names of the vector will be set as \code{names(groups) == V(g)$name}. If \code{V(g)$name==NULL}, the names of the vector will be set by the Vertex index}

\item{colourV}{Colour Vertices based on \code{groups} (default = \code{TRUE})}

\item{alphaV}{Set transparency for Vertices (default = \code{1})}

\item{colourE}{Colour Edges based on \code{groups}. Edges connecting to vertices of the same group will be coloured as the group (default = \code{FALSE})}

\item{alphaE}{Set transparency for Edges. A single numeric, or a vector of length \code{ecount(g)} (default = \code{0.8})}

\item{groupColours}{A list of length \code{groups} with valid colour codes}

\item{defaultEdgeColour}{Default edge colour}

\item{doPlot}{Plot the igraph object}
}
\value{
An igraph object with vertices and/or edges coloured by groups listed in \code{groups}
}
\description{
Identify Vertex and/or Edge groups by colour.
}
\seealso{
Other tools for plotting networks: 
\code{\link{plotNET_BA}()},
\code{\link{plotNET_SW}()},
\code{\link{plotNET_groupWeight}()},
\code{\link{plotNET_prep}()}
}
\concept{tools for plotting networks}
