% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{plotNET_prep}
\alias{plotNET_prep}
\title{Plot Network Based on RQA}
\usage{
plotNET_prep(g, labels = NA, nodesize = c("degree", "hubscore")[1],
  edgeweight = "weight")
}
\arguments{
\item{g}{An igraph object}

\item{labels}{Vertex labels}

\item{nodesize}{Set nodesizes by \code{degree(g, normalised = TRUE)} (default) or \code{hubscore(g)$vector}. If a numeric value is passed all vertex sizes will be set to that value.}

\item{edgeweight}{Set size of edges to \code{"E(g)$weight"} by passing "weight". If a single numeric value is provided all edges will be set to that value.}
}
\value{
an igraph object
}
\description{
Plot Network Based on RQA
}
\seealso{
Other tools for plotting networks: \code{\link{plotNET_BA}},
  \code{\link{plotNET_SW}},
  \code{\link{plotNET_groupColour}},
  \code{\link{plotNET_groupWeight}}
}
