% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotNET_SW}
\alias{plotNET_SW}
\title{Example of Strogatz-Watts small-world network}
\usage{
plotNET_SW(n = 100, k = 5, p = 0.05, doPlot = TRUE)
}
\arguments{
\item{n}{Size of the lattice (integer)}

\item{k}{Neighbourhood size (integer)}

\item{p}{Rewiring probability (between \code{0} and \code{1})}

\item{doPlot}{PLot the igraph object}
}
\value{
A Strogatz-Watts small-world igraph object
}
\description{
A wrapper around \code{\link[igraph:sample_smallworld]{igraph::sample_smallworld()}} with \code{dim=1}
}
\seealso{
\code{\link[igraph:sample_smallworld]{igraph::sample_smallworld()}}

Other tools for plotting networks: 
\code{\link{plotNET_BA}()},
\code{\link{plotNET_groupColour}()},
\code{\link{plotNET_groupWeight}()},
\code{\link{plotNET_prep}()}
}
\concept{tools for plotting networks}
