% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotRED_acf}
\alias{plotRED_acf}
\title{Plot ACF and PACF}
\usage{
plotRED_acf(
  y,
  Lmax = max(round(NROW(y)/4), 10),
  alpha = 0.05,
  doPlot = TRUE,
  returnCorFun = FALSE
)
}
\arguments{
\item{y}{A time series or numeric vector}

\item{Lmax}{Maximum number of lags}

\item{alpha}{Significance level}

\item{doPlot}{Plot output}

\item{returnCorFun}{Return the data}
}
\value{
Either an invisible ggplot2 object r a list containing the plot and the data
}
\description{
Plot ACF and PACF
}
\seealso{
Other Plot redundancy functions: 
\code{\link{plotRED_mif}()}
}
\concept{Plot redundancy functions}
