% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{plotRED_mif}
\alias{plotRED_mif}
\title{Plot various MI functions}
\usage{
plotRED_mif(y, lags = max(round(NROW(y)/4), 10), alpha = 0.05,
  doPlot = TRUE, returnMIFun = TRUE)
}
\arguments{
\item{y}{A \code{Nx1} matrix for auto-mif, a \code{Nx2} matrix or data frame for cross-mif, a \code{Nx3} matrix or data frame for mif between col 1 and 2 conditional on col 3; or a \code{NxM} matrix or data frame for the multi-information function.}

\item{lags}{Maximum number of lags}

\item{alpha}{Significance level}

\item{doPlot}{Plot output}

\item{returnMIFun}{Return the data}
}
\value{
Either an invisible ggplot2 object r a list containing the plot and the data
}
\description{
Plot various MI functions
}
\seealso{
Other Plot redundancy functions: \code{\link{plotRED_acf}}
}
