% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotRN_phaseDensities}
\alias{plotRN_phaseDensities}
\title{Phase Density for each dimension}
\usage{
plotRN_phaseDensities(
  phaseOutput,
  plotCentroid = NA,
  showEpochLegend = TRUE,
  epochColours = NULL,
  epochLabel = "Phase",
  excludeTransients = FALSE,
  excludePhaseNeighbours = FALSE,
  excludeSingularities = TRUE,
  excludeNonrecurring = TRUE,
  excludeVars = "",
  excludePhases = "",
  alphaDensity = 0.4,
  splitFacets = NA,
  showPhaseSize = TRUE,
  returnGraph = FALSE
)
}
\arguments{
\item{phaseOutput}{Output from function \link{rn_phaseInfo}}

\item{excludeTransients}{Should the category "Transient" be excluded from plots? (default = \code{FALSE})}

\item{excludeSingularities}{Should the category "Singularity" be excluded from plots? (default = \code{TRUE})}

\item{excludeNonrecurring}{Should the category "Nonrecurring" be excluded from plots? (default = \code{TRUE})}

\item{excludeVars}{Exclude specific dimension variables by name. Leave empty to include all variables (default = \code{""})}

\item{excludePhases}{Exclude Phases by their name (variable \code{phase_name}). Leave empty to include all Phases (after the other exclusion arguments) (default = \code{""})}

\item{alphaDensity}{Alpha value for the density plots}

\item{splitFacets}{Integer value to indicate if sets of phases should be displayed in different facets? Pass an integer equalt to, or larger than the number of phases found to display all phases for which a density can be calculated (default = \code{NA})}

\item{showPhaseSize}{Show the number of states in each phase in the labels (default = TRUE)}

\item{excludePhaseNeighbour}{Should the category "PhaseNH" be excluded from plots? (default = \code{FALSE})}
}
\value{
data frame with phase series.
}
\description{
Create a ridge plot (requires package \link[ggridges:ggridges-package]{ggridges::ggridges})
}
\examples{

RN <- rn(cumsum(rnorm(100)), emDim = 1, emLag = 1, emRad = NA, weighted = TRUE)
outPhases <- rn_phaseInfo(RN)
plotRN_phaseDensities(outPhases)

}
