% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotRN_phaseDensity}
\alias{plotRN_phaseDensity}
\title{Phase Density for each dimension}
\usage{
plotRN_phaseDensity(
  phaseOutput,
  plotCentroid = NA,
  excludeOther = FALSE,
  excludeNorec = FALSE,
  excludeVars = "",
  excludePhases = "",
  returnGraph = FALSE
)
}
\arguments{
\item{phaseOutput}{Output from function \link{rn_phases}}

\item{excludeOther}{Exclude the default Phase "Other"}

\item{excludeNorec}{Exclude the default Phase "No recurrence"}

\item{excludeVars}{Exclude specific dimension variables by name. Leave empty to include all variables (default = \code{""})}

\item{excludePhases}{Exclude Phases by their name (variable \code{phase_name}). Leave empty to include all Phases (after \code{excludeOther} and \code{excludeNorec}) (default = \code{""})}
}
\value{
data frame with phase series.
}
\description{
Create a ridge plot (requires package \link{ggridges})
}
\examples{

RN <- rn(cumsum(rnorm(100)), emDim = 1, emLag = 1, emRad = NA, weighted = TRUE)
outPhases <- rn_phases(RN)


}
