% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotRN_phaseProfile}
\alias{plotRN_phaseProfile}
\title{Profile Plot}
\usage{
plotRN_phaseProfile(
  phaseOutput,
  plotCentroid = NA,
  showEpochLegend = TRUE,
  epochColours = NULL,
  epochLabel = "Phase",
  excludeOther = FALSE,
  excludeNorec = TRUE,
  excludeVars = "",
  excludePhases = "",
  returnGraph = FALSE,
  colOrder = NA
)
}
\arguments{
\item{phaseOutput}{Output from function \link{rn_phases}}

\item{excludeOther}{Exclude the default Phase "Other"}

\item{excludeNorec}{Exclude the default Phase "No recurrence"}

\item{excludeVars}{Exclude specific dimension variables by name. Leave empty to include all variables (default = \code{""})}

\item{excludePhases}{Exclude Phases by their name (variable \code{phase_name}). Leave empty to include all Phases (after \code{excludeOther} and \code{excludeNorec}) (default = \code{""})}
}
\value{
plot
}
\description{
Plot a profile (values of the dimensions) for each phase.
}
\examples{

RN <- rn(y1 = data.frame(x = rnorm(100), y= rnorm(100)), weighted = TRUE)
phase_out <- rn_phases(RN, returnCentroid = "mean.sd")
plotRN_phaseProfile(phaseOutput = phase_out)


}
