% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotRN_phaseProfiles}
\alias{plotRN_phaseProfiles}
\title{Profile Plot}
\usage{
plotRN_phaseProfiles(
  phaseOutput,
  plotCentroid = NA,
  showEpochLegend = TRUE,
  epochColours = NULL,
  epochLabel = "Phase",
  excludeTransients = FALSE,
  excludePhaseNeighbours = FALSE,
  excludeNonrecurring = TRUE,
  excludeSingularities = TRUE,
  excludeVars = "",
  excludePhases = "",
  showPhaseSize = TRUE,
  returnGraph = FALSE,
  colOrder = NA
)
}
\arguments{
\item{phaseOutput}{Output from function \link{rn_phaseInfo}}

\item{excludeTransients}{Should the category "Transient" be excluded from plots? (default = \code{FALSE})}

\item{excludeNonrecurring}{Should the category "Nonrecurring" be excluded from plots? (default = \code{TRUE})}

\item{excludeSingularities}{Should the category "Singularity" be excluded from plots? (default = \code{TRUE})}

\item{excludeVars}{Exclude specific dimension variables by name. Leave empty to include all variables (default = \code{""})}

\item{excludePhases}{Exclude Phases by their name (variable \code{phase_name}). Leave empty to include all Phases (after the other exclusion arguments) (default = \code{""})}

\item{showPhaseSize}{Show the number of states in each phase in the labels (default = TRUE)}
}
\value{
plot
}
\description{
Plot a profile (values of the dimensions) for each phase.
}
\examples{

RN <- rn(y1 = data.frame(x = rnorm(100), y= rnorm(100)), weighted = TRUE)
phase_out <- rn_phaseInfo(RN, returnCentroid = "mean.sd")
plotRN_phaseProfiles(phaseOutput = phase_out)


}
