% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotRN_phaseProjections}
\alias{plotRN_phaseProjections}
\title{Plot Phase Space Projection}
\usage{
plotRN_phaseProjections(
  RNdist,
  phaseOutput,
  epochColours = NULL,
  showEpochLegend = TRUE,
  epochLabel = "Phase"
)
}
\arguments{
\item{RNdist}{A distance matrix (unthresholded) created with \link{rn} or \link{rp}}

\item{phaseOutput}{Output from function \link{rn_phaseInfo}}

\item{PhaseOutput}{Output from function \link{rn_phaseInfo} based on the thresholded matrix in \code{RNdist}}
}
\value{
Dataframe with coordinates
}
\description{
2D umap projection of multidimensional Phase Space. Categories "No recurrence" and "Other" are removed.
}
