% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotRN_phaseSeries}
\alias{plotRN_phaseSeries}
\title{Phase Series plot}
\usage{
plotRN_phaseSeries(
  phaseOutput,
  showEpochLegend = TRUE,
  epochColours = NULL,
  epochLabel = "Phase",
  excludeOther = FALSE,
  excludeNorec = TRUE,
  excludeVars = "",
  excludePhases = "",
  plotCentroid = FALSE,
  returnGraph = FALSE
)
}
\arguments{
\item{phaseOutput}{Output from function \link{rn_phases}}

\item{excludeOther}{Exclude the default Phase "Other"}

\item{excludeNorec}{Exclude the default Phase "No recurrence"}

\item{excludeVars}{Exclude specific dimension variables by name. Leave empty to include all variables (default = \code{""})}

\item{excludePhases}{Exclude Phases by their name (variable \code{phase_name}). Leave empty to include all Phases (after \code{excludeOther} and \code{excludeNorec}) (default = \code{""})}
}
\value{
phase density plot
}
\description{
Plot the sequence of phases as a time series.
}
\examples{

RN <- rn(y1 = rnorm(100), weighted = TRUE)
phase_out <- rn_phases(RN)
plotRN_phaseDensity(phase_out)


}
