% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotRN_phaseTimeSeries}
\alias{plotRN_phaseTimeSeries}
\title{Phase Series plot}
\usage{
plotRN_phaseTimeSeries(
  phaseOutput,
  showEpochLegend = TRUE,
  epochColours = NULL,
  epochLabel = "Phase",
  excludeVars = "",
  excludePhases = "",
  returnGraph = FALSE
)
}
\arguments{
\item{phaseOutput}{Output from function \link{rn_phaseInfo}}

\item{excludeVars}{Exclude specific dimension variables by name. Leave empty to include all variables (default = \code{""})}

\item{excludePhases}{Exclude Phases by their name (variable \code{phase_name}). Leave empty to include all Phases (after the other exclusion arguments) (default = \code{""})}
}
\value{
phase density plot
}
\description{
Plot the sequence of phases as a time series.
}
\examples{

RN <- rn(y1 = rnorm(100), weighted = TRUE)
phase_out <- rn_phaseInfo(RN)
plotRN_phaseTimeSeries(phase_out)


}
