% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R, R/nlRtsa_SOURCE.old.R
\name{plotRP.crqa}
\alias{plotRP.crqa}
\alias{plotRP.crqa}
\title{plotRP}
\usage{
plotRP.crqa(crqaOutput)

plotRP.crqa(crqaOutput)
}
\arguments{
\item{crqaOutput}{List output from \code{\link[crqa]{crqa}}}

\item{crqaOutput}{List output from \code{\link[crqa]{crqa}}}
}
\value{
A list

A list
}
\description{
Creates a recurrence plot from the sparse matrix output generated by \code{\link[crqa]{crqa}}.

Creates a recurrence plot from the sparse matrix output generated by \code{\link[crqa]{crqa}}.
}
\author{
Fred Hasselman

Fred Hasselman
}
