% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{plotRP_crqa}
\alias{plotRP_crqa}
\title{plotRP}
\usage{
plotRP_crqa(crqaOutput)
}
\arguments{
\item{crqaOutput}{List output from \code{\link[crqa]{crqa}}}
}
\value{
A list
}
\description{
Creates a recurrence plot from the sparse matrix output generated by \code{\link[crqa]{crqa}}.
}
\author{
Fred Hasselman
}
