% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotSUR_hist}
\alias{plotSUR_hist}
\title{Surrogate Test}
\usage{
plotSUR_hist(
  surrogateValues,
  observedValue,
  sides = c("two.sided", "greater", "less")[1],
  binWidth = NULL,
  measureName = "",
  title = "",
  doPlot = TRUE,
  returnOnlyPvalue = FALSE
)
}
\arguments{
\item{surrogateValues}{Vector of measures based on surrogate time series}

\item{observedValue}{The measure obtained from the observed value}

\item{sides}{Is this a 1 or 2-sided test (default = \code{1})}

\item{binWidth}{The size of the histogram bins. The default is to look for the max. number of digits and set the width to \code{1/10^(Ndigits-1)}. If integers are detectec width will be set to 1.}

\item{measureName}{Label for x-axis}

\item{title}{A title for the plot}

\item{doPlot}{Plot a histogram of the distribution (default = \code{TRUE})}

\item{returnOnlyPvalue}{Do not return the graph, just the point p-value (default = \code{FALSE})

alpha Significance threshold for the test. This value is currently calculated from the data as \eqn{\frac{1}{rank}*Nsides}, setting it will not have an effect.}
}
\value{
A point p-value for the observed value, and/or a histogram of the distribution (\code{ggplot2} object).
}
\description{
Surrogate Test
}
