% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{plotTS_multi}
\alias{plotTS_multi}
\title{Plot Multivariate Time Series Data}
\usage{
plotTS_multi(df, timeVec = NA, groupVec = NA, doPlot = TRUE,
  returnPlotData = FALSE)
}
\arguments{
\item{df}{A data frame with time series in columns.}

\item{timeVec}{If numeric, the number of the column in \code{df} which contains a time=keeping variable. If \code{NA}, the time vector will be \code{1:NROW(df)} (default = \code{NA})}

\item{groupVec}{A vector indicating the names of the time series in the columns of \code{df}. If \code{NA}, the column names of \code{df} will be used, excluding the \code{timeVec}, if present. (default = \code{NA})}

\item{doPlot}{Plot the \link{ggplot} object (default = \code{TRUE})}

\item{returnPlotData}{Return the restructered data frame used to create the plot (default = \code{FALSE})}
}
\value{
A \link{ggplot} object.
}
\description{
Plot Multivariate Time Series Data
}
\examples{

# Generate some coloured noise
N <- 512
noises <- seq(-3,3,by=.5)
y <- data.frame(matrix(rep(NA,length(noises)*N), ncol=length(noises)))

for(c in seq_along(noises)){
 y[,c] <- noise_powerlaw(N=N, alpha = noises[c])
 }
 colnames(y) <- paste0(noises)

 plotTS_multi(y)

}
