% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{point_p}
\alias{point_p}
\title{Surrogate Test}
\usage{
point_p(surrogateValues, observedValue, sides = 1,
  alpha = sides/(length(surrogateValues) + 1), doPlot = FALSE,
  measureName = "")
}
\arguments{
\item{surrogateValues}{Vector of measures based on surrogate time series}

\item{observedValue}{The measure obtained from the observed value}

\item{sides}{Is this a 1 or 2-sided test (default = \code{1})}

\item{alpha}{Significane threshold for the test}

\item{doPlot}{Plot a histogram (default = \code{FALSE})}

\item{measureName}{Label for x-axis}
}
\value{
A point p-value for the observed value
}
\description{
Surrogate Test
}
