% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R, R/nlRtsa_SOURCE.old.R
\name{recmat}
\alias{recmat}
\alias{recmat}
\title{Recurrence Matrix}
\usage{
recmat(y1, y2 = NULL, emDim = 1, emLag = 1, to.ts = NULL,
  to.sparse = FALSE, order.by = NULL, method = "Euclidean", ...)

recmat(y1, y2 = NULL, emDim = 1, emLag = 1, to.ts = NULL,
  to.sparse = FALSE, order.by = NULL, method = "Euclidean", ...)
}
\arguments{
\item{y1}{y1}

\item{y2}{y2}

\item{emDim}{emDim}

\item{emLag}{emLag}

\item{to.ts}{to.ts}

\item{to.sparse}{to.sparse}

\item{order.by}{order.by}

\item{method}{method}

\item{...}{other}

\item{y1}{y1}

\item{y2}{y2}

\item{emDim}{emDim}

\item{emLag}{emLag}

\item{to.ts}{to.ts}

\item{to.sparse}{to.sparse}

\item{order.by}{order.by}

\item{method}{method}

\item{...}{other}
}
\value{
RM

RM
}
\description{
Recurrence Matrix

Recurrence Matrix
}
