% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{recmat_plot}
\alias{recmat_plot}
\title{Plot (thresholded) distance matrix}
\usage{
recmat_plot(RM, plotDimensions = FALSE, plotMeasures = FALSE,
  radiusValue = NA, title = "", xlab = "", ylab = "",
  plotSurrogate = NA, doPlot = TRUE, useGtable = TRUE)
}
\arguments{
\item{RM}{A distance matrix or recurrence matrix}

\item{plotDimensions}{Should the state vectors be plotted if they are available as attributes of RM (default = \code{TRUE})}

\item{plotMeasures}{Print common (C)RQA measures in the plot if the matrix is binary}

\item{radiusValue}{If \code{plotMeasures = TRUE} and RM is an unthresholded matrix, this value will be used to calculate recurrence measures. If \code{plotMeasures = TRUE} and RM is already a binary recurence matrix, pass the radius that was used as a threshold to create the matrix for display purposes. If \code{plotMeasures = TRUE} and \code{radiusValue = NA}, function \code{crqa_radius()} will be called with default settings (find a radius that yields .05 recurrence rate). If \code{plotMeasures = FALSE} this setting will be ignored.}

\item{title}{A title for the plot}

\item{xlab}{An x-axis label}

\item{ylab}{An y-axis label}

\item{plotSurrogate}{Should a 2-panel comparison plot based on surrogate time series be added? If \code{RM} has attributes \code{y1} and \code{y2} containing the time series data (i.e. it was created by a call to \code{\link{recmat}}), the following options are available: "RS" (random shuffle), "RP" (randomised phases), "AAFT" (amplitude adjusted fourier transform). If no timeseries data is included, the columns will be shuffled.  NOTE: This is not a surrogate test, just 1 surrogate is created from \code{y1}.}

\item{doPlot}{Draw the plot? (default = \code{TRUE})}

\item{useGtable}{Use package \code{\link{gtable}}. If this results in errors (e.g. viewport settings), set set to FALSE to use package \code{patchwork}. This package is in development, see the warning for instructions on how to install it.}
}
\value{
A nice plot of the recurrence matrix.
}
\description{
Plot (thresholded) distance matrix
}
\seealso{
Other Distance matrix operations: \code{\link{bandReplace}},
  \code{\link{di2bi}}, \code{\link{di2we}},
  \code{\link{dist_hamming}}, \code{\link{lineDists}},
  \code{\link{nzdiags}}, \code{\link{recmat_size}},
  \code{\link{recmat}}
}
