% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{recmat_size}
\alias{recmat_size}
\title{recmat_size}
\usage{
recmat_size(mat, AUTO = NULL, theiler = 0)
}
\arguments{
\item{mat}{A Matrix object}

\item{AUTO}{Is the Matrix an Auto Recurrence Matrix? If so, the length of the diagonal will be subtracted from the matrix size, pass \code{FALSE} to prevent this behaviour. If \code{NULL} (default) \code{AUTO} will take on the value of \code{isSymmetric(mat)}.}

\item{theiler}{Should a Theiler window be applied?}
}
\value{
Matrix size for computation of recurrence measures.
}
\description{
recmat_size
}
\examples{
# Create a 10 by 10 matrix
library(Matrix)
m <- Matrix(rnorm(10),10,10)

recmat_size(m,TRUE,0)   # Subtract diagonal
recmat_size(m,FALSE,0)  # Do not subtract diagonal
recmat_size(m,NULL,0)   # Matrix is symmetrical, AUTO is set to TRUE
recmat_size(m,NULL,1)   # Subtract a Theiler window of 1 around and including the diagonal
}
