% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{repmat}
\alias{repmat}
\title{Repeat Copies of a Matrix}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{A matrix}

\item{m}{Multiply dim(X)[1] m times}

\item{n}{Multiply dim(X)[2] n times}
}
\value{
A repeated matrix
}
\description{
Repeat Copies of a Matrix
}
