% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{return_error}
\alias{return_error}
\title{Return Error}
\usage{
return_error(expr)
}
\arguments{
\item{expr}{The expression}
}
\value{
A list object with 2 fields: 'return' and 'message'.
}
\description{
Return Error
}
\keyword{internal}
