% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{rn}
\alias{rn}
\title{Create a Recurrence Network Matrix}
\usage{
rn(y1, y2 = NULL, emDim = 1, emLag = 1, emRad = NULL,
  directed = FALSE, weighted = FALSE, includeDiagonal = FALSE,
  to.ts = NULL, order.by = NULL, to.sparse = FALSE,
  method = "Euclidean", returnGraph = TRUE, doPlot = FALSE,
  silent = TRUE, ...)
}
\arguments{
\item{y1}{A numeric vector or time series}

\item{y2}{A numeric vector or time series for cross recurrence}

\item{emDim}{The embedding dimensions}

\item{emLag}{The embedding lag}

\item{emRad}{The threshold (emRad) to apply to the distance matrix to create a binary matrix}

\item{directed}{Should the matrix be considered to represent a directed network? (default = \code{FALSE})}

\item{weighted}{Should the matrix be considered to represent a weighted network? (default = \code{FALSE})}

\item{includeDiagonal}{Should the diagonal of the matrix be included when creating the network (default = \code{FALSE})}

\item{to.ts}{Should \code{y1} and \code{y2} be converted to time series objects?}

\item{order.by}{If \code{to.ts = TRUE}, pass a vector of the same length as \code{y1} and \code{y2}. It will be used as the time index, if \code{NA} the vector indices will be used to represent time.}

\item{to.sparse}{Should sparse matrices be used?}

\item{method}{Distance measure to use. Any option that is valid for argument \code{method} of \code{\link[proxy:dist]{proxy::dist()}}. Type \code{proxy::pr_DB$get_entries()} to se a list of all the options. Common methods are: "Euclidean", "Manhattan", "Minkowski", "Chebysev" (or the same but shorter: "L2","L1","Lp" and "max" distance) (default = \code{"Euclidean"})}

\item{returnGraph}{Return an \code{\link[igraph:igraph]{igraph::igraph()}} object (default = \code{TRUE})}

\item{doPlot}{Plot the matrix and its network representation by calling \code{\link[=rp_plot]{rp_plot()}} and \code{\link[=rn_plot]{rn_plot()}} with default settings}

\item{silent}{Silent-ish mode (default = \code{TRUE})}

\item{...}{Any paramters to pass to \code{\link[=rn_plot]{rn_plot()}} if \code{doPlot = TRUE}}
}
\value{
A (Coss-) Recurrence matrix that can be interpreted as an adjacency (or incidence) matrix.
}
\description{
This function serves as a wrapper for function \code{rp()}, it will add some attributes to the matrix related to network representation. These attributes will be used to decide which network type to generate (e.g. undirected, directed, weighted, etc.)
}
\seealso{
Other Distance matrix operations (recurrence network): \code{\link{di2bi}},
  \code{\link{di2we}}, \code{\link{rn_plot}},
  \code{\link{rn_recSpec}}, \code{\link{rn_scaleoGram}}
}
\concept{Distance matrix operations (recurrence network)}
