% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rn.R
\name{rn_findPhases}
\alias{rn_findPhases}
\title{Find Phases}
\usage{
rn_findPhases(
  RN,
  weighted = NA,
  inverseWeight = TRUE,
  directed = NA,
  selectionMethod = c("degree", "strength", "closeness", "betweenness")[1],
  minStatesinPhase = 2,
  maxStatesinPhase = NROW(RN),
  maxPhases = NROW(RN),
  silent = FALSE
)
}
\arguments{
\item{RN}{A matrix produced by the function \link{rn}}

\item{weighted}{Should the matrix be considered to represent a weighted network? (default = \code{FALSE})}

\item{inverseWeight}{Whether to perform the operation \code{1/weight} on the edge weights. This will only have an effect if wieght matters for the selection of most important nodes (e.g., \code{selectionMethod = "strength"}). The default is \code{TRUE}, so if the matrix was weighted by a distance metric (\code{weightedBy = "si"}) edges with smaller distances (recurring coordinates closer to the current coordinate) have greater impact on node strength. If the matrix was weighted by recurrence time (\code{weightedBy = "rt"}) and \code{inverseWeight = TRUE}, recurrent points with shorter recurrence times will have greater impact on the strength calculation. If \code{weightedBy = "rf"}, lower frequencies will end up having more impact. (default = \code{TRUE})}

\item{directed}{Should the matrix be considered to represent a directed network? (default = \code{FALSE})}

\item{selectionMethod}{How will the most "important" node be selected? Should be the name of an \link{igraph} function that returns local (vertex) measures (default = \code{"degree"})}

\item{minStatesinPhase}{A parameter applied after the extraction of phases (limited by \code{maxPhases}). If any extracted phases do not have a minimum number of \code{minStatesinPhase} + \code{1} (= the state that was selected based on node strength), the phase will be removed from the result (default = \code{1})}

\item{maxStatesinPhase}{A parameter applied after the extraction of phases (limited by \code{maxPhases}). If any extracted phases exceeds a maximum number of \code{maxStatesinPhase} + \code{1} (= the state that was selected based on node strength), the phase will be removed from the result (default = \code{NROW(RN)})}

\item{maxPhases}{The maximum number of phases to extract. These will be the phases associated with the highest node degree or node strength. All other recurrent points will be labelled with "Other". If \code{NA}, the value will be set to \code{NROW(RN)}, this will return all the potential phases in the data irrespective of their frequency/strength of recurrence  (default = \code{10})}

\item{silent}{Silent-ish mode}
}
\value{
A data frame with nodes and the phase sequence
}
\description{
Algorithm to find similar recurring states that can be considered phases, based on the node importance (\code{selectionMethod}) of a recurrence network. This makes most sense if the state space dimension can be interpreted.
}
