% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{rn_measures}
\alias{rn_measures}
\title{Recurrence Network Measures}
\usage{
rn_measures(g, cumulative = TRUE, silent = TRUE)
}
\arguments{
\item{g}{An igraph object.}

\item{cumulative}{Only consider out-degree.}

\item{silent}{Siletn(ish) mode}
}
\value{
A list with data frames with common vertex, edge amnd global network measures.
}
\description{
Recurrence Network Measures
}
\examples{

library(igraph)
g <- igraph::sample_gnp(1000, 10/1000)

outList <- rn_measures(g, silent = FALSE)

}
