% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rn.R
\name{rn_phases}
\alias{rn_phases}
\title{Extract Phases from weighted RN}
\usage{
rn_phases(
  RN,
  maxPhases = 5,
  minStatesinPhase = 1,
  maxStatesinPhase = NROW(RN),
  removeSingularities = FALSE,
  standardise = c("none", "mean.sd", "median.mad", "unit")[4],
  returnGraph = FALSE,
  doPhasePlot = FALSE,
  phaseColours = NULL,
  doSpiralPlot = FALSE,
  showEpochLegend = TRUE,
  epochColours = NULL,
  epochLabel = "Phase",
  excludeOther = FALSE,
  excludeNorec = TRUE
)
}
\arguments{
\item{RN}{A matrix produced by the function \link{rn}}

\item{maxPhases}{The maximum number of phases to extract. Use \code{NA} to extract all potential phases (default = \code{5})}

\item{minStatesinPhase}{A parameter applied after the extraction of phases (limited by \code{maxPhases}). If any extracted phases do not have a minimum number of \code{minStatesinPhase} + \code{1} (= the state that was selected based on node strength), the phase will be removed from the result (default = \code{1})}

\item{maxStatesinPhase}{A parameter applied after the extraction of phases (limited by \code{maxPhases}). If any extracted phases exceeds a maximum number of \code{maxStatesinPhase} + \code{1} (= the state that was selected based on node strength), the phase will be removed from the result (default = \code{floor(NROW(RN)/maxPhases)})}

\item{removeSingularities}{Will remove states that recur only once (nodes with \code{degree(g) == 1}) (default = \code{FALSE})}

\item{returnGraph}{Return the graph object objects of the plots that have been produced (default = \code{FALSE})}

\item{doPhasePlot}{Produce a plot with the extracted phases (default = \code{TRUE})}

\item{phaseColours}{Colours for the different phases in the phase plot. If \code{epochColours} also has a value, \code{phaseColours} will be used instead (default = \code{NULL})}

\item{doSpiralPlot}{Produce a plot of the recurrence network with the nodes coloured by phases (default = \code{FALSE})}

\item{showEpochLegend}{Should a legend be shown for the epoch colours? (default = \code{TRUE})}

\item{epochColours}{A vector of length \code{vcount(g)} with colour codes (default = \code{NULL})}

\item{epochLabel}{A title for the epoch legend (default = \code{"Epoch"})}

\item{excludeOther}{Should the phase "Other" be excluded from plots? (default = \code{FALSE})}

\item{excludeNorec}{Should the category "No recurrence" be excluded from plots? (default = \code{TRUE})}

\item{rescaleDimensions}{If the dimensions are not on the same scale, transform by using \code{"standardise"} or \code{"unit"}, otherwise use \code{"none"} (default = \code{"unit"})}
}
\value{
A data frame with information about the phases.
}
\description{
This function will extract phases (regions of attraction in phase space) based on a weighted \code{RN} object created with function \link{rn}.
The assumption is that coordinates in state space that are either close in terms of distance, or are re-visited with short recurrence times,
or with high-frequency, are regions of attraction for the system.
}
\details{
The method used for the identification of phases is on the properties of the \code{RN} object:
\itemize{
\item If weighted by distance \code{"si"}, the inverse distance will be used, which means higher weights correspond to closer states.
\item If weighted by recurrence time \code{"rt"}, the inverse time will be used, which means higher weights correspond to faster recurrence times.
}

The procedure is as follows:
\enumerate{
\item Identify the node with the highest strength
\item Identify the nodes that connect to this node
\item Identify the node with highest strength that does not connect to the node identified in step 1.
\item Repeat.
}
}
\examples{

# Use the ManyAnalysts dataset to create a phase plot with default settings
data("manyAnalystsESM")
df <- manyAnalystsESM[4:10]
RN <- rn(y1 = df, doEmbed = FALSE, weighted = TRUE, weightedBy = "si", emRad = NA)

# This returns 6 phases which have minimally 1 state
rn_phases(RN, maxPhases = 10, doPhasePlot = TRUE)

# Use min. number of states as the extraction criterion
rn_phases(RN, maxPhases = NA, minStatesinPhase = 7, doPhasePlot = TRUE)

}
\seealso{
Other Distance matrix operations (recurrence network): 
\code{\link{di2bi}()},
\code{\link{di2ch}()},
\code{\link{di2we}()},
\code{\link{rn_plot}()},
\code{\link{rn_recSpec}()},
\code{\link{rn}()}
}
\concept{Distance matrix operations (recurrence network)}
