% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{rn_plot}
\alias{rn_plot}
\title{Plot (thresholded) distance matrix as a network}
\usage{
rn_plot(RN, title = "", xlab = "", ylab = "",
  returnOnlyObject = FALSE, useGtable = TRUE)
}
\arguments{
\item{RN}{A distance matrix or recurrence matrix}

\item{title}{A title for the plot}

\item{xlab}{An x-axis label}

\item{ylab}{An y-axis label}

\item{returnOnlyObject}{Return the \code{ggplot} / \code{ggraph} object only, do not draw the plot (default = \code{FALSE})}

\item{useGtable}{Use package \code{\link[=gtable]{gtable()}}. If this results in errors (e.g. viewport settings), set to FALSE (default = \code{TRUE})}
}
\value{
A nice plot of the recurrence network
}
\description{
Plot (thresholded) distance matrix as a network
}
\seealso{
Other Distance matrix operations (recurrence network): \code{\link{di2bi}},
  \code{\link{di2we}}, \code{\link{rn_recSpec}},
  \code{\link{rn_scaleoGram}}, \code{\link{rn}}
}
\concept{Distance matrix operations (recurrence network)}
