% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{rn_recSpec}
\alias{rn_recSpec}
\title{Recurrence Time Spectrum}
\usage{
rn_recSpec(RN, fitRange = NULL, doPlot = TRUE,
  returnOnlyObject = FALSE, returnPowerLaw = FALSE)
}
\arguments{
\item{RN}{A thresholded recurrence matrix generated by function \code{rn()}}

\item{fitRange}{If \code{NULL} the entire range will be used for log-log slope. If a 2-element vector of integers, this will represent the range of recurrence times to use for fitting the log=log slope (e.g. \code{c(1,50)} would fit the first 50 recurrence times).}

\item{doPlot}{Should a plot of the recurrence time spectrum be produced?}

\item{returnOnlyObject}{Return the \code{ggplot} / \code{ggraph} object only, do not draw the plot (default = \code{FALSE})}

\item{returnPowerLaw}{Return results of log-log regression}
}
\value{
A vector of frequencies of recurrence times and a plot (if requested)
}
\description{
Get the recurrence time distribution from a recurrence network.
}
\seealso{
Other Distance matrix operations (recurrence plot): \code{\link{bandReplace}},
  \code{\link{di2bi}}, \code{\link{di2we}},
  \code{\link{dist_hamming}}, \code{\link{rn_scaleoGram}},
  \code{\link{rp_lineDist}}, \code{\link{rp_nzdiags}},
  \code{\link{rp_plot}}, \code{\link{rp_size}},
  \code{\link{rp}}
}
\concept{Distance matrix operations (recurrence plot)}
