% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{rn_scaleoGram}
\alias{rn_scaleoGram}
\title{Recurrence Time Scaleogram}
\usage{
rn_scaleoGram(RN, returnOnlyObject = FALSE)
}
\arguments{
\item{RN}{A thresholded recurrence matrix generated by function \code{rn()}}

\item{returnOnlyObject}{Return the \code{ggplot} / \code{ggraph} object only, do not draw the plot (default = \code{FALSE})}
}
\value{
A \code{ggraph} graph object
}
\description{
Display a recurrence network in a space representing Time (x-axis) x Scale (y-axis). The scale axis will be determined by the latency between the occurence of a value in the (embedded) time series vector and its recurrences in the future (i.e. only the upper triangle of the recurrence matrix will be displayed, excluding the diagonal).
}
\seealso{
Other Distance matrix operations (recurrence plot): \code{\link{bandReplace}},
  \code{\link{di2bi}}, \code{\link{di2we}},
  \code{\link{dist_hamming}}, \code{\link{rn_recSpec}},
  \code{\link{rp_lineDist}}, \code{\link{rp_nzdiags}},
  \code{\link{rp_plot}}, \code{\link{rp_size}},
  \code{\link{rp}}
}
\concept{Distance matrix operations (recurrence plot)}
