% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rn.R
\name{rn_strengthDist}
\alias{rn_strengthDist}
\title{Strength versus Degree scaling relation}
\usage{
rn_strengthDist(g, mode = c("in", "out", "all")[3], doPlot = TRUE)
}
\arguments{
\item{g}{an igraph object representing a weighted Recurrence  Network}

\item{mode}{Evaluate the "in", "out" degree, or "all" edges (default = \code{"all"})}

\item{doPlot}{Plot the scaling relation? (default = TRUE)}
}
\value{
A data frame with local vertex strength and vertex degree, including
}
\description{
Calculates the Recurrence Rate versus Recurrence Time power-law
}
\examples{

y  <- rnorm(100)
RN <- rn(y, emLag=1, emDim=3, emRad=NA, weighted = TRUE, weightedBy = "rt", returnGraph = TRUE)
rn_strengthDist(RN$g)

}
