% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rn.R
\name{rn_transition}
\alias{rn_transition}
\title{Create transition network}
\usage{
rn_transition(
  phaseSequence,
  threshold = NA,
  doMatrixPlot = TRUE,
  doNetworkPlot = TRUE,
  excludeOther = FALSE,
  excludeNorec = TRUE,
  returnGraph = FALSE
)
}
\arguments{
\item{phaseSequence}{A vector with names or numbers that represent a sequence of phases or order parameter dynamics. If a named numeric vector is passed, the names attribute will be used to represent the phases. If the variable \code{phase_name}, generated by \link{rn_phases} is used, the arguments \code{excludeOther} and \code{excludeNorec} wil be evaluated.}

\item{threshold}{Provide a threshold for the relative frequencies. Values below the threshold will be set to \code{0}. Pass \code{NA} to not use a threshold (default = \code{NA})}

\item{doMatrixPlot}{default(\code{TRUE})}

\item{doNetworkPlot}{default(\code{TRUE})}

\item{excludeOther}{Should the phase "Other" be excluded from plots? (default = \code{FALSE})}

\item{excludeNorec}{Should the category "No recurrence" be excluded from plots? (default = \code{TRUE})}

\item{returnGraph}{Return an \code{\link[igraph:aaa-igraph-package]{igraph::igraph()}} object (default = \code{FALSE})}
}
\value{
A transition matrix based on relative frequencies
}
\description{
Create transition network
}
\examples{

# This will output the transition matrix, a plot of the matrix and a transition network plot.
y <- c("Happy", "Happy", "Sad", "Neutral", "Neutral", "Angry", "Sad", "Sad", "Neutral")
TM <- rn_transition(y)

}
\author{
Fred Hasselman

Matti Heino
}
