% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNG}
\alias{RNG}
\title{Random Number Sequences}
\format{A data frame with 24200 rows and 3 variables:
\describe{
  \item{ID}{Participant ID}
  \item{time}{Temporal order}
  \item{number}{A number between 1 and 9}
}}
\source{
\url{https://www.frontiersin.org/articles/10.3389/fnhum.2015.00319/full}
}
\usage{
RNG
}
\description{
A dataset containing sequences of 100 numbers generated bij 242 participants who were instruted to generate random sequences.
}
\references{
Oomens, W., Maes, J. H., Hasselman, F., & Egger, J. I. (2015). A time series approach to random number generation: using recurrence quantification analysis to capture executive behavior. Frontiers in human neuroscience, 9
}
\keyword{datasets}
