% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{roc_noise}
\alias{roc_noise}
\title{roc_noise}
\usage{
roc_noise(y, radius, emDim = 1, emLag = 1, noiseLevel = 0.75,
  standardise = c("mean.sd", "median.mad", "none")[3],
  noiseType = c("normal", "uniform")[1])
}
\arguments{
\item{y}{y}

\item{radius}{radius}

\item{emDim}{embedding Dims}

\item{emLag}{embedding Lag}

\item{noiseLevel}{noise Level}

\item{standardise}{Standardise y? Choose from "mean.sd","median.mad","none".}

\item{noiseType}{Use a Normal distribution of uniform distribution for noiselevels}
}
\value{
data frame for ROC
}
\description{
roc_noise
}
\keyword{internal}
