% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{rocNoise}
\alias{rocNoise}
\title{rocNOISE}
\usage{
rocNoise(y, radius, eDim = 1, eLag = 1, noiseLevel = 0.75)
}
\arguments{
\item{y}{y}

\item{radius}{radius}

\item{eDim}{embedding Dims}

\item{eLag}{embedding Lag}

\item{noiseLevel}{noise Level}
}
\value{
data frame for ROC
}
\description{
rocNOISE
}
\keyword{internal}
