% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{round_trunc}
\alias{round_trunc}
\title{Rounding by truncation}
\usage{
round_trunc(number, digits)
}
\arguments{
\item{number}{number}

\item{digits}{digits}
}
\value{
True rounding
}
\description{
Rounding by truncation
}
\keyword{internal}
