% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{rp_checkfix}
\alias{rp_checkfix}
\title{Check a Recurrence Matrix}
\usage{
rp_checkfix(
  RM,
  checkS4 = TRUE,
  checkAUTO = TRUE,
  checkSPARSE = FALSE,
  checkTSPARSE = FALSE,
  fixS4 = FALSE,
  fixAUTO = TRUE,
  fixSPARSE = FALSE,
  fixTSPARSE = FALSE
)
}
\arguments{
\item{RM}{RM}

\item{checkS4}{checkS4}

\item{checkAUTO}{checkAUTO}

\item{checkSPARSE}{checkSPARSE}

\item{fixS4}{fixS4}

\item{fixAUTO}{fixAUTO}

\item{fixSPARSE}{fixSPARSE}
}
\value{
A checked and/or fixed matrix
}
\description{
Check a Recurrence Matrix
}
\keyword{internal}
