% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{rp_cl_main}
\alias{rp_cl_main}
\title{rp_cl_main}
\usage{
rp_cl_main(
  data,
  emDim = 1,
  emLag = 1,
  emRad = NA,
  DLmin = 2,
  VLmin = 2,
  theiler = 0,
  win = min(length(y1), ifelse(is.null(y2), (length(y1) + 1), length(y2)), na.rm =
    TRUE),
  step = step,
  JRP = FALSE,
  distNorm = c("EUCLIDEAN", "MAX", "MIN", "OP")[[1]],
  returnMeasures = TRUE,
  returnRPvector = FALSE,
  returnLineDist = FALSE,
  doPlot = c("noplot", "rp", "distmat")[[1]],
  path_to_rp = getOption("casnet.path_to_rp"),
  saveOut = FALSE,
  path_out = NULL,
  file_ID = NULL,
  silent = TRUE,
  targetValue = 0.05,
  useParallel = FALSE,
  ...
)
}
\arguments{
\item{emDim}{Embedding dimensions (default = \code{1})}

\item{emLag}{Embedding lag (default = \code{1})}

\item{emRad}{Radius on distance matrix (default = \code{1})}

\item{DLmin}{Minimum length of diagonal structure to be considered a line (default = \code{2})}

\item{VLmin}{Minimum length of vertical structure to be considered a line (default = \code{2})}

\item{theiler}{Theiler window (default = \code{0})}

\item{win}{Window to calculate the (C)RQA (default = minimum of length of \code{y1} or \code{y2})}

\item{step}{Stepsize for sliding windows (default = size of \code{win}, so no sliding window)}

\item{JRP}{Wether to calculate a Joint Recurrence Plot (default = \code{FALSE})}

\item{distNorm}{One of "EUCLIDEAN" (default), \verb{"MAX", "MIN"}, or \code{"OP"} for an Order Pattern recurrence matrix}

\item{returnMeasures}{Return the (C)RQA measures? (default = \code{TRUE})}

\item{returnRPvector}{Return the recurrent points in a dataframe? (default = \code{FALSE})}

\item{returnLineDist}{Return the distribution of diagonal and horizontal line length distances (default = \code{FALSE})}

\item{doPlot}{Produce a plot of the recurrence matrix by calling \code{\link[=rp_plot]{rp_plot()}}, values can be \code{"rp"} (the thresholded recurrence matrix),\code{"distmat"} (the unthresholded recurrence matrix) or \code{"noplot"} (default = \code{"noplot"})}

\item{path_to_rp}{Path to the command line executable (default = path set during installation, use \code{getOption("casnet.path_to_rp")} to see)}

\item{saveOut}{Save the output to files? If \code{TRUE} and \code{path_out = NA}, the current working directory will be used (default = \code{FALSE})}

\item{path_out}{Path to save output if \code{saveOut = TRUE} (default = \code{NULL})}

\item{file_ID}{A file ID which will be a prefix to to the filename if \code{saveOut = TRUE} (default = \code{NULL}, an integer will be added tot the file name to ensure unique files)}

\item{silent}{Do not display any messages (default = \code{TRUE})}

\item{targetValue}{A value passed to \code{est_radius(...,type="fixed", targetMeasure="RR")} if \code{is.na(emRad)==TRUE}. This is useful for windowed analysis, it will estimate a new radius for each window.}

\item{useParallel}{Speed up calculations by using the parallel processing options provided by \code{parallel} to assign a seperate process/core for each window in windowed (C)RQA analysis and \code{\link[parallel:detectCores]{parallel::detectCores()}} with\code{logical = TRUE} to decide on the available cores (default = \code{FALSE})}

\item{...}{Additional parameters (currently not used)}
}
\description{
rp_cl_main
}
\keyword{internal}
