% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{rp_copy_attributes}
\alias{rp_copy_attributes}
\title{Copy Matrix Attributes}
\usage{
rp_copy_attributes(
  source,
  target,
  source_remove = c("names", "row.names", "class", "dim", "dimnames", "x")
)
}
\arguments{
\item{source}{Source matrix}

\item{target}{Target matrix}

\item{source_remove}{Remove these attribute fields from the source before copying.}
}
\value{
The target matrix with attributes copied from the source matrix.
}
\description{
Simple attribute copy used in \code{casnet} to convert between \code{matrix} and \code{Matrix} classes and back.
}
