% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{rp_diagProfile}
\alias{rp_diagProfile}
\alias{crqa_diagprofile}
\title{Diagonal Recurrence Profile}
\usage{
rp_diagProfile(
  RM,
  diagWin = NULL,
  xname = "X-axis",
  yname = "Y-axis",
  theiler = 0,
  DLmin = 2,
  VLmin = 2,
  HLmin = 2,
  DLmax = length(Matrix::diag(RM)) - 1,
  VLmax = length(Matrix::diag(RM)) - 1,
  HLmax = length(Matrix::diag(RM)) - 1,
  doShuffle = FALSE,
  y1 = NA,
  y2 = NA,
  Nshuffle = 19,
  AUTO = NULL,
  chromatic = FALSE,
  matrices = FALSE,
  doPlot = TRUE
)
}
\arguments{
\item{RM}{A binary recurrence matrix}

\item{diagWin}{Window around the line of synchrony}

\item{xname}{Label for x-axis}

\item{yname}{Label for y-axis}

\item{theiler}{Use a theiler window around the line of identity / synchronisation to remove high auto-correlation at short time-lags (default = \code{0})}

\item{DLmin}{Minimal diagonal line length (default = \code{2})}

\item{VLmin}{Minimal vertical line length (default = \code{2})}

\item{HLmin}{Minimal horizontal line length (default = \code{2})}

\item{DLmax}{Maximal diagonal line length (default = length of diagonal -1)}

\item{VLmax}{Maximal vertical line length (default = length of diagonal -1)}

\item{HLmax}{Maximal horizontal line length (default = length of diagonal -1)}

\item{doShuffle}{Should a shuffled baseline be calculated (default = \code{FALSE})}

\item{y1}{The original \code{y1} time series}

\item{y2}{The original \code{y2} time series}

\item{Nshuffle}{How many shuffled versions to make up the baseline? The default is \code{19}, which is the minimum for a one-sided surrogate test.}

\item{AUTO}{Auto-recurrence? (default = \code{FALSE})}

\item{chromatic}{Force chromatic RQA? (default = \code{FALSE})}

\item{matrices}{Return matrices? (default = \code{FALSE})}

\item{doPlot}{Plot}
}
\value{
A plot and/or the data for the plot
}
\description{
Diagonal Recurrence Profile
}
