% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{rp_lineDist}
\alias{rp_lineDist}
\title{Line length distributions}
\usage{
rp_lineDist(RM, DLmin = 2, VLmin = 2, HLmin = 2,
  DLmax = length(Matrix::diag(RP)) - 1,
  VLmax = length(Matrix::diag(RP)) - 1,
  HLmax = length(Matrix::diag(RP)) - 1, d = NULL, theiler = NULL,
  invert = FALSE, AUTO = NULL, chromatic = FALSE, matrices = FALSE,
  doHalf = FALSE)
}
\arguments{
\item{RM}{A thresholded recurrence matrix (binary: 0 - 1)}

\item{DLmin}{Minimal diagonal line length (default = \code{2})}

\item{VLmin}{Minimal vertical line length (default = \code{2})}

\item{HLmin}{Minimal horizontal line length (default = \code{2})}

\item{DLmax}{Maximal diagonal line length (default = length of diagonal -1)}

\item{VLmax}{Maximal vertical line length (default = length of diagonal -1)}

\item{HLmax}{Maximal horizontal line length (default = length of diagonal -1)}

\item{d}{Vector of diagonals to be extracted from matrix \code{RP} before line length distributions are calculated. A one element vector will be interpreted as a windowsize, e.g., \code{d = 50} will extract the diagonal band \code{-50:50}. A two element vector will be interpreted as a band, e.g. \code{d = c(-50,100)} will extract diagonals \code{-50:100}. If \code{length(d) > 2}, the numbers will be interpreted to refer to individual diagonals, \code{d = c(-50,50,100)} will extract diagonals \code{-50,50,100}.}

\item{theiler}{Size of the theiler window, e.g. \code{theiler = 1} removes diagonal bands -1,0,1 from the matrix. If \code{length(d)} is \code{NULL}, 1 or 2, the theiler window is applied before diagonals are extracted. The theiler window is ignored if \code{length(d)>2}, or if it is larger than the matrix or band indicated by parameter \code{d}.}

\item{invert}{Relevant for Recurrence Time analysis: Return the distribution of 0 valued segments in nonzero diagonals/verticals/horizontals. This indicates the time between subsequent line structures.}

\item{AUTO}{Is this an AUTO RQA?}

\item{chromatic}{Chromatic RQA?}

\item{matrices}{Return the matrices ?}

\item{doHalf}{Analyse half of the matrix?}
}
\value{
A list object with distributions of line lengths. If \code{matrices = TRUE} datafr are returned whose columns represent the nonzero diagonals, verticals, or, horizontals.
}
\description{
Extract lengths of diagonal, vertical and horizontal line segments from a recurrence matrix.
}
\details{
Based on the Matlab function \code{linedists} by Stefan Schinkel, Copyright (C) 2009 Stefan Schinkel, University of Potsdam, http://www.agnld.uni-potsdam.de

References:
S. Schinkel, N. Marwan, O. Dimigen & J. Kurths (2009):
"Confidence Bounds of recurrence-based complexity measures
Physics Letters A,  373(26), pp. 2245-2250

Copyright (C) 2009 Stefan Schinkel, University of Potsdam
\url{http://www.agnld.uni-potsdam.de}
}
\seealso{
Other Distance matrix operations: \code{\link{bandReplace}},
  \code{\link{di2bi}}, \code{\link{di2we}},
  \code{\link{dist_hamming}}, \code{\link{rp_nzdiags}},
  \code{\link{rp_plot}}, \code{\link{rp_size}},
  \code{\link{rp}}
}
\author{
Fred Hasselman
}
\concept{Distance matrix operations}
