% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{rp_measures}
\alias{rp_measures}
\alias{crqa_rp}
\title{Get (C)RQA measures based on a binary matrix}
\usage{
rp_measures(
  RM,
  emRad = NA,
  DLmin = 2,
  VLmin = 2,
  HLmin = 2,
  DLmax = length(Matrix::diag(RM)) - 1,
  VLmax = length(Matrix::diag(RM)) - 1,
  HLmax = length(Matrix::diag(RM)) - 1,
  AUTO = NULL,
  theiler = NULL,
  chromatic = FALSE,
  matrices = FALSE,
  doHalf = FALSE,
  Nboot = NULL,
  CL = 0.95,
  targetValue = 0.05,
  doParallel = FALSE,
  silent = TRUE
)
}
\arguments{
\item{RM}{A distance matrix, or a matrix of zeroes and ones (you must set \code{emRad = NA})}

\item{emRad}{Threshold for distance value that counts as a recurrence}

\item{DLmin}{Minimal diagonal line length (default = \code{2})}

\item{VLmin}{Minimal vertical line length (default = \code{2})}

\item{HLmin}{Minimal horizontal line length (default = \code{2})}

\item{DLmax}{Maximal diagonal line length (default = length of diagonal -1)}

\item{VLmax}{Maximal vertical line length (default = length of diagonal -1)}

\item{HLmax}{Maximal horizontal line length (default = length of diagonal -1)}

\item{AUTO}{Auto-recurrence? (default = \code{FALSE})}

\item{theiler}{Use a theiler window around the line of identity / synchronisation to remove high auto-correlation at short time-lags (default = \code{0})}

\item{chromatic}{Force chromatic RQA? (default = \code{FALSE})}

\item{matrices}{Return matrices? (default = \code{FALSE})}

\item{doHalf}{Analyse half of the matrix? (default = \code{FALSE})}

\item{Nboot}{How many bootstrap replications? (default = \code{NULL})}

\item{CL}{Confidence limit for bootstrap results (default = \code{.95})}

\item{targetValue}{A value passed to \code{est_radius(...,type="fixed", targetMeasure="RR", tol = .2)} if \code{is.na(emRad)==TRUE}, it will estimate a radius (default = \code{.05}).}

\item{doParallel}{Speed up calculations by using the parallel processing options provided by \code{parallel} to assign a seperate process/core for each window in windowed (C)RQA analysis and \code{\link[parallel:detectCores]{parallel::detectCores()}} with  \code{logical = TRUE} to decide on the available cores (default = \code{FALSE})}

\item{silent}{Do not display any messages (default = \code{TRUE})}
}
\value{
A list object containing (C)RQA measures (and matrices if requested)
}
\description{
A zoo of measures based on singular recurrent points, diagonal, vertical and horizontal line structures (anisotropic) will be caluclated.
}
\seealso{
Other Recurrence Quantification Analysis: 
\code{\link{rp_cl}()},
\code{\link{rp_measures_main}()}
}
\concept{Recurrence Quantification Analysis}
