% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{rp_nzdiags}
\alias{rp_nzdiags}
\title{rp_nzdiags}
\usage{
rp_nzdiags(
  RM = NULL,
  d = NULL,
  returnVectorList = TRUE,
  returnNZtriplets = FALSE,
  removeNZ = TRUE,
  silent = TRUE
)
}
\arguments{
\item{RM}{A binary (0,1) matrix.}

\item{d}{An optional vector of diagonals to extract.}

\item{returnVectorList}{Return list}

\item{returnNZtriplets}{Return a dataframe with coordinates of only nonzero elements in diagonals (default = \code{FALSE})}

\item{removeNZ}{Remove nonzero diagonals if \code{TRUE}. If \code{FALSE} returns the full diagonals matrix. Use e.g. to plot diagonal recurrence profiles (default = \code{TRUE})}

\item{silent}{Silent-ish mode}
}
\value{
A matrix object with nonzero diagonals as columns and/or a dataframe with coordinates of nonzero diagonal elements
}
\description{
Get all nonzero diagonals of a binary matrix, or, diagonals specified as a vector by argument \code{d}.
}
\seealso{
Other Distance matrix operations (recurrence plot): 
\code{\link{bandReplace}()},
\code{\link{di2bi}()},
\code{\link{di2ch}()},
\code{\link{di2we}()},
\code{\link{dist_hamming}()},
\code{\link{rp_lineDist}()},
\code{\link{rp_plot}()},
\code{\link{rp_size}()},
\code{\link{rp}()}
}
\author{
Fred Hasselman
}
\concept{Distance matrix operations (recurrence plot)}
