% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{rp_plot}
\alias{rp_plot}
\title{Plot (thresholded) distance matrix}
\usage{
rp_plot(RM, plotDimensions = FALSE, plotMeasures = FALSE,
  plotRadiusRRbar = TRUE, markEpochsLOI = NULL,
  markEpochsGrid = NULL, radiusValue = NA, title = "", xlab = "",
  ylab = "", plotSurrogate = NA, returnOnlyObject = FALSE,
  useGtable = TRUE)
}
\arguments{
\item{RM}{A distance matrix or recurrence matrix}

\item{plotDimensions}{Should the state vectors be plotted if they are available as attributes of RM (default = \code{TRUE})}

\item{plotMeasures}{Print common (C)RQA measures in the plot if the matrix is binary}

\item{plotRadiusRRbar}{The \code{Radius-RR-bar} is a colour-bar guide plotted with an unthresholded distance matrix indicating a number of \code{RR} values one would get if a certain distance threshold were chosen (\code{default = TRUE})}

\item{markEpochsLOI}{Pass a factor whose levels indicate different epochs or phases in the time series and use the line of identity to represent the levels by different colours (\code{default = NULL})}

\item{markEpochsGrid}{Pass a list 2 numeric vectors, \code{markEpochsGrid[[1]]} should have length equal to NROW(RM) and \code{markEpochsGrid[[2]]} should have length equal to NCOL(RM). The values in the vectors represent different epochs associated with each time stamp, the change index will become 'breaks' on the \code{y} and \code{x} axis respectively (\code{default = NULL})}

\item{radiusValue}{If \code{plotMeasures = TRUE} and RM is an unthresholded matrix, this value will be used to calculate recurrence measures. If \code{plotMeasures = TRUE} and RM is already a binary recurence matrix, pass the radius that was used as a threshold to create the matrix for display purposes. If \code{plotMeasures = TRUE} and \code{radiusValue = NA}, function \code{crqa_radius()} will be called with default settings (find a radius that yields .05 recurrence rate). If \code{plotMeasures = FALSE} this setting will be ignored.}

\item{title}{A title for the plot}

\item{xlab}{An x-axis label}

\item{ylab}{An y-axis label}

\item{plotSurrogate}{Should a 2-panel comparison plot based on surrogate time series be added? If \code{RM} has attributes \code{y1} and \code{y2} containing the time series data (i.e. it was created by a call to \code{\link{rp}}), the following options are available: "RS" (random shuffle), "RP" (randomised phases), "AAFT" (amplitude adjusted fourier transform). If no timeseries data is included, the columns will be shuffled.  NOTE: This is not a surrogate test, just 1 surrogate is created from \code{y1}.}

\item{returnOnlyObject}{Return the ggplot object only, do not draw the plot (default = \code{TRUE})}

\item{useGtable}{Use package \code{\link{gtable}}. If this results in errors (e.g. viewport settings), set set to FALSE to use package \code{patchwork}. This package is in development, see the warning for instructions on how to install it.}
}
\value{
A nice plot of the recurrence matrix.
}
\description{
Plot (thresholded) distance matrix
}
\seealso{
Other Distance matrix operations: \code{\link{bandReplace}},
  \code{\link{di2bi}}, \code{\link{di2we}},
  \code{\link{dist_hamming}}, \code{\link{rp_lineDist}},
  \code{\link{rp_nzdiags}}, \code{\link{rp_size}},
  \code{\link{rp}}
}
\concept{Distance matrix operations}
