% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{rp_prep}
\alias{rp_prep}
\title{Prepare matrix}
\usage{
rp_prep(
  RP,
  emRad = NULL,
  DLmin = 2,
  VLmin = 2,
  HLmin = 2,
  DLmax = length(Matrix::diag(RP)),
  VLmax = length(Matrix::diag(RP)),
  HLmax = length(Matrix::diag(RP)),
  AUTO = FALSE,
  chromatic = FALSE,
  matrices = FALSE,
  doHalf = FALSE
)
}
\arguments{
\item{RP}{Recurrence plot}

\item{emRad}{Radiuc}

\item{DLmin}{Minimal diagonal line length}

\item{VLmin}{Minimal vertical line length}

\item{HLmin}{Minimal horizontal line length}

\item{DLmax}{Maximal diagonal line length}

\item{VLmax}{Maximal vertical line length}

\item{HLmax}{Maximal horizontal line length}

\item{AUTO}{Is this an AUTO RQA?}

\item{chromatic}{Chromatic RQA?}

\item{matrices}{Return Matrices?}

\item{doHalf}{Analyse half of the matrix?}
}
\value{
A prepped matrix
}
\description{
Prepare matrix
}
\keyword{internal}
