% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqa.R
\name{rqa_calc_lineMeasures}
\alias{rqa_calc_lineMeasures}
\title{rqa_lineMeasures}
\usage{
rqa_calc_lineMeasures(
  y1,
  y2 = NA,
  emRad = NA,
  DLmin = 2,
  VLmin = 2,
  HLmin = 2,
  DLmax = NROW(y1),
  VLmax = NROW(y1),
  HLmax = NROW(y1),
  d = NULL,
  theiler = NA,
  recurrenceTimes = FALSE,
  AUTO = NULL,
  chromatic = FALSE,
  matrices = FALSE
)
}
\description{
rqa_lineMeasures
}
\keyword{internal}
