% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqa.R
\name{rqa_diagProfile}
\alias{rqa_diagProfile}
\alias{crqa_diagprofile}
\title{Diagonal Recurrence Profile}
\usage{
rqa_diagProfile(
  y1 = NULL,
  y2 = NULL,
  emDim = 1,
  emLag = 1,
  emRad = NA,
  targetValue = NA,
  diagWin = NULL,
  xname = "X-axis",
  yname = "Y-axis",
  theiler = 0,
  doShuffle = FALSE,
  shuffleWhich = "y1",
  Nshuffle = 19,
  doEmbed = TRUE,
  AUTO = NULL,
  chromatic = FALSE,
  method = c("Euclidean", "SBD")[1],
  doPlot = TRUE,
  minY = NA,
  returnOnlyPlot = FALSE
)
}
\arguments{
\item{y1}{A numeric vector or time series}

\item{y2}{A numeric vector or time series for cross recurrence}

\item{emDim}{The embedding dimensions}

\item{emLag}{The embedding lag}

\item{emRad}{The threshold (emRad) to apply to the distance matrix to create a binary or weighted matrix. If \code{NULL}, an unthresholded matrix will be created (default = \code{NULL})}

\item{targetValue}{A value passed to \code{est_radius(...,type="fixed", targetMeasure="RR")} if \code{is.na(emRad)==TRUE}.}

\item{diagWin}{Window around the line of synchrony}

\item{xname}{Label for x-axis}

\item{yname}{Label for y-axis}

\item{theiler}{Use a \code{theiler} window around the main diagonal (Line of Identity/Synchronisation) to remove auto-correlations at short time-lags:
\itemize{
\item \code{0} will include the main diagonal in all RQA measure calculations.
\item \code{1} will remove the main diagonal from all RQA measure calculations.
\item \code{NA} (default), will check if the matrix is symmetrical , if so, it will remove the diagonal by setting \code{theiler = 1} (Line of Identity, Auto-RQA), if it is not symmetrical (Line of Synchronisation, Cross-RQA) it will set \code{theiler = 0}.
\item A value greater than \code{1} will remove that many diagonals around and including the diagonal from all RQA measure calculations. So \code{theiler = 2} means exclude \code{2} diagonals around the main diagonal, including the main diagonal itself: \verb{[-1,0,1]}.
If \code{theiler} is a numeric vector of \code{length(theiler) == 2} it is possible to exclude an asymmetrical window. The values are interpreted as end points in a sequence of diagonal ID's, e.g. \code{theiler = c(-1,5)} will exclude \verb{[-1,0,1,2,3,4,5]}. If \code{length(theiler) > 2}, the values will be considered individual diagonal ID's, e.g. \code{theiler = c(-3,-1,0,2,5)}, will exclude only those specific ID's. Also see the note.
}}

\item{doShuffle}{Should a shuffled baseline be calculated (default = \code{FALSE})}

\item{shuffleWhich}{Which of the time series should be shuffled: 'y1' or 'y2'? (default = 'y2')}

\item{Nshuffle}{How many shuffled versions to make up the baseline? The default is \code{19}, which is the minimum for a one-sided surrogate test.}

\item{doEmbed}{If \code{doShuffle = TRUE}, should the data in y1 and y2 be considered embedded time series? The temporal order of all columns in \code{y2} will be randomly shuffled in the same way, keeping coordinates together (default = \code{FALSE})}

\item{AUTO}{Auto-recurrence? (default = \code{FALSE})}

\item{chromatic}{Force chromatic RQA? (default = \code{FALSE})}

\item{method}{Distance measure to use. Any option that is valid for argument \code{method} of \code{\link[proxy:dist]{proxy::dist()}}. Type \code{proxy::pr_DB$get_entries()} to see a list of all the options. Common methods are: \verb{"Euclidean", "Manhattan", "Minkowski", "Chebysev"} (or the same but shorter: \verb{"L2","L1","Lp", "max"} distance). To use the shape based distance for phase-based recurrence use \code{"SBD"} (default = \code{"Euclidean"})}

\item{doPlot}{Plot (default = \code{TRUE})}

\item{minY}{The upper limit of the Y-axis. If \code{NA}, the limit is determined by \code{max(minY,max(RR))}. Set to 1 to always show the theoretical maximum (default = \code{NA})}

\item{returnOnlyPlot}{Don't plot to graphics device, but do return the plot (default = \code{FALSE})}
}
\value{
A plot and/or the data for the plot
}
\description{
Diagonal Recurrence Profile
}
