% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqa.R
\name{rqa_fast}
\alias{rqa_fast}
\title{Fast rqa}
\usage{
rqa_fast(
  y1,
  y2 = NA,
  index,
  theiler,
  emRad,
  symmetrical = TRUE,
  diagonal = FALSE,
  method = "Euclidean",
  withSplits = FALSE,
  splitSize = 2048
)
}
\arguments{
\item{y1}{y1}

\item{y2}{y2}

\item{index}{row}

\item{theiler}{theiler}

\item{emRad}{threshold}

\item{symmetrical}{symmetrical}

\item{diagonal}{diagonals}

\item{method}{distance method}

\item{withSplits}{split long vectors into chunks of size splitSize}

\item{splitSize}{size of splits}
}
\value{
list of vectors
}
\description{
Fast rqa
}
