% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqa.R
\name{rqa_getSeries}
\alias{rqa_getSeries}
\title{Get embedded series for rp() and rqa()}
\usage{
rqa_getSeries(
  y1,
  y2 = NULL,
  emDim = 1,
  emLag = 1,
  chromatic = FALSE,
  doEmbed = TRUE
)
}
\arguments{
\item{y1}{y1}

\item{y2}{y2}

\item{emDim}{emDim}

\item{emLag}{emLag}

\item{chromatic}{chromatic}

\item{doEmbed}{doEmbed}
}
\value{
embedded series
}
\description{
Get embedded series for rp() and rqa()
}
\keyword{internal}
