% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqa.R
\name{rqa_par}
\alias{rqa_par}
\title{Massively Parallel RQA analysis}
\usage{
rqa_par(
  y1,
  y2 = NULL,
  emDim = 1,
  emLag = 1,
  emRad = NULL,
  theiler = NA,
  includeDiagonal = NA,
  AUTO = NULL,
  DLmin = 2,
  VLmin = 2,
  HLmin = 2,
  DLmax = NA,
  VLmax = NA,
  HLmax = NA,
  weighted = FALSE,
  weightedBy = "si",
  method = "Euclidean",
  rescaleDist = c("none", "maxDist", "meanDist")[1],
  targetValue = 0.05,
  chromatic = FALSE,
  recurrenceTimes = FALSE,
  doPlot = FALSE,
  doEmbed = TRUE,
  anisotropyHV = FALSE,
  asymmetryUL = FALSE,
  silent = TRUE,
  ...
)
}
\arguments{
\item{y1}{A numeric vector or time series}

\item{y2}{A numeric vector or time series for cross recurrence}

\item{emDim}{The embedding dimensions}

\item{emLag}{The embedding lag}

\item{emRad}{The threshold (emRad) to apply to the distance matrix to create a binary or weighted matrix. If \code{NULL}, an unthresholded matrix will be created (default = \code{NULL})}

\item{theiler}{Use a \code{theiler} window around the main diagonal (Line of Identity/Synchronisation) to remove auto-correlations at short time-lags:
\itemize{
\item \code{0} will include the main diagonal in all RQA measure calculations.
\item \code{1} will remove the main diagonal from all RQA measure calculations.
\item \code{NA} (default), will check if the matrix is symmetrical , if so, it will remove the diagonal by setting \code{theiler = 1} (Line of Identity, Auto-RQA), if it is not symmetrical (Line of Synchronisation, Cross-RQA) it will set \code{theiler = 0}.
\item A value greater than \code{1} will remove that many diagonals around and including the diagonal from all RQA measure calculations. So \code{theiler = 2} means exclude \code{2} diagonals around the main diagonal, including the main diagonal itself: \verb{[-1,0,1]}.
If \code{theiler} is a numeric vector of \code{length(theiler) == 2} it is possible to exclude an asymmetrical window. The values are interpreted as end points in a sequence of diagonal ID's, e.g. \code{theiler = c(-1,5)} will exclude \verb{[-1,0,1,2,3,4,5]}. If \code{length(theiler) > 2}, the values will be considered individual diagonal ID's, e.g. \code{theiler = c(-3,-1,0,2,5)}, will exclude only those specific ID's. Also see the note.
}}

\item{AUTO}{Auto-recurrence? (default = \code{FALSE})}

\item{DLmin}{Minimal diagonal line length (default = \code{2})}

\item{VLmin}{Minimal vertical line length (default = \code{2})}

\item{HLmin}{Minimal horizontal line length (default = \code{2})}

\item{DLmax}{Maximal diagonal line length (default = length of diagonal -1)}

\item{VLmax}{Maximal vertical line length (default = length of diagonal -1)}

\item{HLmax}{Maximal horizontal line length (default = length of diagonal -1)}

\item{weighted}{If \code{FALSE} a binary matrix will be returned. If \code{TRUE} every value larger than \code{emRad} will be \code{0}, but values smaller than \code{emRad} will be retained (default = \code{FALSE})}

\item{weightedBy}{After setting values smaller than \code{emRad} to \code{0}, what should the recurrent values represent? The default is to use the state space similarity (distance/proximity) values as weights (\code{"si"}). Other option are \code{"rt"} for \emph{recurrence time} and \code{"rf"} for \emph{recurrence time frequency} (default = \code{"si"})}

\item{method}{Distance measure to use. Any option that is valid for argument \code{method} of \code{\link[proxy:dist]{proxy::dist()}}. Type \code{proxy::pr_DB$get_entries()} to see a list of all the options. Common methods are: "Euclidean", "Manhattan", "Minkowski", "Chebysev" (or the same but shorter: "L2","L1","Lp" and "max" distance) (default = \code{"Euclidean"})}

\item{rescaleDist}{Should the distance matrix be rescaled? Options are "none", "maxDist" to create a unit scale, "meanScale" to creat z-scores based on the mean distance. (default = \code{"none"})}

\item{targetValue}{A value passed to \code{est_radius(...,type="fixed", targetMeasure="RR")} if \code{is.na(emRad)==TRUE}.}

\item{chromatic}{Perform a chromatic RQA. This assumes the recurring values represent the labels of an unordered categorical variable (default = \code{FALSE})}

\item{recurrenceTimes}{Return measures based on 'white lines', the recurrence times (default = \code{FALSE})}

\item{doPlot}{Plot the matrix by calling \code{\link[=rp_plot]{rp_plot()}} with default settings}

\item{doEmbed}{If \code{FALSE}, a distance matrix will be returned that is not embedded by \code{emDim} and \code{emLag} (Multidimensional RQA). If \code{y1} and/or \code{y2} are data frames, the columns will be used as the state space dimensions (default = \code{TRUE})}

\item{anisotropyHV}{Return anisotropy ratio measures based on Horizontal and Vertical lines. The ratios are calculated as \code{(horizontal - vertical) / (horizontal + vertical)}. So a value of 0 means no anisotropy, negative ratios indicate the measures based on vertical lines had  higher values, positive ratios indicate the measures based on horizontal lines had higher values  (default = \code{FALSE})}

\item{asymmetryUL}{Return asymmetry ratio measures based on Upper and Lower triangles. The ratios are calculated as \code{(upper - lower) / (upper + lower)}. So a value of 0 means no asymmetry, negative ratios indicate the measures based on the lower triangle had the higher values, positive ratios indicate measures based on the upper triangle had higher values (default = \code{FALSE})}

\item{silent}{Silent-ish mode}

\item{...}{Any parameters to pass to \code{\link[=rp_plot]{rp_plot()}} if \code{doPlot = TRUE}}
}
\value{
RQA
}
\description{
Calculate (C)RQA measures without creating a recurrence matrix. Can handle very large time series and requires package \link{parallel} to be installed.
}
