% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqa.R
\name{rqa_plot}
\alias{rqa_plot}
\title{Plot large RP}
\usage{
rqa_plot(RM, imagePath = NA, wMax = 2048, hMax = 2048)
}
\arguments{
\item{RM}{RM}

\item{imagePath}{imagePath}

\item{wMax}{wMax}

\item{hMax}{hMax}
}
\value{
image
}
\description{
Plot large RP
}
\keyword{internal}
