% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{sa2fd.default}
\alias{sa2fd.default}
\title{sa2df.default}
\usage{
\method{sa2fd}{default}(sa, ...)
}
\arguments{
\item{sa}{Self-affinity parameter}

\item{...}{Other arguments}
}
\description{
sa2df.default
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}
}
\seealso{
Other SA to FD converters: \code{\link{sa2fd.dfa}},
  \code{\link{sa2fd.psd}}
}
\author{
Fred Hasselman
}
\keyword{internal}
