% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R, R/nlRtsa_SOURCE.old.R
\name{sa2fd.sda}
\alias{sa2fd.sda}
\alias{sa2fd.sda}
\title{Informed Dimension estimate from SDA slope.}
\usage{
\method{sa2fd}{sda}(sa)

\method{sa2fd}{sda}(sa)
}
\arguments{
\item{sa}{Self-afinity parameter estimate based on SDA slope (e.g., \code{\link{fd.sda}})).}

\item{sa}{Self-afinity parameter estimate based on SDA slope (e.g., \code{\link{fd.sda}})).}
}
\value{
An informed estimate of the Fractal Dimension, see Hasselman(2013) for details.

An informed estimate of the Fractal Dimension, see Hasselman(2013) for details.
}
\description{
Conversion formula: Standardised Dispersion Analysis (SDA) estimate of self-affinity parameter (\code{SA}) to an informed estimate of the fractal dimension (FD).

Conversion formula: Standardised Dispersion Analysis (SDA) estimate of self-affinity parameter (\code{SA}) to an informed estimate of the fractal dimension (FD).
}
\details{
Note that for some signals different PSD slope values project to a single SDA slope. That is, SDA cannot distinguish dplyr::between all variaties of power-law scaling in the frequency domain.

Note that for some signals different PSD slope values project to a single SDA slope. That is, SDA cannot distinguish dplyr::between all variaties of power-law scaling in the frequency domain.
}
\examples{
# Informed FD of white noise
sa2fd.sda(-0.5)

# Informed FD of Brownian noise
sa2fd.sda(-1)

# Informed FD of blue noise
sa2fd.sda(-0.9)
# Informed FD of white noise
sa2fd.sda(-0.5)

# Informed FD of Brownian noise
sa2fd.sda(-1)

# Informed FD of blue noise
sa2fd.sda(-0.9)
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}

Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}
}
\seealso{
Other SA to FD converters: \code{\link{sa2fd.default}},
  \code{\link{sa2fd.dfa}}, \code{\link{sa2fd.psd}}

Other SA to FD converters: \code{\link{sa2fd.default}},
  \code{\link{sa2fd.dfa}}, \code{\link{sa2fd.psd}}
}
\author{
Fred Hasselman

Fred Hasselman
}
