% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{sa2fd_dfa}
\alias{sa2fd_dfa}
\title{Informed Dimension estimate from DFA slope (H)}
\usage{
sa2fd_dfa(sa, ...)
}
\arguments{
\item{sa}{Self-Afinity parameter estimate based on DFA slope (e.g., \code{\link[=fd_sda]{fd_sda()}})).}

\item{...}{Other arguments}
}
\value{
An informed estimate of the Fractal Dimension, see Hasselman(2013) for details.
}
\description{
Conversion formula: Detrended Fluctuation Analysis (DFA) estimate of the Hurst exponent (a self-affinity parameter \code{sa}) to an informed estimate of the (fractal) dimension (FD).
}
\details{
The DFA slope (H) will be converted to a dimension estimate using:

\deqn{D_{DFA}\approx 2-(\tanh(\log(3)*sa)) }{D_{DFA} ~ 2-(tanh(log(3)*sa)) }
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. https://doi.org/10.3389/fphys.2013.00075
}
\author{
Fred Hasselman
}
