% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{sa2fd_sda}
\alias{sa2fd_sda}
\title{Informed Dimension estimate from SDA slope.}
\usage{
sa2fd_sda(sa, ...)
}
\arguments{
\item{sa}{Self-afinity parameter estimate based on SDA slope (e.g., \code{\link[=fd_sda]{fd_sda()}})).}

\item{...}{Other arguments}
}
\value{
An informed estimate of the Fractal Dimension, see Hasselman(2013) for details.
}
\description{
Conversion formula: Standardised Dispersion Analysis (SDA) estimate of self-affinity parameter (\code{SA}) to an informed estimate of the fractal dimension (FD).
}
\details{
Note that for some signals different PSD slope values project to a single SDA slope. That is, SDA cannot distinguish dplyr::between all variaties of power-law scaling in the frequency domain.
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. https://doi.org/10.3389/fphys.2013.00075
}
\author{
Fred Hasselman
}
