% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{scale.R}
\alias{scale.R}
\title{Scale.R}
\usage{
\method{scale}{R}(x, mn = min(x, na.rm = T), mx = max(x, na.rm = T),
  lo = 0, hi = 1)
}
\arguments{
\item{x}{Input vector or data frame.}

\item{mn}{Minimum value of original, defaults to \code{min(x, na.rm = TRUE)}.}

\item{mx}{Maximum value of original, defaults to \code{max(x, na.rm = TRUE)}.}

\item{lo}{Maximum value to rescale to, defaults to \code{1}.}

\item{hi}{Minimum value to rescale to, defaults to \code{0}.}
}
\description{
Rescale a vector to a user defined range defined by user.
}
\details{
Three uses:
\enumerate{
\item scale.R(x)             - Scale x to data range: min(x.out)==0;      max(x.out)==1
\item scale.R(x,mn,mx)       - Scale x to arg. range: min(x.out)==mn==0;  max(x.out)==mx==1
\item scale.R(x,mn,mx,lo,hi) - Scale x to arg. range: min(x.out)==mn==lo; max(x.out)==mx==hi
}
}
\examples{
# Works on numeric objects
somenumbers <- cbind(c(-5,100,sqrt(2)),c(exp(1),0,-pi))

scale.R(somenumbers)
scale.R(somenumbers,mn=-100)
# Values < mn will return < lo (default=0)
# Values > mx will return > hi (default=1)
scale.R(somenumbers,mn=-1,mx=99)

scale.R(somenumbers,lo=-1,hi=1)
scale.R(somenumbers,mn=-10,mx=101,lo=-1,hi=4)
}
