% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_log}
\alias{scale_log}
\title{scale_log}
\usage{
scale_log(
  scale.min,
  scale.max,
  scale.ratio = 2,
  scale.res = NULL,
  coerce = NULL
)
}
\arguments{
\item{scale.min}{min}

\item{scale.max}{max}

\item{scale.ratio}{ratio}

\item{scale.res}{reolution}

\item{coerce}{coerce?}
}
\value{
output
}
\description{
Adaptation from discontinued package \code{ifultools} by William Constantine.
}
\keyword{internal}
