% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{SDA}
\alias{SDA}
\title{Standardised Dispersion Analysis}
\usage{
SDA(x, front = FALSE)
}
\arguments{
\item{x}{x}

\item{front}{front}
}
\value{
output
}
\description{
Standardised Dispersion Analysis
}
\note{
Based on "dispersion" from discontinued package fractal.
}
\keyword{internal}
