% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_command_line_crp.R
\name{set_command_line_rp}
\alias{set_command_line_rp}
\title{Set command line RQA executable}
\usage{
set_command_line_rp()
}
\value{
Message informing whether the procedure was succesful.
}
\description{
Sets the location of the command line RQA executable
}
