% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOps.R
\name{setTheiler}
\alias{setTheiler}
\title{Set theiler window on a distance matrix or recurrence matrix.}
\usage{
setTheiler(RM, theiler = NA, silent = FALSE)
}
\arguments{
\item{RM}{A distance matrix (set \code{emRad = NA} to estimate a radius), or a matrix of zeroes and ones}

\item{theiler}{Use a \code{theiler} window around the main diagonal (Line of Identity/Synchronisation) to remove auto-correlations at short time-lags:
\itemize{
\item \code{0} will include the main diagonal in all RQA measure calculations.
\item \code{1} will remove the main diagonal from all RQA measure calculations.
\item \code{NA} (default), will check if the matrix is symmetrical , if so, it will remove the diagonal by setting \code{theiler = 0} (Line of Identity, Auto-RQA), if it is not symmetrical (Line of Synchronisation, Cross-RQA) it will set \code{theiler = 1}.
\item A value greater than \code{1} will remove that many diagonals around and including the diagonal from all RQA measure calculations. So \code{theiler = 2} means exclude \code{2} diagonals around the main diagonal, including the main diagonal itself: \verb{[-1,0,1]}.
If \code{theiler} is a numeric vector of \code{length(theiler) == 2} it is possible to exclude an asymmetrical window. The values are interpreted as end points in a sequence of diagonal ID's, e.g. \code{theiler = c(-1,5)} will exclude \verb{[-1,0,1,2,3,4,5]}. If \code{length(theiler) > 2}, the values will be considered individual diagonal ID's, e.g. \code{theiler = c(-3,-1,0,2,5)}, will exclude only those specific ID's. Also see the note.
}}

\item{silent}{Silent-ish mode}
}
\value{
The matrix with the diagonals indicated in the \code{theiler} argument set to either \code{max(RM)+1} (if \code{RM} is a distance matrix) or \code{0} (if \code{RM} is a recurrence matrix).
}
\description{
Set theiler window on a distance matrix or recurrence matrix.
}
