% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{sliceTS}
\alias{sliceTS}
\title{Slice columns of a matrix in epochs}
\usage{
sliceTS(TSmat, epochSz = 1)
}
\arguments{
\item{TSmat}{A matrix with tieseries as columns}

\item{epochSz}{Epoch size}
}
\value{
A list with epochs
}
\description{
Slice columns of a matrix in epochs
}
