% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ssg_gwf2long}
\alias{ssg_gwf2long}
\title{Import GrdWare files}
\usage{
ssg_gwf2long(gwf_name, delta_t = 0.01, returnOnlyData = TRUE,
  saveLongFormat = FALSE)
}
\arguments{
\item{gwf_name}{Name of the GridWare project file. A directory named \code{../gwf_name_trjs} must be present at the location of the project file.}

\item{delta_t}{Time between two samples or sampling frequency}

\item{returnOnlyData}{Just return the data, do not return a list object with data, variable info and preferences.}

\item{saveLongFormat}{Save the long format trajectory data as a \code{.csv} file in the same location as \code{gwf_name}}
}
\value{
A data frame containing State Space Grid trajectories, or a list object with additional info.
}
\description{
Import GrdWare files
}
