% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{symHelper}
\alias{symHelper}
\title{Correct for plateaus in symbolic series}
\usage{
symHelper(sym_num)
}
\arguments{
\item{sym_num}{}
}
\value{
data frame
}
\description{
Correct for plateaus in symbolic series
}
\keyword{internal}
