% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R, R/nlRtsa_SOURCE.old.R
\name{tau}
\alias{tau}
\alias{tau}
\title{tau}
\usage{
tau(y, selection.methods = c("first.minimum"), maxLag = length(y)/4, ...)

tau(y, selection.methods = c("first.minimum"), maxLag = length(y)/4, ...)
}
\arguments{
\item{y}{Time series}

\item{selection.methods}{Selecting an optimal embedding lag (default: Return "first.e.decay", "first.zero", "first.minimum", "first.value", where value is 1/exp(1))}

\item{maxLag}{Maximal lag to consider (default: 1/5 of timeseries length)}

\item{...}{Additional parameters}

\item{value}{Threshold value for selection method first.value (default: 1/exp(1))}

\item{nbins}{Number of bins for average mutual information function (default: 1/3 of number of unique values in timeseries)}

\item{y}{Time series}

\item{selection.methods}{Selecting an optimal embedding lag (default: Return "first.e.decay", "first.zero", "first.minimum", "first.value", where value is 1/exp(1))}

\item{value}{Threshold value for selection method first.value (default: 1/exp(1))}

\item{maxLag}{Maximal lag to consider (default: 1/5 of timeseries length)}

\item{nbins}{Number of bins for average mutual information function (default: 1/3 of number of unique values in timeseries)}

\item{...}{Additional parameters}
}
\value{
The ami function with minima

The ami function with minima
}
\description{
A wrapper for nonlinearTseries::timeLag

A wrapper for nonlinearTseries::timeLag
}
