% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{transM}
\alias{transM}
\title{Estimate the transition matrix for computing the critical threshold values on
mutual information.}
\usage{
transM(x, n_bins)
}
\arguments{
\item{x}{A \code{vector} of values.}

\item{n_bins}{The number of bins for the entropy calculation.}
}
\value{
A \code{list} with the following components:
\describe{
\item{\code{xn}}{An \code{[n x 2]} matrix of the original and discretized vectors.}
\item{\code{MM}}{Transition probability matrix from bin-i to bin-j.}
}
}
\description{
\code{mutual_info} is a helper function for estimating the transition matrix
used in creating resampled vectors for the (1 - alpha)\% critical threshold
value on the mutual info.
}
\author{
Mark Scheuerell (https://mdscheuerell.github.io/muti/)
}
\keyword{internal}
