% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{trim_fill}
\alias{trim_fill}
\title{Trim of Fill vector}
\usage{
trim_fill(x, y, action = c("fill", "trim.cut", "trim.NA")[1],
  type = c("end", "center", "front")[1], padding = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}

\item{action}{Use \code{"fill"} to fill the shortest vector with \code{padding} (default); \code{"trim.cut"} to trim the longest vector to the length of the shortest; \code{"trim.NA"} to fill the longest vector with \code{NA}}

\item{type}{Should trimming or filling take place at the \code{"end"} (default), or \code{"front"} of the vector? The option \code{"center"} will try to distribute trimming by \code{NA} or filling by \code{padding} evenly across the front and end of the vector.}

\item{padding}{A value to use for padding (default = \code{0})}
}
\value{
A list with two vectors of equal length.
}
\description{
Trim the largest vector by cutting it, or filling it with `NA`.
Fill the shortest vector with padding.
}
\seealso{
il_mi
}
