% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R, R/nlRtsa_SOURCE.old.R
\name{try.CATCH}
\alias{try.CATCH}
\alias{try.CATCH}
\title{tryCatch both warnings (with value) and errors}
\usage{
try.CATCH(expr)

try.CATCH(expr)
}
\arguments{
\item{expr}{an \R expression to evaluate}

\item{expr}{an \R expression to evaluate}
}
\value{
a list with 'value' and 'warning', where value' may be an error caught.

a list with 'value' and 'warning', where value' may be an error caught.
}
\description{
TRY ... CATCH

TRY ... CATCH
}
\details{
In longer simulations, aka computer experiments,
 you may want to
 1) catch all errors and warnings (and continue)
 2) store the error or warning messages

 Here's a solution  (see \R-help mailing list, Dec 9, 2010):

Catch *and* save both errors and warnings, and in the case of
a warning, also keep the computed result.

In longer simulations, aka computer experiments,
 you may want to
 1) catch all errors and warnings (and continue)
 2) store the error or warning messages

 Here's a solution  (see \R-help mailing list, Dec 9, 2010):

Catch *and* save both errors and warnings, and in the case of
a warning, also keep the computed result.
}
\author{
Martin Maechler; Copyright (C) 2010-2012  The R Core Team

Martin Maechler; Copyright (C) 2010-2012  The R Core Team
}
\keyword{internal}
