% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_changeindex}
\alias{ts_changeindex}
\title{Find change indices}
\usage{
ts_changeindex(y, returnRectdata = TRUE, groupVar = NULL,
  labelVar = NULL, discretize = FALSE, nbins = 5)
}
\arguments{
\item{y}{An indicator variable representing different levels of a variable or factor}

\item{returnRectdata}{Return a dataframe suitable for shading a \code{ggplot2} graph with \code{\link[ggplot2]{geom_rect}}}

\item{groupVar}{Pass a value (length 1) or variable (length of y) that can be used as a variable to join the indices by if \code{returnRectdata = TRUE}}

\item{labelVar}{If \code{y} is not a character vector, provide a vector of labels equal to \code{length(y)}}

\item{discretize}{If \code{y} is a continuous variable, setting \code{discretize = TRUE} will partition the values of \code{y} into \code{nbins} number of bins, each value of \code{y} will be replaced by its bin number.}

\item{nbins}{Number of bins to use to change a continuous \code{y} (if \code{discretize = TRUE}) into a variable with \code{nbins} levels}
}
\value{
Either a vector with the indices of change in \code{y}, or, a data frame with variables \code{xmin,xmax,ymin,ymax,label}
}
\description{
Find change indices
}
