% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_checkfix}
\alias{ts_checkfix}
\title{Check and Fix a vector}
\usage{
ts_checkfix(y, checkNumericVector = TRUE, checkTimeVector = TRUE,
  checkPow2 = TRUE, checkScale = TRUE, checkSummationOrder = TRUE,
  checkNonStationarity = TRUE, checkNonHomogeneity = TRUE,
  fixNumericVector = FALSE, fixTimeVector = FALSE, fixPow2 = FALSE,
  fixNA = TRUE, fixScale = FALSE, fixSummationOrder = FALSE,
  fixNonStationarity = FALSE, fixNonHomogeneity = FALSE)
}
\arguments{
\item{y}{A time series object or numeric vector}

\item{checkNumericVector}{checkVector}

\item{checkTimeVector}{checkTimeVector}

\item{checkPow2}{checkPow2}

\item{checkScale}{checkScale}

\item{checkSummationOrder}{checkSummationOrder}

\item{checkNonStationarity}{checkNonStationarity}

\item{checkNonHomogeneity}{checkNonHomogeneity}

\item{fixTimeVector}{fixTimeVector}

\item{fixPow2}{foxPow2}

\item{fixNA}{fixNA}

\item{fixScale}{fixScale}

\item{fixSummationOrder}{fixSummationOrder}

\item{fixNonStationarity}{fixNonStationarity}

\item{fixNonHomogeneity}{fixNonHomogeneity}

\item{fixVector}{fixVector}
}
\value{
A 'check' report and/or a 'fixed' vector y.
}
\description{
Check and Fix a vector
}
