% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_coarsegrain}
\alias{ts_coarsegrain}
\title{Course grain a time series}
\usage{
ts_coarsegrain(
  y,
  grain = 2,
  summaryFunction = c("mean", "median", "min", "max")[1],
  retainLength = FALSE
)
}
\arguments{
\item{y}{A numeric vector}

\item{grain}{The bin size in which to summarise the values (default = \code{2})}

\item{summaryFunction}{How should the data be summarized in the bins?}

\item{retainLength}{Return only the bin values (\code{FALSE}), or retain the length of the original series? (default = \code{TRUE})}
}
\value{
A coarse grained version of \code{y}.
}
\description{
Generate a course grained version of a time series by summarising values into bins.
}
\examples{

set.seed(1234)
y <- rnorm(100)
y1 <- ts_coarsegrain(y, grain = 3)
y2 <- ts_coarsegrain(y, grain = 3, retainLength = TRUE)
y3 <- ts_coarsegrain(y, grain = 3, retainLength = TRUE, summaryFunction = "max")

t1 <- seq(1,length(y), by = 3)

plot(t1+1, y1, col = "red3", type = "l", ylim = c(-3,3), xlab = "time", ylab = "Y")
lines(y, col = "grey70")
lines(y2, col = "steelblue")
lines(y3, col = "green3")
legend(60, -1.3, legend=c("Original", "Mean", "Mean + Retain Length", "Max + Retain Length"),
lty = 1, col=c("grey70", "red3", "steelblue","green3"), cex = 0.7)


}
