% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_cp}
\alias{ts_cp}
\title{Change Profile}
\usage{
ts_cp(y, win, align = "right", keepNA = TRUE)
}
\arguments{
\item{y}{Time series}

\item{win}{A window in which}

\item{align}{Alignment of the window, see \code{\link[zoo:rollapply]{zoo::rollapply()}} (default = \code{"right"})}

\item{keepNA}{Remove \code{NA} or return \code{y} with \code{NA} (default = \code{TRUE})}
}
\value{
Transformed time series
}
\description{
Change Profile
}
\references{
Hasselman, F. and Bosman, A.M.T. (2020). Studying Complex Adaptive Systems With Internal States: A Recurrence Network Approach to the Analysis of Multivariate Time-Series Data Representing Self-Reports of Human Experience. Frontiers of Applied Mathematics and Statistics, 6:9. doi: 10.3389/fams.2020.00009
}
