% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_detrend}
\alias{ts_detrend}
\title{Detrend a time series}
\usage{
ts_detrend(y, order = 1)
}
\arguments{
\item{y}{A time series ot numeric vector}

\item{order}{order Order of polynomial trend to remove}
}
\value{
Residuals after detrending polynomial of order \code{order}
}
\description{
Detrend a time series
}
\seealso{
Other Time series operations: \code{\link{ts_center}},
  \code{\link{ts_levels}}, \code{\link{ts_peaks}},
  \code{\link{ts_slice}}, \code{\link{ts_standardise}},
  \code{\link{ts_trimfill}}
}
\author{
Fred Hasselman
}
