% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_discrete}
\alias{ts_discrete}
\title{Discrete representation}
\usage{
ts_discrete(y, nbins = ceiling(2 * NROW(y)^(1/3)), keepNA = TRUE)
}
\arguments{
\item{y}{Numeric vector or time series to be discretised.}

\item{nbins}{Number of bins to use for calculating transfer probabilities (default = \code{ceiling(2*length(y)^(1/3))})}

\item{keepNA}{If \code{TRUE}, any \code{NA} values will first be removed and later re-inserted into the discretised time series.}
}
\value{
A discretised version of \code{y}
}
\description{
Return a discrete representation of \code{y} by binning the observed values and returning the transfer probabilities.
}
