% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_rasterize}
\alias{ts_rasterize}
\title{Turn a 1D time series vector into a 2D curve}
\usage{
ts_rasterize(y, unitSquare = FALSE, toSparse = TRUE, resolution = 2)
}
\arguments{
\item{y}{A 1D time series object or numeric verctor.}

\item{unitSquare}{Convert the series to a unit square? (default = \code{FALSE})}

\item{toSparse}{Convert to sparse Matrix (default = \code{FALSE})}

\item{resolution}{Factor by which dimensions will be multiplied (default = \code{2})}
}
\value{
A (sparse) matrix representing the time series as a curve in 2D space
}
\description{
Turn a 1D time series vector into a 2D curve
}
\examples{

y <- rnorm(100)
plot(ts(y))

y_img <- ts_rasterize(y)
image(y_img,col=c("white","black"))


}
