% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_sd}
\alias{ts_sd}
\title{Standard Deviation estimates}
\usage{
ts_sd(y, na.rm = TRUE, type = c("Bessel", "unadjusted")[1],
  silent = TRUE)
}
\arguments{
\item{y}{Time series or numeric vector}

\item{na.rm}{Remove missing values}

\item{type}{Apply Bessel's correction (divide by N-1) or return unadjusted SD (divide by N)}

\item{silent}{Silent-ish mode (default = \code{TRUE})}
}
\value{
Standard deviation of \code{y}
}
\description{
Calculates the population estimate of the standard deviation, or the unadjusted standard deviation.
}
