% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_slice}
\alias{ts_slice}
\title{Slice columns of a matrix in epochs}
\usage{
ts_slice(y, epochSz = 4)
}
\arguments{
\item{y}{A matrix with timeseries as columns}

\item{epochSz}{Epoch size}
}
\value{
A list with epochs
}
\description{
Slice columns of a matrix in epochs
}
\seealso{
Other Time series operations: \code{\link{ts_center}},
  \code{\link{ts_detrend}}, \code{\link{ts_embed}},
  \code{\link{ts_levels}}, \code{\link{ts_peaks}},
  \code{\link{ts_standardise}}, \code{\link{ts_trimfill}}
}
\author{
Fred Hasselman
}
\concept{Time series operations}
