% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_slope}
\alias{ts_slope}
\title{Calculate Kendall's tau in sliding window or around change point.}
\usage{
ts_slope(y, win = NA, changepoint = NA, doPlot = FALSE)
}
\arguments{
\item{y}{A numeric vector}

\item{win}{Numeric vector with 1 or 2 values. If \code{changepoint != NA}, Kendall's tau will be calculated in a window around the change point. If one value is provided a symmetric window, otherwise \code{changepoint - win[1]} and \code{changepoint + win[2]}. If \code{changepoint == NA} Kendall's tau will be calculated in a sliding  window.}

\item{changepoint}{If not \code{NA}, it has to be an index of \code{y}. If \code{win} is \code{NA} Kendall's tau will be calculated in \code{1:changepoint}, otherwise the values of \code{win} will be used to create a window. If both \code{changepoint} and \code{win} are \code{NA} Kendall's tau will be calculated on all of \code{y}.}

\item{doPlot}{provide a plot}
}
\value{
A data frame
}
\description{
Kendall's tau at different lags or change point ranges
}
\examples{

ts_slope(y=rnorm(100), doPlot = TRUE)

}
