% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_standardise}
\alias{ts_standardise}
\title{Standardise a vector}
\usage{
ts_standardise(y, na.rm = TRUE, type = c("mean.sd", "median.mad")[1],
  adjustN = TRUE)
}
\arguments{
\item{y}{A time series or numeric vector}

\item{na.rm}{Set the \code{na.rm} field}

\item{type}{Center on the \code{"mean"} and divide by \code{sd} (default), or center on \code{"median"} and divide by \code{mad}}

\item{adjustN}{Use the population SD estimator (\code{N-1}), or \code{N} (default = \code{TRUE})}
}
\value{
A standardised vector
}
\description{
Standardise a vector
}
\seealso{
Other Time series operations: \code{\link{ts_center}},
  \code{\link{ts_detrend}}, \code{\link{ts_embed}},
  \code{\link{ts_levels}}, \code{\link{ts_peaks}},
  \code{\link{ts_slice}}, \code{\link{ts_trimfill}}
}
\author{
Fred Hasselman
}
\concept{Time series operations}
