% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_symbolic}
\alias{ts_symbolic}
\title{Symbolic representation}
\usage{
ts_symbolic(y, keepNA = TRUE, usePlateaus = FALSE, doPlot = FALSE)
}
\arguments{
\item{y}{Numeric vector or time series to be discretised.}

\item{keepNA}{If \code{TRUE}, any \code{NA} values will first be removed and later re-inserted into the discretised time series.}

\item{usePlateaus}{Treat consequative "same" values after "peak" or "trough" as a "peak"/"trough".}

\item{doPlot}{Create a plot of the symbolized series.}
}
\value{
A discretised version of \code{y}
}
\description{
Return a discrete representation of \code{y} by binning the observed values and returning the transfer probabilities.
}
