% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_windower}
\alias{ts_windower}
\title{Get sliding window indices}
\usage{
ts_windower(y, win = length(y), step = round(win/2), overlap = NA,
  adjustY = NA)
}
\arguments{
\item{y}{A time series or numeric vector}

\item{win}{Size of the window to slide across \code{y}}

\item{step}{Size of steps between windows. Can be larger than \code{win}, but is ignored if \code{overlap} is not {NA}.}

\item{overlap}{A value between \code{[0 .. 1]}. If overlap is not \code{NA} (default), the value of \code{step} is ignored and set to \code{floor(overlap*win)}. This produces indices in which the size of \code{step} is always smaller than \code{win}, e.g. for fluctuation analyses that use binning procedures to represent time scales.}

\item{adjustY}{If not \code{NA}, or, \code{FALSE} a list object with fields that match one or more arguments of \link[casnet]{ts_trimfill} (except for \code{x,y}), e.g. \code{list(action="trim.NA",type="end",padding=NA,silent=TRUE)}. See \code{Return value} below for details.}
}
\value{
If \code{adjustY = FALSE}, or, a list object with fields that represent arguments of \link[casnet]{ts_trimfill}, then the (adjusted) vector \code{y} is returned with an attribute \code{"windower"}. This is a list object with fields that contain the indices for each window that fits on \code{y}, given \code{win}, \code{step} or \code{overlap} and the settings of \code{adjustY}. If \code{adjustY = NA}, only the list object is returned.
}
\description{
Get sliding window indices
}
