% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{win_slide}
\alias{win_slide}
\title{Sliding window analysis}
\usage{
win_slide(FUN = c("rp", "rp_measures", "rp_diagprofile", "rn",
  "rn_measures")[1], y1, y2 = NULL, win = min(length(y1),
  ifelse(is.null(y2), (length(y1) + 1), length(y2)), na.rm = TRUE),
  step = win, animate = FALSE, silent = TRUE, ...)
}
\arguments{
\item{FUN}{Which function should be used in the sliding window analysis? Options are: \code{rp}, \code{rp_measures}, \code{rp_diagprofile}, \code{rn}, \code{rn_measures} (default = \code{rp})}

\item{y1}{A numeric vector or time series}

\item{y2}{A numeric vector or time series (optional)}

\item{win}{Window size in which to evaluate \code{fun} (default = minimum of length of \code{y1} or \code{y2})}

\item{step}{Stepsize for sliding windows (default = size of \code{win}, so no sliding window)}

\item{animate}{If \code{TRUE} and the function argument \code{doPlot = TRUE}, an animation of the graph output for subsequent windows will be generated using package \link{gganimate} (default = \code{FALSE})}

\item{silent}{Silent-ish mode}

\item{...}{Any parameters to pass to the function \code{fun}}
}
\value{
A list object with the output of \code{fun} for each window at each step.
}
\description{
Sliding window analysis
}
\keyword{internal}
