% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{as.numeric_character}
\alias{as.numeric_character}
\title{Character vector to named numeric vector}
\usage{
as.numeric_character(x, sortUnique = FALSE, keepNA = FALSE)
}
\arguments{
\item{x}{A character vector}

\item{sortUnique}{Should the unique character values be sorted? (default = \code{FALSE})}

\item{keepNA}{Keep NA values (\code{TRUE}), or remove them (default = \code{FALSE})}
}
\value{
A named numeric vector
}
\description{
Converts a character vector to a named numeric vector, with the character elements as names.
}
\examples{

f <- letters
as.numeric_character(f)

}
