% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{as.numeric_factor}
\alias{as.numeric_factor}
\title{Numeric factor to numeric vector}
\usage{
as.numeric_factor(x, keepNA = FALSE, sortUnique = FALSE)
}
\arguments{
\item{x}{A factor based on numeric values.}

\item{keepNA}{Keep NA values (\code{TRUE}), or remove them (default = \code{FALSE})}

\item{sortUnique}{Should the unique character values be sorted? (default = \code{FALSE})}
}
\value{
A numeric vector with factor levels as names.
}
\description{
Converts a factor with numeric levels to a numeric vector, using the values of the levels.
}
\examples{

f <- factor(round(runif(10,0,9)))
as.numeric_factor(f)

# Add NAs
f <- factor(c(round(runif(9,0,9)),NA))
as.numeric_factor(f)
as.numeric_factor(f, keepNA = TRUE)


}
