% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R, R/nlRtsa_SOURCE.old.R
\name{bandReplace}
\alias{bandReplace}
\alias{bandReplace}
\title{bandReplace}
\usage{
bandReplace(mat, lower, upper, value = NA)

bandReplace(mat, lower, upper, value = NA)
}
\arguments{
\item{mat}{A Matrix}

\item{lower}{Lower diagonal to be included in the band (should be $\\leq 0$)}

\item{upper}{Upper diagonal to be included in the band (should be $\\geq 0$)}

\item{value}{A single value to replace all values in the selected band}

\item{mat}{A Matrix}

\item{lower}{Lower diagonal to be included in the band (should be $\\leq 0$)}

\item{upper}{Upper diagonal to be included in the band (should be $\\geq 0$)}

\item{value}{A single value to replace all values in the selected band}
}
\value{
A matrix in which the values in the selected diagonals have been replaced

A matrix in which the values in the selected diagonals have been replaced
}
\description{
Sets a band of matrix diagonals to any given value

Sets a band of matrix diagonals to any given value
}
\examples{
# Create a 10 by 10 matrix
library(Matrix)
m <- Matrix(rnorm(10),10,10)

bandReplace(m,-1,1,0)   # Replace diagonal and adjacent bands with 0 (Theiler window of 1)
# Create a 10 by 10 matrix
library(Matrix)
m <- Matrix(rnorm(10),10,10)

bandReplace(m,-1,1,0)   # Replace diagonal and adjacent bands with 0 (Theiler window of 1)
}
