% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOps.R
\name{bandReplace}
\alias{bandReplace}
\title{Replace matrix diagonals}
\usage{
bandReplace(mat, lower, upper, value = NA, silent = TRUE)
}
\arguments{
\item{mat}{A Matrix}

\item{lower}{Lower diagonal to be included in the band (should be \eqn{\le 0})}

\item{upper}{Upper diagonal to be included in the band (should be \eqn{\ge 0})}

\item{value}{A single value to replace all values in the selected band (default = \code{NA})}

\item{silent}{Operate in silence, only (some) warnings will be shown (default = \code{TRUE})}
}
\value{
A matrix in which the values in the selected diagonals have been replaced
}
\description{
Sets a band of matrix diagonals to any given value
}
\examples{
# Create a 10 by 10 matrix
library(Matrix)
m <- Matrix(rnorm(10),10,10)

bandReplace(m,-1,1,0)   # Replace diagonal and adjacent bands with 0 (Theiler window of 1)
}
\seealso{
Other Distance matrix operations (recurrence plot): 
\code{\link{di2bi}()},
\code{\link{di2ch}()},
\code{\link{di2we}()},
\code{\link{dist_hamming}()},
\code{\link{rp_lineDist}()},
\code{\link{rp_nzdiags}()},
\code{\link{rp_plot}()},
\code{\link{rp_size}()},
\code{\link{rp}()}
}
\author{
Fred Hasselman
}
\concept{Distance matrix operations (recurrence plot)}
