% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet-package.r
\docType{package}
\name{casnet-package}
\alias{casnet}
\alias{casnet-package}
\title{casnet: A Toolbox for Studying Complex Adaptive Systems and Networks}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A collection of analytic tools for studying signals recorded from complex adaptive systems or networks: (Cross) Recurrence Quantification Analyses (Continuous and Categorical (C)RQA, Chromatic RQA, Anisotropic RQA). Fluctuation Analyses (DFA, PSD slope, SDA, Multifractal DFA, Wavelet Singularity Spectrum). Network based time series analyses and visualisation (Recurrence Networks, spiral graphs). Multivariate (network based) time series analyses (Dynamic Complexity, Multiplex Recurrence Networks).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FredHasselman/casnet}
  \item Report bugs at \url{https://github.com/FredHasselman/casnet/issues}
}

}
\author{
\strong{Maintainer}: Fred Hasselman \email{f.hasselman@pwo.ru.nl} (\href{https://orcid.org/0000-0003-1384-8361}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Merlijn Olthof \email{m.olthof@pwo.ru.nl} [contributor]
}

}
\keyword{internal}
