% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_cl_main}
\alias{crqa_cl_main}
\title{crqa_cl_main}
\usage{
crqa_cl_main(data, emDim = 1, emLag = 1, emRad = NA, DLmin = 2,
  VLmin = 2, theiler = 0, win = min(length(y1), ifelse(is.null(y2),
  (length(y1) + 1), length(y2)), na.rm = TRUE), step = step, JRP = FALSE,
  distNorm = c("EUCLIDEAN", "MAX", "MIN", "OP")[[1]], returnMeasures = TRUE,
  returnRPvector = FALSE, returnLineDist = FALSE,
  plot_recmat = c("noplot", "recmat", "distmat")[[1]],
  path_to_rp = getOption("casnet.path_to_rp"), saveOut = FALSE,
  path_out = NULL, file_ID = NULL, silent = TRUE, targetValue = 0.05,
  useParallel = FALSE, ...)
}
\arguments{
\item{emDim}{y1}

\item{emLag}{y1}

\item{emRad}{y1}

\item{DLmin}{y1}

\item{VLmin}{y1}

\item{theiler}{y1}

\item{win}{y1}

\item{step}{y1}

\item{JRP}{y1}

\item{distNorm}{y1}

\item{returnMeasures}{y1}

\item{returnRPvector}{y1}

\item{returnLineDist}{y1}

\item{path_to_rp}{y1}

\item{saveOut}{y1}

\item{path_out}{y1}

\item{file_ID}{y1}

\item{silent}{y1}

\item{targetValue}{y1}

\item{...}{}

\item{y1}{y1}

\item{y2}{y1}

\item{returnDistance}{y1}
}
\description{
crqa_cl_main
}
\keyword{internal}
