% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_diagPofile}
\alias{crqa_diagPofile}
\title{Diagonal Recurrence Profile}
\usage{
crqa_diagPofile(RM, diagWin = NULL, xname = "", yname = "",
  DLmin = 2, VLmin = 2, HLmin = 2, DLmax = length(Matrix::diag(RM))
  - 1, VLmax = length(Matrix::diag(RM)) - 1,
  HLmax = length(Matrix::diag(RM)) - 1, doShuffle = FALSE, y1 = NA,
  y2 = NA, Nshuffle = 19, AUTO = NULL, chromatic = FALSE,
  matrices = FALSE, doPlot = TRUE)
}
\arguments{
\item{RM}{A binary recurrence matrix}

\item{diagWin}{Window around the line of synchrony}

\item{xname}{Label for x-axis}

\item{yname}{Label for y-axis}

\item{DLmin}{DLmin}

\item{VLmin}{VLmin}

\item{HLmin}{HLmin}

\item{DLmax}{DLmax}

\item{VLmax}{VLmax}

\item{HLmax}{HLmax}

\item{doShuffle}{Should a shuffled baseline be calculated (default = \code{FALSE})}

\item{y1}{The original \code{y1} time series}

\item{y2}{The original \code{y2} time series}

\item{Nshuffle}{How many shuffled versions to make up the baseline? The default is \code{19}, which is the minimum for a one-sided surrogate test.}

\item{AUTO}{AUTO}

\item{chromatic}{chromatic}

\item{matrices}{matrices}

\item{doPlot}{Plot}
}
\value{
A plot and/or the data for the plot
}
\description{
Diagonal Recurrence Profile
}
