% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_mat}
\alias{crqa_mat}
\title{Get bootsrapped (C)RQA measures based on a recurrence matrix}
\usage{
crqa_mat(RM, threshold = NULL, DLmin = 2, VLmin = 2, HLmin = 2,
  DLmax = length(diag(RM)) - 1, VLmax = length(diag(RM)) - 1,
  HLmax = length(diag(RM)) - 1, AUTO = NULL, chromatic = FALSE,
  matrices = FALSE, doHalf = FALSE, Nboot = NULL, CL = 0.95)
}
\arguments{
\item{RM}{A distance matrix, or a matrix of zeroes and ones (you must set \code{radius = NULL})}

\item{DLmin}{Minimal diagonal line length (default = \code{2})}

\item{VLmin}{Minimal vertical line length (default = \code{2})}

\item{HLmin}{Minimal horizontal line length (default = \code{2})}

\item{DLmax}{Maximal diagonal line length (default = length of diagonal -1)}

\item{VLmax}{Maximal vertical line length (default = length of diagonal -1)}

\item{HLmax}{Maximal horizontal line length (default = length of diagonal -1)}

\item{AUTO}{Auto-recurrence? (default = \code{FALSE})}

\item{chromatic}{Force chromatic RQA? (default = \code{FALSE})}

\item{matrices}{Return matrices? (default = \code{FALSE})}

\item{doHalf}{Analyse half of the matrix? (default = \code{FALSE})}

\item{Nboot}{How many bootstrap replications? (default = \code{NULL})}

\item{CL}{Confidence limit for bootstrap results (default = \code{.95})}

\item{radius}{Threshold for distance value that counts as a recurrence}
}
\value{
A list object containing (C)RQA measures (and matrices if requested)
}
\description{
A zoo of measures based on singular recurrent points, diagonal, vertical and horizontal line structures will be caluclated.
}
\seealso{
Other Recurrence Quantification Analysis: \code{\link{crqa_cl}},
  \code{\link{crqa_mat_measures}},
  \code{\link{crqa_parameters}}, \code{\link{crqa_radius}}
}
