% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_parameters}
\alias{crqa_parameters}
\title{Find optimal (C)RQA parameters}
\usage{
crqa_parameters(y, maxDim = 10, maxLag = floor(length(y)/(maxDim + 1)),
  lagMethods = c("first.minimum", "global.minimum", "max.lag"),
  emLag = NULL, nnSizes = c(2, 5, 10, 15), nnRadius = 5,
  nnThres = 10, theiler = 0, estimateLags = TRUE,
  estimateDimensions = TRUE, diagPlot = TRUE, silent = TRUE, ...)
}
\arguments{
\item{y}{A numeric vector or time series}

\item{maxDim}{Maximum number of embedding dimensions (default = \code{10})}

\item{maxLag}{Maximum embedding lag to consider. Defaults to \code{floor(length(y)/(maxDim+1))}.}

\item{lagMethods}{A range of embedding lags to consider when calling \code{\link[nonlinearTseries]{timeLag}} with \code{technique="ami"}, valid options are c("first.e.decay", "first.zero", "first.minimum")}

\item{emLag}{Optimal embedding lag (delay), e.g., provided by optimising algorithm. Leave empty to estimate (default = \code{NULL})}

\item{nnSizes}{Points whose distance is \code{nnSize} times further apart than the estimated size of the attractor will be declared false neighbours. See the argument \code{atol} in \code{\link[fractal]{FNN}} (default = \code{c(2,5)})}

\item{nnRadius}{If the ratio of the distance between two points in successive dimensions is larger than \code{nnRadius}, the points are declared false neighbours. See the argument \code{rtol} in \code{\link[fractal]{FNN}} (default = \code{c(5,10)})}

\item{nnThres}{Threshold for selecting optimal parameter in percentage points (default = \code{10})}

\item{theiler}{Theiler window on distance matrix (default = \code{0})}

\item{estimateLags}{Get optimal embedding lags}

\item{estimateDimensions}{Get optimal embedding dimension}

\item{diagPlot}{Plot the results}

\item{silent}{Silent-ish mode}

\item{...}{Other parameters passed to \code{\link[nonlinearTseries]{timeLag}}}
}
\value{
A list object containing the optimal values and iteration history.
}
\description{
A wrapper for various algorithms used to find optimal embedding delay, number of embedding dimensions and radius.
}
\details{
A number of functions are called to determie optimal parameters for delay embedding a time series:

\itemize{
\item{Embedding lag (\eqn{\tau}, \code{emLag}): The default is to call \code{\link[casnet]{est_emLag}}, which is a wrapper around \code{\link[nonlinearTseries]{timeLag}} with \code{technique="ami"} to get lags based on the mutual information function.}
\item{Embedding dimension (\code{m}, \code{emDim}): The default is to call \code{\link[casnet]{est_emDim}}, which is a wrapper around \code{\link[fractal]{FNN}}}
}
}
\seealso{
Other Recurrence Quantification Analysis: \code{\link{crqa_cl}},
  \code{\link{crqa_radius}},
  \code{\link{crqa_rp_measures}}, \code{\link{crqa_rp}}
}
\concept{Recurrence Quantification Analysis}
