% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_parameters}
\alias{crqa_parameters}
\title{crqa_parameters}
\usage{
crqa_parameters(y, maxDim = 5, maxLag = round(length(y)/(maxDim + 1)),
  emLag = tau(y, selection.methods = c("first.e.decay", "first.zero",
  "first.minimum"), maxLag = maxLag), emLags = tau(y, maxLag = maxLag),
  ami.method = c("first.minimum"), nnSizes = c(0.1, 0.3, 0.6, 0.9),
  nnThres = 0.1, diagPlot = TRUE)
}
\arguments{
\item{y}{Timeseries}

\item{emLag}{Optimal embedding lag (delay), e.g., provided by optimising algorithm.}

\item{emLags}{A range of embedding lags to consider (default: )}

\item{maxDims}{Maximum number of embedding dimensions}

\item{emDims}{A range of embedding dimensions (defsult: 1 - maxDims)}

\item{nSize}{The neighbourhood size used to estimate the number of nearest neighbours of a coordinate (default: 10% of all points)}
}
\description{
A wrapper for various algorithms used to find optimal embedding delay, number of embedding dimensions and radius.
}
