% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_parameters}
\alias{crqa_parameters}
\title{Find optimal (C)RQA parameters}
\usage{
crqa_parameters(y, maxDim = 10, maxLag = floor(length(y)/(maxDim + 1)),
  lagMethods = "first.minimum", emLag = NULL, nnRadius = 10,
  nnSizes = c(0.005, 0.01, 0.005, 0.1), nnThres = 0.1, theiler = 0,
  diagPlot = TRUE, silent = TRUE, ...)
}
\arguments{
\item{y}{A numeric vector or time series}

\item{maxDim}{Maximum number of embedding dimensions (default = \code{min(c(10, length(y)),na.rm = TRUE)})}

\item{maxLag}{Maximum embedding lag. Defaults to \code{floor(length(y)/(maxDim+1))}.}

\item{lagMethods}{Optimal embedding lag (delay), e.g., provided by optimising algorithm.}

\item{nnSizes}{Neighbourhood size}

\item{nnThres}{Threshold}

\item{theiler}{Theiler window on distance matrix (default = \code{0})}

\item{diagPlot}{Plot the results}

\item{silent}{Silent-ish mode}

\item{...}{Other parameters passed to \code{\link[nonlinearTseries]{timeLag}}}

\item{lagMethods}{A range of embedding lags to consider when calling \code{\link[nonlinearTseries]{timemLag}} with \code{technique="ami"}, valid options are c("first.e.decay", "first.zero", "first.minimum")}
}
\value{
A list object containing the optimal values and iteration history.
}
\description{
A wrapper for various algorithms used to find optimal embedding delay, number of embedding dimensions and radius.
}
\details{
A number of functions are called to determie optimal parameters for delay embedding a time series:

\itemize{
\item{Embedding lag (\eqn{\tau}, \code{emLag}): The default is to call \code{\link[casnet]{est_emLag}}, which is a wrapper around \code{\link[nonlinearTseries]{timeLag}} with \code{technique="ami"} to get lags based on the mutual information function.}
\item{Embedding dimension (\code{m, \code{emLag}}): The default is to call \code{\link[casnet]{est_emDim}}, which is a wrapper around \code{\link[nonlinearTseries]{estimateEmbeddingDim}}}
}
}
\seealso{
Other Recurrence Quantification Analysis: \code{\link{crqa_cl}},
  \code{\link{crqa_mat_measures}}, \code{\link{crqa_mat}},
  \code{\link{crqa_radius}}
}
