% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_radius}
\alias{crqa_radius}
\title{Find a radius (fix RR)}
\usage{
crqa_radius(RM, startRadius = 0.1 * max(RM, na.rm = TRUE), targetRR = 0.05,
  tol = 0.1, maxIter = 100, AUTO = FALSE, theiler = 0,
  histIter = FALSE)
}
\arguments{
\item{RM}{RM}

\item{startRadius}{startRadius}

\item{targetRR}{targetRR}

\item{tol}{tol}

\item{maxIter}{maxIter}

\item{AUTO}{AUTO}

\item{theiler}{theiler}

\item{histIter}{histIter}
}
\description{
Find a radius (fix RR)
}
