% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{crqa_rp_measures}
\alias{crqa_rp_measures}
\title{Get (C)RQA measures from a Recurrence Matrix}
\usage{
crqa_rp_measures(RM, emRad = NULL, DLmin = 2, VLmin = 2, HLmin = 2,
  DLmax = length(Matrix::diag(RM)), VLmax = length(Matrix::diag(RM)),
  HLmax = length(Matrix::diag(RM)), AUTO = NULL, theiler = NULL,
  chromatic = FALSE, matrices = FALSE, doHalf = FALSE,
  Nboot = NULL, CL = 0.95, doParallel = FALSE)
}
\arguments{
\item{RM}{A distance matrix, or a matrix of zeroes and ones (you must set \code{emRad = NA})}

\item{emRad}{Threshold for distance value that counts as a recurrence}

\item{DLmin}{Minimal diagonal line length (default = \code{2})}

\item{VLmin}{Minimal vertical line length (default = \code{2})}

\item{HLmin}{Minimal horizontal line length (default = \code{2})}

\item{DLmax}{Maximal diagonal line length (default = length of diagonal -1)}

\item{VLmax}{Maximal vertical line length (default = length of diagonal -1)}

\item{HLmax}{Maximal horizontal line length (default = length of diagonal -1)}

\item{AUTO}{Auto-recurrence? (default = \code{FALSE})}

\item{theiler}{= Use a theiler window around the line of identity / synchronisation to remove high auto-correlation at short time-lags (default = \code{0})}

\item{chromatic}{Force chromatic RQA? (default = \code{FALSE})}

\item{matrices}{Return matrices? (default = \code{FALSE})}

\item{doHalf}{Analyse half of the matrix? (default = \code{FALSE})}

\item{Nboot}{How many bootstrap replications? (default = \code{NULL})}

\item{CL}{Confidence limit for bootstrap results (default = \code{.95})}

\item{doParallel}{Speed up calculations by using the parallel processing options provided by \code{parallel} to assign a seperate process/core for each window in windowed (C)RQA analysis using \code{\link[purrr:map2]{purrr::map2()}} to assign data and \code{\link[parallel:detectCores]{parallel::detectCores()}} with  \code{logical = TRUE} to decide on the available cores (default = \code{FALSE})}
}
\description{
Use \code{crqa_rp}
}
\seealso{
Other Recurrence Quantification Analysis: \code{\link{crqa_cl}},
  \code{\link{crqa_parameters}}, \code{\link{crqa_radius}},
  \code{\link{crqa_rp}}
}
\concept{Recurrence Quantification Analysis}
\keyword{internal}
