% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.R
\name{dc_ccp}
\alias{dc_ccp}
\title{Cumulative Complexity Peaks (CCP)}
\usage{
dc_ccp(
  df_win,
  alpha_item = 0.05,
  alpha_time = 0.05,
  doPlot = FALSE,
  useVarNames = TRUE,
  colOrder = TRUE,
  useTimeVector = NA,
  timeStamp = "01-01-1999"
)
}
\arguments{
\item{df_win}{A data frame containing series of Dynamic Complexity values obtained by running function \code{\link[=dc_win]{dc_win()}}}

\item{alpha_item}{The significance level of the one-sided Z-test used to determine which peaks are \verb{> 0}.}

\item{alpha_time}{The significance level of the one-sided Z-test used to determine if the number of significant peaks (as determined  by \code{alpha_item}) at a specific time stamp are \verb{> 0}.}

\item{doPlot}{If \code{TRUE} shows a Complexity Resonance Diagram of the Dynamic Complexity and returns an invisible \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object. (default = \code{FALSE})}

\item{useVarNames}{Use the column names of \code{df} as variable names in the Complexity Resonance Diagram (default = \code{TRUE})}

\item{colOrder}{If \code{TRUE}, the order of the columns in \code{df} determines the of variables on the y-axis. Use \code{FALSE} for alphabetic/numeric order. Use \code{NA} to sort by by mean value of Dynamic Complexity (default = \code{TRUE})}

\item{useTimeVector}{Parameter used for plotting. A vector of length \code{NROW(df)}, containing date/time information (default = \code{NA})}

\item{timeStamp}{If \code{useTimeVector} is not \code{NA}, a character string that can be passed to \code{\link[lubridate:stamp]{lubridate::stamp()}} to format the the dates/times passed in \code{useTimeVector} (default = \code{"01-01-1999"})}
}
\value{
A list with a dataframe of binary complexity peak indices and a cumulative complexity peak index, a CCP diagram.
}
\description{
Computes significant peaks in the dynamic complexity time series.
}
\references{
Haken H, & Schiepek G. (2006). \emph{Synergetik in der Psychologie. Selbstorganisation verstehen und gestalten}. Hogrefe, Göttingen.

Schiepek, G. (2003). A Dynamic Systems Approach to Clinical Case Formulation. \emph{European Journal of Psychological Assessment, 19}, 175-184. https://doi.org/10.1027//1015-5759.19.3.175

Schiepek, G., & Strunk, G. (2010). The identification of critical fluctuations and phase transitions in short term and coarse-grained time series-a method for the real-time monitoring of human change processes. \emph{Biological cybernetics, 102(3)}, 197-207. https://doi.org/10.1007/s00422-009-0362-1
}
\seealso{
Other Dynamic Complexity functions: 
\code{\link{dc_d}()},
\code{\link{dc_f}()},
\code{\link{dc_win}()},
\code{\link{plotDC_ccp}()},
\code{\link{plotDC_lvl}()},
\code{\link{plotDC_res}()}
}
\author{
Merlijn Olthof

Fred Hasselman
}
\concept{Dynamic Complexity functions}
