% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{dc_f}
\alias{dc_f}
\title{Fluctuation Intensity}
\usage{
dc_f(df, col_first, col_last, win = NROW(df), scale_min, scale_max)
}
\arguments{
\item{df}{A dataframe containing multivariate time series data from 1 person. Rows should indicate time,  columns should indicate variables. The multivariate time series should be on the same scale. If nescessary, rescale variables.}

\item{col_first}{The first column of the dataframe that should be analyzed.}

\item{col_last}{The last column of the dataframe that should be analyzed.}

\item{win}{Size of window in which to calculate Dynamic Complexity. If \code{win < NROW(df)} the window will move along the time series with a stepsize of \code{1} (default = \code{NROW(df)})}

\item{scale_min}{Theoretical minimum value of thescale.}

\item{scale_max}{Theoretical maximum value of the scale.}
}
\value{
dataframe
}
\description{
Fluctuation intensity is one of two components of which the product is the Dynamic Complexity measure.
}
\seealso{
Use \code{\link[=dc_win]{dc_win()}} to get the dynamic complexity measure.

Other Dynamic Complexity functions: \code{\link{dc_ccp}},
  \code{\link{dc_d}}, \code{\link{dc_win}},
  \code{\link{plotDC_ccp}}, \code{\link{plotDC_lvl}},
  \code{\link{plotDC_res}}
}
\concept{Dynamic Complexity functions}
