% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{dc_win}
\alias{dc_win}
\title{Dynamic Complexity}
\usage{
dc_win(df, col_first, col_last, win = NROW(df), scale_min, scale_max,
  doPlot = FALSE, useVarNames = TRUE, colOrder = TRUE,
  useTimeVector = NA, timeStamp = "01-01-1999")
}
\arguments{
\item{df}{A dataframe containing multivariate time series data from 1 person. Rows should indicate time,  columns should indicate variables. The multivariate time series should be on the same scale. If nescessary, rescale variables.}

\item{col_first}{The first column of the dataframe that should be analyzed.}

\item{col_last}{The last column of the dataframe that should be analyzed.}

\item{win}{Size of window in which to calculate Dynamic Complexity. If \code{win < NROW(df)} the window will move along the time series with a stepsize of \code{1} (default = \code{NROW(df)})}

\item{scale_min}{Theoretical minimum value of thescale.}

\item{scale_max}{Theoretical maximum value of the scale.}

\item{doPlot}{If \code{TRUE} shows a Complexity Resonance Diagram and returns an invisible \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object. If \code{FALSE}, plot will be drawn and no object will be returned (default = \code{FALSE})}

\item{useVarNames}{Use the column names of \code{df} as variable names in the Complexity Resonance Diagram (default = \code{TRUE})}

\item{colOrder}{If \code{TRUE}, the order of the columns in \code{df} determines the of variables on the y-axis. Use \code{FALSE} for alphabetic/numeric order. Use \code{NA} to sort by by mean value of Dynamic Complexity (default = \code{TRUE})}

\item{useTimeVector}{Parameter used for plotting. Either a variable name, or a vector of length \code{NROW(df)}, containing date/time information (default = \code{NA})}

\item{timeStamp}{If \code{useTimeVector} is not \code{NA}, a character string that can be passed to \code{\link[lubridate:stamp]{lubridate::stamp()}} to format the the dates/times passed in \code{useTimeVector} (default = \code{"01-01-1999"})}
}
\value{
If \code{doPlot = TRUE}, a list object containing a data frame of Dynamic Complexity values and a \code{ggplot2} object of the dynamic complexity resonance diagram (e.g. Schiepek et al., 2016). If \code{doPlot = FALSE} only the data frame with Dynamic Complexity series is returned.
}
\description{
Calculates Dynamic Complexity, a complexity index for short and coarse-grained time series (Schiepek & Strunk, 2010; Schiepek, 2003; Haken & Schiepek 2006).
}
\references{
Schiepek, G., & Strunk, G. (2010). The identification of critical fluctuations and phase transitions in short term and coarse-grained time series-a method for the real-time monitoring of human change processes. Biological cybernetics, 102(3), 197-207.

Schiepek, G. (2003). A Dynamic Systems Approach to Clinical Case Formulation. European Journal of Psychological Assessment, 19, 175-184.

Haken, H. & Schiepek, G. (2006, 2. Aufl. 2010). Synergetik in der Psychologie. Selbstorganisation verstehen und gestalten. G?ttingen: Hogrefe.

Schiepek, G. K., St?ger-Schmidinger, B., Aichhorn, W., Sch?ller, H., & Aas, B. (2016). Systemic case formulation, individualized process monitoring, and state dynamics in a case of dissociative identity disorder. Frontiers in psychology, 7, 1545.
}
\seealso{
Other Dynamic Complexity functions: \code{\link{dc_ccp}},
  \code{\link{dc_d}}, \code{\link{dc_f}},
  \code{\link{plotDC_ccp}}, \code{\link{plotDC_res}}
}
\author{
Merlijn Olthof
}
\concept{Dynamic Complexity functions}
