% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{di2bi}
\alias{di2bi}
\title{di2bi}
\usage{
di2bi(distmat, radius, convMat = FALSE)
}
\arguments{
\item{distmat}{Distance matrix}

\item{radius}{The radius or threshold value}

\item{convMat}{Should the matrix be converted to an object of type Matrix}
}
\value{
A matrix with only 0s and 1s
}
\description{
Distance matrix to binary matrix based on threshold value
}
