% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{di2bi}
\alias{di2bi}
\title{Distance to binary matrix

Distance matrix to binary matrix based on threshold value}
\usage{
di2bi(distmat, radius, convMat = FALSE)
}
\arguments{
\item{distmat}{Distance matrix}

\item{radius}{The radius or threshold value}

\item{convMat}{Should the matrix be converted to an object of type Matrix}
}
\value{
A matrix with only 0s and 1s
}
\seealso{
Other Distance Matrix operations: \code{\link{bandReplace}},
  \code{\link{di2we}}, \code{\link{dist_hamming}},
  \code{\link{lineDists}}, \code{\link{nzdiags}},
  \code{\link{recmat_plot}}, \code{\link{recmat_size}},
  \code{\link{recmat}}
}
