% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOps.R
\name{di2bi}
\alias{di2bi}
\title{Distance to binary matrix}
\usage{
di2bi(distmat, emRad = NA, convMat = FALSE)
}
\arguments{
\item{distmat}{Distance matrix}

\item{emRad}{The radius or threshold value}

\item{convMat}{Should the matrix be converted from a \code{distmat} object of class \code{\link[Matrix:Matrix]{Matrix::Matrix()}} to \code{\link[base:matrix]{base::matrix()}} (or vice versa)}
}
\value{
A (sparse) matrix with only 0s and 1s
}
\description{
Distance matrix to binary matrix based on threshold value
}
\seealso{
Other Distance matrix operations (recurrence plot): 
\code{\link{bandReplace}()},
\code{\link{di2ch}()},
\code{\link{di2we}()},
\code{\link{dist_hamming}()},
\code{\link{rp_lineDist}()},
\code{\link{rp_nzdiags}()},
\code{\link{rp_plot}()},
\code{\link{rp_size}()},
\code{\link{rp}()}

Other Distance matrix operations (recurrence network): 
\code{\link{di2ch}()},
\code{\link{di2we}()},
\code{\link{rn_phases}()},
\code{\link{rn_plot}()},
\code{\link{rn_recSpec}()},
\code{\link{rn}()}
}
\concept{Distance matrix operations (recurrence network)}
\concept{Distance matrix operations (recurrence plot)}
