% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{di2we}
\alias{di2we}
\title{Distance 2 weighted matrix}
\usage{
di2we(distmat, radius, convMat = FALSE)
}
\arguments{
\item{distmat}{Distance matrix}

\item{radius}{The radius or threshold value}

\item{convMat}{Should the matrix be converted to an object of type Matrix}
}
\value{
A matrix with 0s and leaves the values < threshold distance value
}
\description{
Distance matrix to weighted matrix based on threshold value
}
\seealso{
Other Distance matrix operations: \code{\link{bandReplace}},
  \code{\link{di2bi}}, \code{\link{dist_hamming}},
  \code{\link{lineDists}}, \code{\link{nzdiags}},
  \code{\link{recmat_plot}}, \code{\link{recmat_size}},
  \code{\link{recmat}}
}
