% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est_emDim}
\alias{est_emDim}
\title{Estimate number of embedding dimensions}
\usage{
est_emDim(
  y,
  delay = est_emLag(y),
  maxDim = 15,
  threshold = 0.95,
  max.relative.change = 0.1,
  doPlot = FALSE,
  ...
)
}
\arguments{
\item{y}{Time series or numeric vector}

\item{delay}{Embedding lag}

\item{maxDim}{Maximum number of embedding dimensions}

\item{threshold}{See \code{\link[nonlinearTseries:estimateEmbeddingDim]{nonlinearTseries::estimateEmbeddingDim()}}}

\item{max.relative.change}{See \code{\link[nonlinearTseries:estimateEmbeddingDim]{nonlinearTseries::estimateEmbeddingDim()}}}

\item{doPlot}{Plot}

\item{...}{Other arguments (not in use)}
}
\value{
Embedding dimensions
}
\description{
A wrapper for nonlinearTseries::estimateEmbeddingDim
}
\details{
A wrapper for \link[nonlinearTseries:estimateEmbeddingDim]{nonlinearTseries::estimateEmbeddingDim}
}
\seealso{
Other Estimate Recurrence Parameters: 
\code{\link{est_emLag}()},
\code{\link{est_parameters_roc}()},
\code{\link{est_parameters}()},
\code{\link{est_radius}()}
}
\concept{Estimate Recurrence Parameters}
