% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est_emLag}
\alias{est_emLag}
\title{Estimate embedding lag (tau)}
\usage{
est_emLag(y, selection.methods = "first.minimum", maxLag = length(y)/4, ...)
}
\arguments{
\item{y}{Time series or numeric vector}

\item{selection.methods}{Selecting an optimal embedding lag (default: Return "first.e.decay", "first.zero", "first.minimum", "first.value", where value is 1/exp(1))}

\item{maxLag}{Maximal lag to consider (default: 1/4 of timeseries length)}

\item{...}{Additional parameters}
}
\value{
The ami function with requested minima
}
\description{
A wrapper for \link[nonlinearTseries:timeLag]{nonlinearTseries::timeLag}
}
\seealso{
Other Estimate Recurrence Parameters: 
\code{\link{est_emDim}()},
\code{\link{est_parameters_roc}()},
\code{\link{est_parameters}()},
\code{\link{est_radius}()}
}
\concept{Estimate Recurrence Parameters}
