% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est_radius_rqa}
\alias{est_radius_rqa}
\title{Estimate Radius without building a recurrence matrix}
\usage{
est_radius_rqa(
  y1 = NULL,
  y2 = NULL,
  AUTO = NULL,
  method = "Euclidean",
  startRadius = NULL,
  targetValue = 0.05,
  tol = 0.01,
  maxIter = 100,
  theiler = NA,
  histIter = FALSE,
  standardise = c("mean.sd", "median.mad", "none")[3],
  radiusOnFail = c("tiny", "huge", "percentile")[1],
  silent = FALSE
)
}
\arguments{
\item{startRadius}{The starting value for the radius (default = SD of time series values)}

\item{targetValue}{When argument \code{type} is set to "fixed", the value represents the target value for the measure in \code{targetMeasure} (default = \code{RR = .05}).}

\item{tol}{Tolerance for achieving \code{targetValue} for \code{targetMeasure} (default = \code{0.01})}

\item{maxIter}{If \code{type = "fixed"}: Maximum number of iterations to reach targetValue.}

\item{theiler}{Size of theiler window (default \code{0})}

\item{histIter}{Return iteration history? (default = \code{FALSE})}

\item{standardise}{Standardise \code{y} if \code{type == "optimal"}}

\item{radiusOnFail}{Radius to return when search fails \verb{"tiny" = 0 + ,Machine.double.eps}, this will likely cause a matrix full of zeros. \verb{"huge" = 1 + max. distance}, which will give a matrix full of ones, \verb{"minimum" = minimum distance in matrix}.}

\item{silent}{Silent-ish}

\item{noiseLevel}{Noise level to construct the \verb{signal + noiseLevel *} \eqn{N(\mu=0,\sigma=1)} (default = \code{0.75})}

\item{noiseType}{Type}

\item{plotROC}{Generates an ROC plot if \code{type = "optimal"}}
}
\value{
A dataframe listing settings used to search for the radius, the radius found given the settings and the recurrence rate produced by the radius (either 1 row or the entire iteration history)
}
\description{
Find a fixed or optimal radius.
}
\seealso{
Other Estimate Recurrence Parameters: 
\code{\link{est_emDim}()},
\code{\link{est_emLag}()},
\code{\link{est_parameters_roc}()},
\code{\link{est_parameters}()},
\code{\link{est_radius}()}
}
\concept{Estimate Recurrence Parameters}
