% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{fd_psd}
\alias{fd_psd}
\title{Power Spectral Density Slope (PSD).}
\usage{
fd_psd(
  y,
  fs = NULL,
  standardise = TRUE,
  detrend = TRUE,
  fitMethod = c("lowest25", "Wijnants", "Hurvich-Deo")[3],
  doPlot = FALSE,
  returnPlot = FALSE,
  returnPLAW = FALSE,
  returnInfo = FALSE,
  silent = FALSE,
  noTitle = FALSE,
  tsName = "y"
)
}
\arguments{
\item{y}{A numeric vector or time series object.}

\item{fs}{Sample rate (default = \code{NULL})}

\item{standardise}{standardise the series (default = \code{TRUE}).}

\item{detrend}{Subtract linear trend from the series (default = \code{TRUE}).}

\item{fitMethod}{Method to decide on a frequency range for log-log fit. Can be one of: "lowest25","Wijnants","Hurvich-Deo" (default). See details for more info.}

\item{doPlot}{Return the log-log spectrum with linear fit (default = \code{TRUE}).}

\item{returnPlot}{Return ggplot2 object (default = \code{FALSE})}

\item{returnPLAW}{Return the power law data (default = \code{FALSE})}

\item{returnInfo}{Return all the data used in SDA (default = \code{FALSE})}

\item{silent}{Run in silent-ish mode (default = \verb{TRUE)}}

\item{noTitle}{Do not generate a title (only the subtitle)}

\item{tsName}{Name of y added as a subtitle to the plot}
}
\value{
A list object containing:
\itemize{
\item A data matrix \code{PLAW} with columns \code{freq.norm}, \code{size} and \code{bulk}.
\item Estimate of scaling exponent \code{alpha} based on a fit over the lowest 25\\% frequencies (\code{low25}), or using the HD estimate \code{HD}.
\item Estimate of the the Fractal Dimension (\code{FD}) using conversion formula's reported in Hasselman(2013).
\item Information output by various functions.
}
}
\description{
Estimate Alpha, Hurst Exponent and Fractal Dimension through log-log slope.
}
\details{
Calls function \code{\link[stats:spec.pgram]{stats::spec.pgram()}} to estimate the scaling exponent of a timeseries based on the periodogram frequency spectrum. After detrending and normalizing the signal (if requested), \code{\link[stats:spec.pgram]{stats::spec.pgram()}} is called using a cosine taper = 0.5.

A line is fitted on the periodogram in log-log coordinates. The full range is fitted as well as one of three fit-ranges:
\itemize{
\item \code{lowest25} - The 25\\% lowest frequencies
\item \code{Wijnants} - The 50 lowest frequencies (Wijnants et al., 2012)
\item \code{Hurvich-Deo} - The Hurvich-Deo estimate (Hurvich & Deo, 1999)
}
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}

Hurvich, C.M., & Deo, R.R. (1999). Plug-in Selection of the Number of Frequencies in Regression Estimates of the Memory Parameter of a Long Memory Time Series. \emph{Journal of Time Series Analysis, 20(3)}, 331–341.
}
\seealso{
Other Fluctuation Analyses: 
\code{\link{fd_RR}()},
\code{\link{fd_allan}()},
\code{\link{fd_dfa}()},
\code{\link{fd_mfdfa}()},
\code{\link{fd_sda}()},
\code{\link{fd_sev}()}
}
\author{
Fred Hasselman
}
\concept{Fluctuation Analyses}
