% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{fd_psd}
\alias{fd_psd}
\title{Power Spectral Density Slope (PSD).}
\usage{
fd_psd(y, fs = NULL, standardise = TRUE, detrend = TRUE,
  doPlot = TRUE, returnPLAW = FALSE, returnInfo = FALSE,
  silent = TRUE)
}
\arguments{
\item{y}{A numeric vector or time series object.}

\item{fs}{Sample rate (default = \code{NULL})}

\item{standardise}{standardise the series (default = \code{TRUE}).}

\item{detrend}{Subtract linear trend from the series (default = \code{TRUE}).}

\item{doPlot}{Return the log-log spectrum with linear fit (default = \code{TRUE}).}

\item{returnPLAW}{Return the power law data (default = \code{FALSE})}

\item{returnInfo}{Return all the data used in DFA (default = \code{FALSE})}

\item{silent}{Run in silent-ish mode (default = \code{TRUE)})}
}
\value{
A list object containing:
\itemize{
\item A data matrix \code{PLAW} with columns \code{freq.norm}, \code{size} and \code{bulk}.
\item Estimate of scaling exponent \code{alpha} based on a fit over the lowest 25\% frequencies (\code{low25}), or using the HD estimate \code{HD}.
\item Estimate of the the Fractal Dimension (\code{FD}) using conversion formula's reported in Hasselman(2013).
\item Information output by various functions.
}
}
\description{
Estimate Alpha, Hurst Exponent and Fractal Dimension through log-log slope.
}
\details{
Calls function \code{\link[sapa]{SDF}} to estimate the scaling exponent of a timeseries based on the periodogram frequency spectrum. After detrending and normalizing the signal (if requested), \code{SDF} is called using a Tukey window (\code{raised cosine \link[sapa]{taper}}).

A line is fitted on the periodogram in log-log coordinates. Two fit-ranges are used: The 25\% lowest frequencies and the Hurvich-Deo estimate (\code{\link[fractal]{HDEst}}).
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}
}
\seealso{
Other Fluctuation Analyses: \code{\link{MFDFA}},
  \code{\link{fd_RR}}, \code{\link{fd_dfa}}
}
\author{
Fred Hasselman
}
\concept{Fluctuation Analyses}
