% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{fd_sev}
\alias{fd_sev}
\title{Calculate FD using Sevcik's method}
\usage{
fd_sev(
  y,
  detrend = FALSE,
  adjustSumOrder = FALSE,
  smallNapprox = FALSE,
  doPlot = FALSE,
  returnPlot = FALSE,
  returnPLAW = FALSE,
  returnInfo = FALSE,
  silent = FALSE,
  noTitle = FALSE,
  tsName = "y"
)
}
\arguments{
\item{y}{A time series or numeric vector}

\item{detrend}{Subtract linear trend from the series (default = \code{TRUE}).}

\item{adjustSumOrder}{Adjust the time series (summation or differencing), based on the global scaling exponent, see e.g. \url{https://www.frontiersin.org/files/Articles/23948/fphys-03-00141-r2/image_m/fphys-03-00141-t001.jpg}{Ihlen (2012)} (default = \code{TRUE})}

\item{smallNapprox}{Force use of small sample approximation (default for N < 128)}

\item{doPlot}{Return the log-log scale versus fluctuation plot with linear fit (default = \code{TRUE}).}

\item{returnPlot}{Return ggplot2 object (default = \code{FALSE})}

\item{returnPLAW}{Return the power law data (default = \code{FALSE})}

\item{returnInfo}{Return all the data used in DFA (default = \code{FALSE})}

\item{silent}{Silent-ish mode}

\item{noTitle}{Do not generate a title (only the subtitle)}

\item{tsName}{Name of y added as a subtitle to the plot}
}
\value{
An FD estimate
}
\description{
Calculate FD using Sevcik's method
}
\references{
Sevcik, C. (1998). A procedure to Estimate the Fractal Dimension of Waveforms. Paper available at http://arxiv.org/pdf/1003.5266.pdf
}
\seealso{
Other Fluctuation Analyses: 
\code{\link{fd_RR}()},
\code{\link{fd_allan}()},
\code{\link{fd_dfa}()},
\code{\link{fd_mfdfa}()},
\code{\link{fd_psd}()},
\code{\link{fd_sda}()}
}
\author{
Fred Hasselman
}
\concept{Fluctuation Analyses}
