% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R, R/nlRtsa_SOURCE.old.R
\name{gg.plotHolder}
\alias{gg.plotHolder}
\alias{gg.plotHolder}
\title{gg.plotHolder}
\usage{
gg.plotHolder(useArial = F, afmPATH = "~/Dropbox")

gg.plotHolder(useArial = F, afmPATH = "~/Dropbox")
}
\arguments{
\item{useArial}{Use the Arial font (requires \code{.afm} font files in the \code{afmPath})}

\item{afmPATH}{Path to Arial \code{.afm} font files.}

\item{useArial}{Use the Arial font (requires \code{.afm} font files in the \code{afmPath})}

\item{afmPATH}{Path to Arial \code{.afm} font files.}
}
\value{
A blank \code{ggplot2} object that can be used in concordance with \code{grid.arrange}.

A blank \code{ggplot2} object that can be used in concordance with \code{grid.arrange}.
}
\description{
gg.plotHolder

gg.plotHolder
}
\examples{
# Create a plot with marginal distributions.
library(ggplot2)
library(scales)

df <- data.frame(x = rnorm(n = 100), y = rnorm(n = 100), group = factor(sample(x=c(0,1), size = 100, replace = TRUE)))

scatterP <- ggplot(df, aes(x = x, y =y, colour = group)) + geom_point() + gg.theme()
xDense <- ggplot(df, aes(x = x, fill = group)) + geom_density(aes(y= ..count..),trim=FALSE, alpha=.5) + gg.theme("noax") + theme(legend.position = "none")
yDense <- ggplot(df, aes(x = y, fill = group)) + geom_density(aes(y= ..count..),trim=FALSE, alpha=.5) + coord_flip() + gg.theme("noax") + theme(legend.position = "none")

library(gridExtra)
grid.arrange(xDense, gg.plotHolder(), scatterP, yDense, ncol=2, nrow=2, widths=c(4, 1.4), heights=c(1.4, 4))
# Create a plot with marginal distributions.
library(ggplot2)
library(scales)

df <- data.frame(x = rnorm(n = 100), y = rnorm(n = 100), group = factor(sample(x=c(0,1), size = 100, replace = TRUE)))

scatterP <- ggplot(df, aes(x = x, y =y, colour = group)) + geom_point() + gg.theme()
xDense <- ggplot(df, aes(x = x, fill = group)) + geom_density(aes(y= ..count..),trim=FALSE, alpha=.5) + gg.theme("noax") + theme(legend.position = "none")
yDense <- ggplot(df, aes(x = y, fill = group)) + geom_density(aes(y= ..count..),trim=FALSE, alpha=.5) + coord_flip() + gg.theme("noax") + theme(legend.position = "none")

library(gridExtra)
grid.arrange(xDense, gg.plotHolder(), scatterP, yDense, ncol=2, nrow=2, widths=c(4, 1.4), heights=c(1.4, 4))
}
