% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{HurvichDeo}
\alias{HurvichDeo}
\title{HurvichDeo}
\usage{
HurvichDeo(nr, spec, A = 0.3, delta = 6/7)
}
\arguments{
\item{nr}{NFT}

\item{spec}{sdf}

\item{A}{A}

\item{delta}{delta}
}
\value{
Hurvich-Deo estimate
}
\description{
Adaptation from discontinued package \code{fractal} by William Constantine and Donald Percival.
}
\keyword{internal}
