% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{inf_MSE}
\alias{inf_MSE}
\title{Multi-Scale Entropy}
\usage{
inf_MSE(
  y,
  summaryFunction = c("mean", "median", "min", "max")[1],
  retainLength = FALSE,
  m = 2,
  r = 0.2,
  D = NA,
  fs = NULL,
  transformBefore = TRUE,
  standardise = c("none", "mean.sd", "median.mad")[1],
  detrend = FALSE,
  polyOrder = 1,
  adjustSumOrder = FALSE,
  scaleMin = 1,
  scaleMax = floor(NROW(y)/10),
  scaleDiff = 1,
  scaleS = NA,
  overlap = 0,
  minData = 4,
  relativeEntropy = FALSE,
  doPlot = FALSE,
  returnPlot = FALSE,
  returnPLAW = FALSE,
  returnInfo = FALSE,
  silent = FALSE,
  noTitle = FALSE,
  tsName = "y"
)
}
\arguments{
\item{y}{A numeric vector}

\item{summaryFunction}{How should the data be summarized in the bins?}

\item{retainLength}{Return only the bin values (\code{FALSE}), or retain the length of the original series? (default = \code{TRUE})}

\item{m}{The size of the window in which tho evaluate whether a pattern repeats (default = \code{2})}

\item{r}{A factor that will determine the threshold for similarity of values, calculated as r x D (default = \code{0.2})}

\item{D}{Commonly the standard deviation of the time series, the similarity threshold will be calculated as r x D. Note that if the series is detrended and/or standardised and \code{D = NA} the standard deviation will be calculated after the transformations (default = \code{NA})}

\item{fs}{Sample rate}

\item{transformBefore}{Detrend/standardise before coarse graining. If set to \code{FALSE}, each coarsegrained series will be detrended/standardised separately (default = \code{TRUE})}

\item{standardise}{Standardise by the series using \code{\link[=ts_standardise]{ts_standardise()}} with \code{adjustN = FALSE} (default = "mean.sd")}

\item{polyOrder}{Order of polynomial trend to remove if \code{removeTrend = "poly"}}

\item{scaleDiff}{Steps between scaleMin and scaleMax (default = \code{1})}

\item{returnInfo}{Return all the data used in DFA (default = \code{FALSE})}

\item{silent}{Silent-ish mode}
}
\value{
MSE
}
\description{
Calculate the Multi-Scale (Sample) Entropy of a time series.
}
\note{
The transformation settings (detrending and/or normalisation) and \code{transformBefore} will determine the value of the product r x D if \code{D = NA}.
}
\examples{

y <- rnorm(100)

out <- inf_MSE(y, r=.5, scaleMin=1, scaleMax=10, returnInfo = TRUE, doPlot = FALSE)


}
\seealso{
Other Information based complexity measures: 
\code{\link{inf_SampEn}()}
}
\concept{Information based complexity measures}
