% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{inf_MSE}
\alias{inf_MSE}
\title{Multi-Scale Entropy}
\usage{
inf_MSE(
  y,
  summaryFunction = c("mean", "median", "min", "max")[1],
  retainLength = FALSE,
  m = 2,
  r = 0.2,
  D = NA,
  fs = NULL,
  transformBefore = TRUE,
  removeTrend = c("no", "poly", "adaptive", "bridge")[1],
  polyOrder = 1,
  standardise = c("none", "mean.sd", "median.mad")[1],
  adjustSumOrder = FALSE,
  scaleMin = 1,
  scaleMax = floor(NROW(y)/10),
  scaleS = NA,
  overlap = 0,
  dataMin = 4,
  relativeEntropy = FALSE,
  doPlot = FALSE,
  returnPlot = FALSE,
  returnPLAW = FALSE,
  returnInfo = FALSE,
  silent = FALSE,
  noTitle = FALSE,
  tsName = "y"
)
}
\arguments{
\item{y}{A numeric vector}

\item{summaryFunction}{How should the data be summarized in the bins? Can be \code{"mean"}, \code{"median"}, \code{"min"}, \code{"max"}, or, \code{"maxFreq"}. Value \code{"maxFreq"} is for categorical data and will pick the most frequently occurring category within the bin (default = \code{"mean"})}

\item{retainLength}{Return only the bin values (\code{FALSE}), or retain the length of the original series? (default = \code{TRUE})}

\item{m}{The size of the window in which tho evaluate whether a pattern repeats (default = \code{2})}

\item{r}{A factor that will determine the threshold for similarity of values, calculated as r x D (default = \code{0.2})}

\item{D}{Commonly the standard deviation of the time series, the similarity threshold will be calculated as r x D. Note that if the series is detrended and/or standardised and \code{D = NA} the standard deviation will be calculated after the transformations (default = \code{NA})}

\item{fs}{Sample rate}

\item{transformBefore}{Detrend/standardise before coarse graining. If set to \code{FALSE}, each coarsegrained series will be detrended/standardised separately (default = \code{TRUE})}

\item{removeTrend}{Method to use for global detrending (default = \code{"poly"})}

\item{polyOrder}{Order of global polynomial trend to remove if \code{removeTrend = "poly"}. If \code{removeTrend = "adaptive"} polynomials \code{1} to \code{polyOrder} will be evaluated and the best fitting curve (R squared) will be removed (default = \code{1})}

\item{standardise}{Standardise the series using \code{\link[=ts_standardise]{ts_standardise()}} with \code{adjustN = FALSE} (default = "mean.sd")}

\item{adjustSumOrder}{Adjust the time series (summation or difference), based on the global scaling exponent, see e.g. \href{https://www.frontiersin.org/files/Articles/23948/fphys-03-00141-r2/image_m/fphys-03-00141-t001.jpg}{Ihlen (2012)} (default = \code{FALSE})}

\item{scaleMin}{Minimum scale (in data points) to use for log-log regression (default = \code{4})}

\item{scaleMax}{Maximum scale (in data points) to use for log-log regression. This value will be ignored if \code{dataMin} is not \code{NA}, in which case bins of size \verb{< dataMin} will be removed (default = \code{stats::nextn(floor(NROW(y)/4), factors = 2)})}

\item{scaleS}{If not \code{NA}, it should be a numeric vector listing the scales on which to evaluate the detrended fluctuations. Arguments \verb{scaleMax, scaleMin, scaleResolution} and \code{dataMin} will be ignored (default = \code{NA})}

\item{overlap}{A number in \verb{[0 ... 1]} representing the amount of 'bin overlap' when calculating the fluctuation. This reduces impact of arbitrary time series begin and end points. If \code{length(y) = 1024} and overlap is \code{.5}, a scale of \code{4} will be considered a sliding window of size \code{4} with step-size \code{floor(.5 * 4) = 2}, so for scale \code{128} step-size will be \code{64} (default = \code{NA})}

\item{relativeEntropy}{The relative entropy, SampEn / (-1 * log(1/length(y))) will be returned (default = \code{FALSE})}

\item{doPlot}{Output the log-log scale versus fluctuation plot with linear fit by calling function \code{plotFD_loglog()} (default = \code{TRUE})}

\item{returnPlot}{Return ggplot2 object (default = \code{FALSE})}

\item{returnPLAW}{Return the power law data (default = \code{FALSE})}

\item{returnInfo}{Return all the data used in SDA (default = \code{FALSE})}

\item{silent}{Silent-ish mode (default = \code{FALSE})}

\item{noTitle}{Do not generate a title (only the subtitle) (default = \code{FALSE})}

\item{tsName}{Name of y added as a subtitle to the plot (default = \code{"y"})}
}
\value{
MSE
}
\description{
Calculate the Multi-Scale (Sample) Entropy of a time series.
}
\note{
The transformation settings (detrending and/or normalisation), \code{transformBefore} will determine the value of the product r x D if \code{D = NA}.
}
\examples{

y <- rnorm(100)

out <- inf_MSE(y, r=.5, scaleMin=1, scaleMax=10, returnInfo = TRUE, doPlot = FALSE)


}
\seealso{
Other Information based complexity measures: 
\code{\link{inf_SampEn}()}
}
\concept{Information based complexity measures}
