% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyAnalystsESM.R
\docType{data}
\name{manyAnalystsESM}
\alias{manyAnalystsESM}
\title{Data from the Many Analysts study.}
\description{
Data from the Many Analysts study.
}
\references{
Bastiaansen et al. (2019). Time to get personal? The impact of researchers’ choices on the selection of treatment targets using the experience sampling methodology. \href{https://doi.org/10.31234/osf.io/c8vp7}{PsyArXiv}

Many Analysts ESM \href{https://osf.io/h3djy/}{Data set}
}
\author{
Bastiaansen et al. (2019). Corresponding author: L.F. Bringmann \email{l.f.bringmann@rug.nl}
}
\keyword{data}
