% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrn.R
\name{mi_mat}
\alias{mi_mat}
\title{Mutual Information variations}
\usage{
mi_mat(y, ID1, ID2, discreteBins = ceiling(2 * NROW(ID1)^(1/3)))
}
\arguments{
\item{y}{A matrix with time series in columns}

\item{ID1}{ids}

\item{ID2}{ids}

\item{discreteBins}{Number of bins to use when discretizing the time series}
}
\value{
mi in nats
}
\description{
Mutual Information variations
}
\seealso{
Other Redundancy measures (mutual information): 
\code{\link{mi_interlayer}()},
\code{\link{mif}()}
}
\concept{Redundancy measures (mutual information)}
