% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{monoH}
\alias{monoH}
\title{mono Hurst}
\usage{
monoH(
  TSm,
  scaleS,
  polyOrder = 1,
  returnPLAW = FALSE,
  returnSegments = FALSE,
  removeRMSbelow = .Machine$double.eps
)
}
\arguments{
\item{TSm}{TS matrix with 2 columns \code{t} (1st) and \code{y} (second)}

\item{scaleS}{scales to evaluate}

\item{polyOrder}{If numeric: order to use for polynomial detrendiing, if "adaptive" will use the best fitting polynomial.}
}
\description{
mono Hurst
}
\keyword{internal}
