% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{normalise}
\alias{normalise}
\title{Normalise a vector by Mean/SD, or Median/MAD}
\usage{
normalise(numvec, na.rm = TRUE, type = c("mean", "median")[1])
}
\arguments{
\item{numvec}{A numeric vector}

\item{na.rm}{Set the \code{na.rm} field}

\item{type}{Center on the \code{"mean"} and divide by \code{sd} (default), or center on \code{"median"} and divide by \code{mad}}
}
\value{
A normalised vector
}
\description{
Normalise a vector by Mean/SD, or Median/MAD
}
\author{
Fred Hasselman
}
