% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{nzdiags}
\alias{nzdiags}
\title{nzdiags}
\usage{
nzdiags(A = NULL, d = NULL)
}
\arguments{
\item{A}{A binary (0,1) matrix.}

\item{d}{An optional vecotor of diagonals to extract.}
}
\description{
Get all nonzero diagonals of a binary matrix, or, diagonals specified as a vector by argument \code{d}. Output mimics Matlab's \code{[B,d] = spdiags(A)} function.
}
\author{
Fred Hasselman
}
