% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotMRN_win}
\alias{plotMRN_win}
\title{Plot windowed Multiplex Recurrence Network measures}
\usage{
plotMRN_win(
  df_mrn,
  measures = "mi",
  mrnWeights = "mi",
  showSeries = FALSE,
  doPlot = TRUE
)
}
\arguments{
\item{df_mrn}{Output from function \code{\link[=mrn]{mrn()}} with arguments set for a windowed analysis.}

\item{measures}{Character vector indicating which measures should be plotted. Valid elements in the vector are: \code{"mi"} for inter-layer mutual information,\code{"eo"} for edge overlap, or any function name from package \link{igraph} that can be applied to extract vertex-based measures from a weighted network. This is the multiplex recurrence network with a number of nodes equal to the number of layers and edge weights set by argument \code{mrnWeights}. (default = \code{"mi"})}

\item{mrnWeights}{Which measure should be used for the MRN edge-weights? Valid options are \code{"mi"} for inter-layer mutual information,\code{"eo"} for edge overlap. (default = \code{"mi"})}

\item{showSeries}{Should the time series that constitute the layers be plotted below the windowed MRN measures? This will only work if the \code{\link[=mrn]{mrn()}} was called using graphs that were generated by function \code{\link[=rn]{rn()}} with \code{returnGraph = TRUE}, in which case they will have a property \code{V(g)$y1} (and if applicable \code{V(g)$y2}), which will be plotted. If more than 1 measure is requested in \code{measures} and \code{showSeries = TRUE} a seperate plot for each measure will be produced. (default = \code{FALSE})}

\item{doPlot}{Plot the igraph object.}
}
\value{
a ggplot object
}
\description{
Plot windowed Multiplex Recurrence Network measures
}
\seealso{
Other Tools for windowed analyses: 
\code{\link{ts_windower}()}
}
\concept{Tools for plotting multiplex recurrence networks}
\concept{Tools for windowed analyses}
