% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rn.R
\name{rn}
\alias{rn}
\title{Create a Recurrence Network Matrix}
\usage{
rn(
  y1,
  y2 = NULL,
  emDim = 1,
  emLag = 1,
  emRad = NULL,
  theiler = 0,
  directed = FALSE,
  cumulative = TRUE,
  weighted = FALSE,
  weightedBy = c("si", "rt", "rf")[1],
  rescaleWeights = FALSE,
  fs = NA,
  to.ts = NULL,
  order.by = NULL,
  to.sparse = FALSE,
  method = "Euclidean",
  targetValue = 0.05,
  returnGraph = FALSE,
  doPlot = FALSE,
  doEmbed = TRUE,
  silent = TRUE,
  ...
)
}
\arguments{
\item{y1}{A numeric vector or time series}

\item{y2}{A numeric vector or time series for cross recurrence}

\item{emDim}{The embedding dimensions}

\item{emLag}{The embedding lag}

\item{emRad}{The threshold (emRad) to apply to the distance matrix to create a binary or weighted matrix. If \code{NULL}, an unthresholded matrix will be created (default = \code{NULL})}

\item{theiler}{Use a \code{theiler} window around the main diagonal (Line of Identity/Synchronisation) to remove auto-correlations at short time-lags:
\itemize{
\item \code{0} will include the main diagonal in all RQA measure calculations.
\item \code{1} will remove the main diagonal from all RQA measure calculations.
\item \code{NA} (default), will check if the matrix is symmetrical , if so, it will remove the diagonal by setting \code{theiler = 0} (Line of Identity, Auto-RQA), if it is not symmetrical (Line of Synchronisation, Cross-RQA) it will set \code{theiler = 1}.
\item A value greater than \code{1} will remove that many diagonals around and including the diagonal from all RQA measure calculations. So \code{theiler = 2} means exclude \code{2} diagonals around the main diagonal, including the main diagonal itself: \verb{[-1,0,1]}.
If \code{theiler} is a numeric vector of \code{length(theiler) == 2} it is possible to exclude an asymmetrical window. The values are interpreted as end points in a sequence of diagonal ID's, e.g. \code{theiler = c(-1,5)} will exclude \verb{[-1,0,1,2,3,4,5]}. If \code{length(theiler) > 2}, the values will be considered individual diagonal ID's, e.g. \code{theiler = c(-3,-1,0,2,5)}, will exclude only those specific ID's. Also see the note.
}}

\item{directed}{Should the matrix be considered to represent a directed network? (default = \code{FALSE})}

\item{cumulative}{To make the network represent cumulative time, set \code{directed = TRUE} and \code{cumulative = TRUE}. This will set the upper triangle of the recurrence matrix to \code{0} and ensures that the network edges represent recurrent values that have occurred in the \code{past} relative to the current observed value (node). If \code{directed = FALSE} the argument is ignored (default = \code{TRUE}).}

\item{weighted}{Should the matrix be considered to represent a weighted network? (default = \code{FALSE})}

\item{weightedBy}{After setting values smaller than \code{emRad} to \code{0}, what should the recurrent values represent? The default is to use the state space similarity (distance/proximity) values as weights (\code{"si"}). Other option are \code{"rt"} for \emph{recurrence time} and \code{"rf"} for \emph{recurrence time frequency}, Because vertices represent time points in \eqn{\epsilon}-thresholded recurrence networks, a difference of two vertex-indices represents duration. If an edge \code{e1} connects \code{v1} and \code{v10} then the \emph{recurrence time} will be the difference of the vertex indices, \code{9}, and the \emph{recurrence time frequency} will be \code{1/9}.}

\item{rescaleWeights}{If set to \code{TRUE} and \code{weighted = TRUE}, all weight values will be rescaled to \verb{[0,1]}, where \code{0} means no recurrence relation and \code{1} the maximum weight value.}

\item{fs}{Sample frequency: A numeric value interpreted as the \verb{number of observed samples per unit of time}. If the weights represent recurrence times (\code{"rt"}), they will be divided by the value in \code{fs}. If the weights represent recurrence time frequencies (\code{"rf"}), they will be multiplied by the value of \code{fs} (default = \code{NA})}

\item{to.ts}{Should \code{y1} and \code{y2} be converted to time series objects?}

\item{order.by}{If \code{to.ts = TRUE}, pass a vector of the same length as \code{y1} and \code{y2}. It will be used as the time index, if \code{NA} the vector indices will be used to represent time.}

\item{to.sparse}{Should sparse matrices be used?}

\item{method}{Distance measure to use. Any option that is valid for argument \code{method} of \code{\link[proxy:dist]{proxy::dist()}}. Type \code{proxy::pr_DB$get_entries()} to see a list of all the options. Common methods are: "Euclidean", "Manhattan", "Minkowski", "Chebysev" (or the same but shorter: "L2","L1","Lp" and "max" distance) (default = \code{"Euclidean"})}

\item{targetValue}{A value passed to \code{est_radius(...,type="fixed", targetMeasure="RR")} if \code{is.na(emRad)==TRUE}.}

\item{returnGraph}{Return an \code{\link[igraph:aaa-igraph-package]{igraph::igraph()}} object (default = \code{FALSE})}

\item{doPlot}{Plot the matrix by calling \code{\link[=rp_plot]{rp_plot()}} with default settings}

\item{doEmbed}{If \code{FALSE}, a distance matrix will be returned that is not embedded by \code{emDim} and \code{emLag} (Multidimensional RQA). If \code{y1} and/or \code{y2} are data frames, the columns will be used as the state space dimensions (default = \code{TRUE})}

\item{silent}{Silent-ish mode}

\item{...}{Any paramters to pass to \code{\link[=rn_plot]{rn_plot()}} if \code{doPlot = TRUE}}
}
\value{
A (Coss-) Recurrence matrix that can be interpreted as an adjacency (or incidence) matrix.
}
\description{
This function serves as a wrapper for function \code{rp()}, it will add some attributes to the matrix related to network representation. These attributes will be used to decide which network type to generate (e.g. undirected, directed, weighted, etc.)
}
\seealso{
Other Distance matrix operations (recurrence network): 
\code{\link{di2bi}()},
\code{\link{di2ch}()},
\code{\link{di2we}()},
\code{\link{rn_phases}()},
\code{\link{rn_plot}()},
\code{\link{rn_recSpec}()}
}
\concept{Distance matrix operations (recurrence network)}
