% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rn.R
\name{rn_plot}
\alias{rn_plot}
\title{Plot (thresholded) distance matrix as a network}
\usage{
rn_plot(
  RN,
  plotDimensions = FALSE,
  plotMeasures = FALSE,
  drawGrid = FALSE,
  markEpochsLOI = NULL,
  radiusValue = NA,
  title = "",
  xlabel = "",
  ylabel = "",
  plotSurrogate = NA,
  returnOnlyObject = FALSE
)
}
\arguments{
\item{RN}{A distance matrix or recurrence matrix}

\item{plotDimensions}{Should the state vectors be plotted if they are available as attributes of RM (default = \code{TRUE})}

\item{plotMeasures}{Print common (C)RQA measures in the plot if the matrix is binary (default = \code{FALSE})}

\item{drawGrid}{Draw a grid on the recurrence plot (default = \code{FALSE})}

\item{markEpochsLOI}{Pass a factor whose levels indicate different epochs or phases in the time series and use the line of identity to represent the levels by different colours (default = \code{NULL})}

\item{radiusValue}{If \code{plotMeasures = TRUE} and RM is an unthresholded matrix, this value will be used to calculate recurrence measures. If \code{plotMeasures = TRUE} and RM is already a binary recurrence matrix, pass the radius that was used as a threshold to create the matrix for display purposes. If \code{plotMeasures = TRUE} and \code{radiusValue = NA}, function \code{est_radius()} will be called with default settings (find a radius that yields .05 recurrence rate). If \code{plotMeasures = FALSE} this setting will be ignored.}

\item{title}{A title for the plot}

\item{xlabel}{An x-axis label}

\item{ylabel}{An y-axis label}

\item{plotSurrogate}{Should a 2-panel comparison plot based on surrogate time series be added? If \code{RM} has attributes \code{y1} and \code{y2} containing the time series data (i.e. it was created by a call to \code{\link[=rp]{rp()}}), the following options are available: "RS" (random shuffle), "RP" (randomised phases), "AAFT" (amplitude adjusted fourier transform). If no timeseries data is included, the columns will be shuffled.  NOTE: This is not a surrogate test, just 1 surrogate is created from \code{y1}.}

\item{returnOnlyObject}{Return the ggplot object only, do not draw the plot (default = \code{TRUE})}
}
\value{
A nice plot of the recurrence network
}
\description{
Plot (thresholded) distance matrix as a network
}
\seealso{
Other Distance matrix operations (recurrence network): 
\code{\link{di2bi}()},
\code{\link{di2ch}()},
\code{\link{di2we}()},
\code{\link{rn_phases}()},
\code{\link{rn_recSpec}()},
\code{\link{rn}()}
}
\concept{Distance matrix operations (recurrence network)}
