% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{rp}
\alias{rp}
\title{Create a Distance Matrix}
\usage{
rp(y1, y2 = NULL, emDim = 1, emLag = 1, emRad = NULL,
  to.ts = NULL, order.by = NULL, to.sparse = FALSE,
  method = "Euclidean", doPlot = FALSE, silent = TRUE, ...)
}
\arguments{
\item{y1}{A numeric vector or time series}

\item{y2}{A numeric vector or time series for cross recurrence}

\item{emDim}{The embedding dimensions}

\item{emLag}{The embedding lag}

\item{emRad}{The threshold (emRad) to apply to the distance matrix to create a binary matrix}

\item{to.ts}{Should \code{y1} and \code{y2} be converted to time series objects?}

\item{order.by}{If \code{to.ts = TRUE}, pass a vector of the same length as \code{y1} and \code{y2}. It will be used as the time index, if \code{NA} the vector indices will be used to represent time.}

\item{to.sparse}{Should sparse matrices be used?}

\item{method}{Distance measure to use. Any option that is valid for argument \code{method} of \code{\link[proxy]{dist}}. Type \code{proxy::pr_DB$get_entries()} to se a list of all the options. Common methods are: "Euclidean", "Manhattan", "Minkowski", "Chebysev" (or the same but shorter: "L2","L1","Lp" and "max" distance) (default = \code{"Euclidean"})}

\item{doPlot}{Plot the matrix by calling \code{\link{rp_plot}} with defult settings}

\item{silent}{Silent-ish mode}

\item{...}{Any paramters to pass to \code{\link{rp_plot}} if \code{doPlot = TRUE}}
}
\value{
A (Coss-) Recurrence matrix
}
\description{
Create a Distance Matrix
}
\seealso{
Other Distance matrix operations: \code{\link{bandReplace}},
  \code{\link{di2bi}}, \code{\link{di2we}},
  \code{\link{dist_hamming}}, \code{\link{rp_lineDist}},
  \code{\link{rp_nzdiags}}, \code{\link{rp_plot}},
  \code{\link{rp_size}}
}
\concept{Distance matrix operations}
