% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{rp_diagProfile}
\alias{rp_diagProfile}
\alias{crqa_diagprofile}
\title{Diagonal Recurrence Profile}
\usage{
rp_diagProfile(
  RM,
  diagWin = NULL,
  xname = "X-axis",
  yname = "Y-axis",
  theiler = 0,
  DLmin = 2,
  VLmin = 2,
  HLmin = 2,
  DLmax = length(Matrix::diag(RM)),
  VLmax = length(Matrix::diag(RM)),
  HLmax = length(Matrix::diag(RM)),
  doShuffle = FALSE,
  y1 = NA,
  y2 = NA,
  Nshuffle = 19,
  AUTO = NULL,
  chromatic = FALSE,
  matrices = FALSE,
  doPlot = TRUE
)
}
\arguments{
\item{RM}{A binary recurrence matrix}

\item{diagWin}{Window around the line of synchrony}

\item{xname}{Label for x-axis}

\item{yname}{Label for y-axis}

\item{theiler}{Use a \code{theiler} window around the main diagonal (Line of Identity/Synchronisation) to remove auto-correlations at short time-lags:
\itemize{
\item \code{0} will include the main diagonal in all RQA measure calculations.
\item \code{1} will remove the main diagonal from all RQA measure calculations.
\item \code{NA} (default), will check if the matrix is symmetrical , if so, it will remove the diagonal by setting \code{theiler = 0} (Line of Identity, Auto-RQA), if it is not symmetrical (Line of Synchronisation, Cross-RQA) it will set \code{theiler = 1}.
\item A value greater than \code{1} will remove that many diagonals around and including the diagonal from all RQA measure calculations. So \code{theiler = 2} means exclude \code{2} diagonals around the main diagonal, including the main diagonal itself: \verb{[-1,0,1]}.
If \code{theiler} is a numeric vector of \code{length(theiler) == 2} it is possible to exclude an asymmetrical window. The values are interpreted as end points in a sequence of diagonal ID's, e.g. \code{theiler = c(-1,5)} will exclude \verb{[-1,0,1,2,3,4,5]}. If \code{length(theiler) > 2}, the values will be considered individual diagonal ID's, e.g. \code{theiler = c(-3,-1,0,2,5)}, will exclude only those specific ID's. Also see the note.
}}

\item{DLmin}{Minimal diagonal line length (default = \code{2})}

\item{VLmin}{Minimal vertical line length (default = \code{2})}

\item{HLmin}{Minimal horizontal line length (default = \code{2})}

\item{DLmax}{Maximal diagonal line length (default = length of diagonal -1)}

\item{VLmax}{Maximal vertical line length (default = length of diagonal -1)}

\item{HLmax}{Maximal horizontal line length (default = length of diagonal -1)}

\item{doShuffle}{Should a shuffled baseline be calculated (default = \code{FALSE})}

\item{y1}{The original \code{y1} time series}

\item{y2}{The original \code{y2} time series}

\item{Nshuffle}{How many shuffled versions to make up the baseline? The default is \code{19}, which is the minimum for a one-sided surrogate test.}

\item{AUTO}{Auto-recurrence? (default = \code{FALSE})}

\item{chromatic}{Force chromatic RQA? (default = \code{FALSE})}

\item{matrices}{Return matrices? (default = \code{FALSE})}

\item{doPlot}{Plot}
}
\value{
A plot and/or the data for the plot
}
\description{
Diagonal Recurrence Profile
}
