% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{rp_plot}
\alias{rp_plot}
\title{Plot (thresholded) distance matrix as a recurrence plot}
\usage{
rp_plot(
  RM,
  plotDimensions = FALSE,
  plotDimensionLegend = FALSE,
  plotMeasures = FALSE,
  plotRadiusRRbar = TRUE,
  drawGrid = FALSE,
  drawDiagonal = TRUE,
  drawNA = FALSE,
  markEpochsLOI = NULL,
  radiusValue = NA,
  title = "",
  xlabel = "",
  ylabel = "",
  plotSurrogate = NA,
  returnOnlyObject = FALSE
)
}
\arguments{
\item{RM}{A distance matrix or recurrence matrix, preferably generated by function \link{rp} or \link{rn}.}

\item{plotDimensions}{Should the state vectors be plotted if they are available as attributes of RM (default = \code{TRUE})}

\item{plotDimensionLegend}{If \code{plotDimensions = TRUE} plot a simple legend (default = \code{FALSE})}

\item{plotMeasures}{Print common (C)RQA measures in the plot if the matrix is binary (default = \code{FALSE})}

\item{plotRadiusRRbar}{The \code{Radius-RR-bar} is a colour-bar guide plotted with an unthresholded distance matrix indicating a number of \code{RR} values one would get if a certain distance threshold were chosen (default = \code{TRUE})}

\item{drawGrid}{Draw a grid on the recurrence plot (default = \code{FALSE})}

\item{drawDiagonal}{One usually omits the main diagonal, the Line Of Incidence (LOI) from the calculation of Auto-RQA measures., it is however common to plot it. Set to \code{FALSE} to omit the LOI from an Auto-Recurrence Plot (default = \code{TRUE})}

\item{drawNA}{Draw \code{NA} values in the recurrence plot? If \code{FALSE} then \code{NA} values will be considered non-recurring (default = \code{FALSE})}

\item{markEpochsLOI}{Pass a factor whose levels indicate different epochs or phases in the time series and use the line of identity to represent the levels by different colours (default = \code{NULL})}

\item{radiusValue}{If \code{plotMeasures = TRUE} and RM is an unthresholded matrix, this value will be used to calculate recurrence measures. If \code{plotMeasures = TRUE} and RM is already a binary recurrence matrix, pass the radius that was used as a threshold to create the matrix for display purposes. If \code{plotMeasures = TRUE} and \code{radiusValue = NA}, function \code{est_radius()} will be called with default settings (find a radius that yields .05 recurrence rate). If \code{plotMeasures = FALSE} this setting will be ignored.}

\item{title}{A title for the plot}

\item{xlabel}{An x-axis label}

\item{ylabel}{An y-axis label}

\item{plotSurrogate}{Should a 2-panel comparison plot based on surrogate time series be added? If \code{RM} has attributes \code{y1} and \code{y2} containing the time series data (i.e. it was created by a call to \code{\link[=rp]{rp()}}), the following options are available: "RS" (random shuffle), "RP" (randomised phases), "AAFT" (amplitude adjusted fourier transform). If no timeseries data is included, the columns will be shuffled.  NOTE: This is not a surrogate test, just 1 surrogate is created from \code{y1}.}

\item{returnOnlyObject}{Return the ggplot object only, do not draw the plot (default = \code{TRUE})}
}
\value{
A nice plot of the recurrence matrix.
}
\description{
Plot (thresholded) distance matrix as a recurrence plot
}
\seealso{
Other Distance matrix operations (recurrence plot): 
\code{\link{bandReplace}()},
\code{\link{di2bi}()},
\code{\link{di2ch}()},
\code{\link{di2we}()},
\code{\link{dist_hamming}()},
\code{\link{rp_lineDist}()},
\code{\link{rp_nzdiags}()},
\code{\link{rp_size}()},
\code{\link{rp}()}
}
\concept{Distance matrix operations (recurrence plot)}
