% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{SDA}
\alias{SDA}
\title{Standardised Dispersion Analysis}
\usage{
SDA(x, front = FALSE, scaleS)
}
\arguments{
\item{x}{x}

\item{front}{front}

\item{scaleS}{If not \code{NA}, it should be a numeric vector listing the scales on which to evaluate the detrended fluctuations. Arguments \verb{scaleMax, scaleMin, scaleResolution} and \code{dataMin} will be ignored (default = \code{NA})}
}
\value{
output
}
\description{
Adaptation from discontinued package \code{fractal} by William Constantine and Donald Percival.
}
\keyword{internal}
