% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{SDA}
\alias{SDA}
\title{Standardised Dispersion Analysis}
\usage{
SDA(x, front = FALSE)
}
\arguments{
\item{x}{x}

\item{front}{front}
}
\value{
output
}
\description{
Adaptation from discontinued package \code{fractal} by William Constantine and Donald Percival.
}
\keyword{internal}
