% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{sliceTS}
\alias{sliceTS}
\title{Slice columns of a matrix in epochs}
\usage{
sliceTS(y, epochSz = 4)
}
\arguments{
\item{y}{A matrix with timeseries as columns}

\item{epochSz}{Epoch size}
}
\value{
A list with epochs
}
\description{
Slice columns of a matrix in epochs
}
