% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ssg_winnowing}
\alias{ssg_winnowing}
\title{ssg_winnowing}
\usage{
ssg_winnowing(durations, screeCut)
}
\arguments{
\item{durations}{durations frame}

\item{screeCut}{cutoff}
}
\value{
attractor frame
}
\description{
ssg_winnowing
}
