% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{symHelper}
\alias{symHelper}
\title{Correct for plateaus in symbolic series}
\usage{
symHelper(sym_label, sym_num)
}
\arguments{
\item{sym_label}{Labels}

\item{sym_num}{Numeric labels}
}
\value{
data frame
}
\description{
Correct for plateaus in symbolic series
}
\keyword{internal}
