% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_diff}
\alias{ts_diff}
\title{Derivative of time series}
\usage{
ts_diff(
  y,
  order = 1,
  addColumns = TRUE,
  keepDerivatives = FALSE,
  maskEdges = NULL,
  silent = TRUE
)
}
\arguments{
\item{y}{A timeseries object or numeric vector or a matrix in which columns are variables and rows are numeric values observed over time.}

\item{order}{How many times should the difference iteration be applied? (default = \code{1})}

\item{addColumns}{Should the derivative(s) be added to the input vector/matrix as columns? (default = \code{TRUE})}

\item{keepDerivatives}{If \code{TRUE} and \code{order > 1}, all derivatives from \code{1:order} will be returned as a matrix )default = \code{FALSE})}

\item{maskEdges}{Mask the values at the edges of the derivatives by any numeric type that is not \code{NULL} (default = \code{NULL})}

\item{silent}{Silent-ish mode}
}
\value{
Depending on the setting of \code{addColumns} and the object type passed as \code{y}, a vector of equal length as \code{y} iteratively differenced by \code{order} times; a matrix with derivatives, or a matrix with original(s) and derivative(s).
}
\description{
Iteratively differenced series up to \code{order}. The same length as the original series is recovered by calculating the mean of two vectors for each iteration: One with a duplicated first value and one with a duplicated last value.
}
\note{
The values at the edges of the derivatives represent endpoint averages and should be excluded from any subsequent analyses. Set argument \code{maskEdges} to a value of your choice.
}
\examples{

# Here's an interesting numeric vector
y<-c(1,1,2,3,5,8,13,21,34,55,89,144,233,377,610,987,1597,2584,4181,6765,10946,17711,28657,46368)

# Return the first order derivative as a vector
ts_diff(y=y,addColumns=FALSE)

# Return original and derivative as a matrix
plot(stats::ts(ts_diff(y=y, addColumns=TRUE)))

# Works on multivariate data objects with mixed variable types
df <- data.frame(x=letters, y=1:26, z=sin(1:26))

# Returns only derivatives of the numeric colunmns
ts_diff(y=df,addColumns=FALSE)

# Returns original data with derivatives of the numeric columns
ts_diff(y=df, order=4, addColumns=TRUE)

# Plot logistic S-curve and derivatives 1 to 3
S <- stats::plogis(seq(-5,5,.1))
plot(stats::ts(ts_diff(S, order=3, keepDerivatives = TRUE)))
abline(v=which(seq(-5,5,.1)==0), col= "red3", lwd=3)

# Plot again, but with masked edges
(maskEdge <- ts_diff(S, order=3, keepDerivatives = TRUE, maskEdges = NA))
plot(stats::ts(maskEdge))
abline(v=which(seq(-5,5,.1)==0), col= "red3", lwd=3)

}
\seealso{
Other Time series operations: 
\code{\link{ts_center}()},
\code{\link{ts_changeindex}()},
\code{\link{ts_checkfix}()},
\code{\link{ts_detrend}()},
\code{\link{ts_discrete}()},
\code{\link{ts_duration}()},
\code{\link{ts_embed}()},
\code{\link{ts_integrate}()},
\code{\link{ts_levels}()},
\code{\link{ts_peaks}()},
\code{\link{ts_permtest_block}()},
\code{\link{ts_permtest_transmat}()},
\code{\link{ts_rasterize}()},
\code{\link{ts_sd}()},
\code{\link{ts_slice}()},
\code{\link{ts_standardise}()},
\code{\link{ts_sumorder}()},
\code{\link{ts_symbolic}()},
\code{\link{ts_trimfill}()},
\code{\link{ts_windower}()}
}
\concept{Time series operations}
