% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_embed}
\alias{ts_embed}
\title{Delay embedding of a time series}
\usage{
ts_embed(y, emDim, emLag, returnOnlyIndices = FALSE, silent = TRUE)
}
\arguments{
\item{y}{Time series}

\item{emDim}{Embedding dimension}

\item{emLag}{Embedding lag}

\item{returnOnlyIndices}{Return only the index of y for each surrogate dimension, not the values (default = \code{FALSE})}

\item{silent}{Silent-ish mode}
}
\value{
The lag embedded time series
}
\description{
Create a state vector based on an embedding lag and a number of embedding dimanesions.
}
\seealso{
Other Time series operations: \code{\link{ts_center}},
  \code{\link{ts_detrend}}, \code{\link{ts_levels}},
  \code{\link{ts_peaks}}, \code{\link{ts_slice}},
  \code{\link{ts_standardise}}, \code{\link{ts_trimfill}}
}
\author{
Fred Hasselman
}
\concept{Time series operations}
