% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_embed}
\alias{ts_embed}
\title{Delay embedding of a time series}
\usage{
ts_embed(y, emDim, emLag, returnOnlyIndices = FALSE, silent = TRUE)
}
\arguments{
\item{y}{Time series}

\item{emDim}{Embedding dimension}

\item{emLag}{Embedding lag}

\item{returnOnlyIndices}{Return only the index of y for each surrogate dimension, not the values (default = \code{FALSE})}

\item{silent}{Silent-ish mode}
}
\value{
The lag embedded time series
}
\description{
Create a state vector based on an embedding lag and a number of embedding dimanesions.
}
\note{
If \code{emLag = 0}, the assumption is the columns in \code{y} represent the dimensions and \code{y} will be returned with attributes \code{emLag = 0} and \code{emDim = NCOL(y)}. If \code{emLag > 0} and \code{NCOL(y)>1} the first column of \code{y} will used for embedding and a warning will be triggered.
}
\seealso{
Other Time series operations: 
\code{\link{ts_center}()},
\code{\link{ts_changeindex}()},
\code{\link{ts_checkfix}()},
\code{\link{ts_detrend}()},
\code{\link{ts_diff}()},
\code{\link{ts_discrete}()},
\code{\link{ts_duration}()},
\code{\link{ts_integrate}()},
\code{\link{ts_levels}()},
\code{\link{ts_peaks}()},
\code{\link{ts_permtest_block}()},
\code{\link{ts_permtest_transmat}()},
\code{\link{ts_rasterize}()},
\code{\link{ts_sd}()},
\code{\link{ts_slice}()},
\code{\link{ts_standardise}()},
\code{\link{ts_sumorder}()},
\code{\link{ts_symbolic}()},
\code{\link{ts_trimfill}()},
\code{\link{ts_windower}()}

Other Time series operations: 
\code{\link{ts_center}()},
\code{\link{ts_changeindex}()},
\code{\link{ts_checkfix}()},
\code{\link{ts_detrend}()},
\code{\link{ts_diff}()},
\code{\link{ts_discrete}()},
\code{\link{ts_duration}()},
\code{\link{ts_integrate}()},
\code{\link{ts_levels}()},
\code{\link{ts_peaks}()},
\code{\link{ts_permtest_block}()},
\code{\link{ts_permtest_transmat}()},
\code{\link{ts_rasterize}()},
\code{\link{ts_sd}()},
\code{\link{ts_slice}()},
\code{\link{ts_standardise}()},
\code{\link{ts_sumorder}()},
\code{\link{ts_symbolic}()},
\code{\link{ts_trimfill}()},
\code{\link{ts_windower}()}
}
\author{
Fred Hasselman
}
\concept{Time series operations}
