% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_integrate}
\alias{ts_integrate}
\title{Create a timeseries profile}
\usage{
ts_integrate(y)
}
\arguments{
\item{y}{A 1D timeseries}
}
\description{
Create a timeseries profile
}
\examples{
y <- runif(1000,-3,3)
plot(ts(y))
y_i <- ts_integrate(y)
plot(ts(y_i))

}
