% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_integrate}
\alias{ts_integrate}
\title{Create a timeseries profile}
\usage{
ts_integrate(y)
}
\arguments{
\item{y}{A 1D timeseries}
}
\value{
The profile
}
\description{
Create a timeseries profile
}
\examples{
y <- runif(1000,-3,3)
plot(ts(y))
y_i <- ts_integrate(y)
plot(ts(y_i))

}
\seealso{
Other Time series operations: 
\code{\link{ts_center}()},
\code{\link{ts_changeindex}()},
\code{\link{ts_checkfix}()},
\code{\link{ts_detrend}()},
\code{\link{ts_diff}()},
\code{\link{ts_discrete}()},
\code{\link{ts_duration}()},
\code{\link{ts_embed}()},
\code{\link{ts_levels}()},
\code{\link{ts_peaks}()},
\code{\link{ts_permtest_block}()},
\code{\link{ts_permtest_transmat}()},
\code{\link{ts_rasterize}()},
\code{\link{ts_sd}()},
\code{\link{ts_slice}()},
\code{\link{ts_standardise}()},
\code{\link{ts_sumorder}()},
\code{\link{ts_symbolic}()},
\code{\link{ts_trimfill}()},
\code{\link{ts_windower}()}
}
\concept{Time series operations}
