% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_levels}
\alias{ts_levels}
\title{Detect levels in time series}
\usage{
ts_levels(y, minLevelChange = 10,
  minLevelDuration = round(minLevelChange/3), changeSensitivity = 0.01,
  maxLevels = 30, method = c("anova", "poisson", "class", "exp")[1])
}
\arguments{
\item{y}{A time series of numeric vector}

\item{minLevelChange}{A factor indicating how much the signal should change before is is consider a new level (default = \code{20})}

\item{minLevelDuration}{Minimum durtion (length) of a level (default = \code{round(minLevelChange/3)})}

\item{changeSensitivity}{A factor indicating a criterion}

\item{maxLevels}{Maximum number of levels in one series (default = \code{30})}

\item{method}{The partitioning method to use, see the manual pages of \link[rpart]{rpart} for details.}
}
\value{
A list object with fields \code{tree} and \code{pred}
}
\description{
Use recursive partitioning function (\link[rpart]{rpart} from \code{rpart} to perform a 'classification' of relatively stable levels in a timeseries.
}
\seealso{
Other Time series operations: \code{\link{ts_center}},
  \code{\link{ts_detrend}}, \code{\link{ts_embed}},
  \code{\link{ts_peaks}}, \code{\link{ts_slice}},
  \code{\link{ts_standardise}}, \code{\link{ts_trimfill}}
}
\author{
Fred Hasselman
}
