% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlRtsa_SOURCE.R
\name{ts_permtest_block}
\alias{ts_permtest_block}
\title{Permutation Test: Block Randomisation}
\usage{
ts_permtest_block(y1, y2, Nperms = 99, sim = "geom", l = 3,
  alpha = 0.05, returnBootObject = FALSE, ...)
}
\arguments{
\item{y1}{Time series 1}

\item{y2}{Time series 2}

\item{Nperms}{Number of permutations (default = \code{99})}

\item{sim}{Value passed to the \code{sim} argument of \code{\link[boot]{tsboot}} valid options are: \code{"model","fixed","geom","scramble"} (default = \code{"geom"})}

\item{l}{Block sizes to use, see \code{\link[boot]{tsboot}} for details (default = \code{3})}

\item{alpha}{Alpha level for deciding significance  (default = \code{0.05})}

\item{returnBootObject}{Return the \code{boot} object (default = \code{FALSE})}

\item{...}{Other arguments passed to function \code{\link[boot]{tsboot}}}
}
\value{
A data frame with the difference time series and variables indicating N and significance.
}
\description{
Permutation Test: Block Randomisation
}
