% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_permtest_block}
\alias{ts_permtest_block}
\title{Permutation Test: Block Randomisation}
\usage{
ts_permtest_block(
  y1,
  y2 = NULL,
  targetValue = 0,
  Nperms = 19,
  sim = "geom",
  l = 3,
  alpha = 0.05,
  returnBootObject = FALSE,
  ...
)
}
\arguments{
\item{y1}{Time series 1. The goal of the permutation test will be to decide whether the difference \code{y1-targetValue != 0} for each time point, given \code{alpha}.}

\item{y2}{An optional second time series. If this timeseries is provided then the goal of the permutation test will be the to decide wether the difference \code{y2-y1 != targetValue} for each time point, given \code{alpha}.}

\item{targetValue}{The target value for the permutation test. If \code{NULL}, the function will return a data frame with the block randomised surrogates  columns (default = \code{0})}

\item{Nperms}{Number of permutations (default = \code{19})}

\item{sim}{Value passed to the \code{sim} argument of \code{\link[boot:tsboot]{boot::tsboot()}} valid options are: \verb{"model","fixed","geom","scramble"} (default = \code{"geom"})}

\item{l}{Block sizes to use, see \code{\link[boot:tsboot]{boot::tsboot()}} for details (default = \code{3})}

\item{alpha}{Alpha level for deciding significance  (default = \code{0.05})}

\item{returnBootObject}{Return the \code{boot} object (default = \code{FALSE})}

\item{...}{Other arguments passed to function \code{\link[boot:tsboot]{boot::tsboot()}}}
}
\value{
A data frame with the difference time series and variables indicating N and significance.
}
\description{
Use block randomistion to get a permutation test evaluation of the deviation of an observed value at each time point from a target value. To do block permutation without any tests, pass \code{NULL} for argument \code{targetValue}.
}
\examples{

\donttest{set.seed(4321)
y1 <- rnorm(5000)
y2 <- y1-(mean(y1)+rnorm(1))

ts_permtest_block(y1 = y1, y2 = y2)}

}
\seealso{
Other Time series operations: 
\code{\link{ts_center}()},
\code{\link{ts_changeindex}()},
\code{\link{ts_checkfix}()},
\code{\link{ts_detrend}()},
\code{\link{ts_diff}()},
\code{\link{ts_discrete}()},
\code{\link{ts_duration}()},
\code{\link{ts_embed}()},
\code{\link{ts_integrate}()},
\code{\link{ts_levels}()},
\code{\link{ts_peaks}()},
\code{\link{ts_permtest_transmat}()},
\code{\link{ts_rasterize}()},
\code{\link{ts_sd}()},
\code{\link{ts_slice}()},
\code{\link{ts_standardise}()},
\code{\link{ts_sumorder}()},
\code{\link{ts_symbolic}()},
\code{\link{ts_trimfill}()},
\code{\link{ts_windower}()}
}
\concept{Time series operations}
