% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_rasterize}
\alias{ts_rasterize}
\title{Turn a 1D time series vector into a 2D curve}
\usage{
ts_rasterize(y, unitSquare = FALSE, toSparse = TRUE, resolution = 2)
}
\arguments{
\item{y}{A 1D time series object or numeric vector.}

\item{unitSquare}{Convert the series to a unit square? (default = \code{FALSE})}

\item{toSparse}{Convert to sparse Matrix (default = \code{FALSE})}

\item{resolution}{Factor by which dimensions will be multiplied (default = \code{2})}
}
\value{
A (sparse) matrix representing the time series as a curve in 2D space
}
\description{
Turn a 1D time series vector into a 2D curve
}
\examples{

\donttest{
y <- rnorm(100)
plot(ts(y))

y_img <- ts_rasterize(y)
image(y_img,col=c("white","black"))}


}
\seealso{
Other Time series operations: 
\code{\link{ts_center}()},
\code{\link{ts_changeindex}()},
\code{\link{ts_checkfix}()},
\code{\link{ts_detrend}()},
\code{\link{ts_diff}()},
\code{\link{ts_discrete}()},
\code{\link{ts_duration}()},
\code{\link{ts_embed}()},
\code{\link{ts_integrate}()},
\code{\link{ts_levels}()},
\code{\link{ts_peaks}()},
\code{\link{ts_permtest_block}()},
\code{\link{ts_permtest_transmat}()},
\code{\link{ts_sd}()},
\code{\link{ts_slice}()},
\code{\link{ts_standardise}()},
\code{\link{ts_sumorder}()},
\code{\link{ts_symbolic}()},
\code{\link{ts_trimfill}()},
\code{\link{ts_windower}()}
}
\concept{Time series operations}
