% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_sd}
\alias{ts_sd}
\title{Standard Deviation estimates}
\usage{
ts_sd(y, na.rm = TRUE, type = c("Bessel", "unadjusted")[1], silent = TRUE)
}
\arguments{
\item{y}{Time series or numeric vector}

\item{na.rm}{Remove missing values before calculations}

\item{type}{Apply Bessel's correction (divide by N-1) or return unadjusted SD (divide by N)}

\item{silent}{Silent-ish mode (default = \code{TRUE})}
}
\value{
Standard deviation of \code{y}
}
\description{
Calculates the population estimate of the standard deviation, or the unadjusted standard deviation.
}
\seealso{
Other Time series operations: 
\code{\link{ts_center}()},
\code{\link{ts_changeindex}()},
\code{\link{ts_checkfix}()},
\code{\link{ts_detrend}()},
\code{\link{ts_diff}()},
\code{\link{ts_discrete}()},
\code{\link{ts_duration}()},
\code{\link{ts_embed}()},
\code{\link{ts_integrate}()},
\code{\link{ts_levels}()},
\code{\link{ts_peaks}()},
\code{\link{ts_permtest_block}()},
\code{\link{ts_permtest_transmat}()},
\code{\link{ts_rasterize}()},
\code{\link{ts_slice}()},
\code{\link{ts_standardise}()},
\code{\link{ts_sumorder}()},
\code{\link{ts_symbolic}()},
\code{\link{ts_trimfill}()},
\code{\link{ts_windower}()}
}
\concept{Time series operations}
