% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_standardise}
\alias{ts_standardise}
\title{Standardise a vector}
\usage{
ts_standardise(
  y,
  na.rm = TRUE,
  keepNAvalues = TRUE,
  type = c("mean.sd", "median.mad")[1],
  adjustN = TRUE
)
}
\arguments{
\item{y}{A time series or numeric vector}

\item{na.rm}{Set the \code{na.rm} field}

\item{keepNAvalues}{If \code{na.rm = TRUE} and \code{keepNAvalues = TRUE}, any \code{NA} values in \code{y} will be re-inserted after transformation.}

\item{type}{Center on the \code{"mean"} and divide by \code{sd} (default), or center on \code{"median"} and divide by \code{mad}}

\item{adjustN}{If \code{TRUE}, apply Bessel's correction (divide by \code{N-1}) or return the unadjusted \code{SD} (divide by \code{N}) (default = \code{TRUE})}
}
\value{
A standardised vector
}
\description{
Standardise a vector
}
\seealso{
Other Time series operations: 
\code{\link{ts_center}()},
\code{\link{ts_changeindex}()},
\code{\link{ts_checkfix}()},
\code{\link{ts_detrend}()},
\code{\link{ts_diff}()},
\code{\link{ts_discrete}()},
\code{\link{ts_duration}()},
\code{\link{ts_embed}()},
\code{\link{ts_integrate}()},
\code{\link{ts_levels}()},
\code{\link{ts_peaks}()},
\code{\link{ts_permtest_block}()},
\code{\link{ts_permtest_transmat}()},
\code{\link{ts_rasterize}()},
\code{\link{ts_sd}()},
\code{\link{ts_slice}()},
\code{\link{ts_sumorder}()},
\code{\link{ts_symbolic}()},
\code{\link{ts_trimfill}()},
\code{\link{ts_windower}()}
}
\author{
Fred Hasselman
}
\concept{Time series operations}
