% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_symbolic}
\alias{ts_symbolic}
\title{Symbolic representation}
\usage{
ts_symbolic(y, keepNA = TRUE, usePlateaus = FALSE, doPlot = FALSE)
}
\arguments{
\item{y}{Numeric vector or matrix to be discretised. Will return attributes with labels for each column in the matrix.}

\item{keepNA}{If \code{TRUE}, any \code{NA} values will first be removed and later re-inserted into the discretised time series. (default = \code{TRUE})}

\item{usePlateaus}{Give consecutive \code{"same"} values after \code{"peak"} or \code{"trough"} a \code{"peak"} or \code{"trough"} label instrad of \code{"same"}. (default = \code{FALSE})}

\item{doPlot}{Create a plot of the symbolized series. (default = \verb{FALSE)})}
}
\value{
A symbolic version of \code{y}
}
\description{
Return a discrete representation of \code{y} by transforming it into an unordered categorical variable which indicates whether a value went up, down, or remained the same relative to the previous value.
}
\seealso{
Other Time series operations: 
\code{\link{ts_center}()},
\code{\link{ts_changeindex}()},
\code{\link{ts_checkfix}()},
\code{\link{ts_detrend}()},
\code{\link{ts_diff}()},
\code{\link{ts_discrete}()},
\code{\link{ts_duration}()},
\code{\link{ts_embed}()},
\code{\link{ts_integrate}()},
\code{\link{ts_levels}()},
\code{\link{ts_peaks}()},
\code{\link{ts_permtest_block}()},
\code{\link{ts_permtest_transmat}()},
\code{\link{ts_rasterize}()},
\code{\link{ts_sd}()},
\code{\link{ts_slice}()},
\code{\link{ts_standardise}()},
\code{\link{ts_sumorder}()},
\code{\link{ts_trimfill}()},
\code{\link{ts_windower}()}
}
\concept{Time series operations}
