% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts_symbolic}
\alias{ts_symbolic}
\title{Symbolic representation}
\usage{
ts_symbolic(y, keepNA = TRUE, usePlateaus = FALSE, doPlot = FALSE)
}
\arguments{
\item{y}{Numeric vector or time series to be discretised.}

\item{keepNA}{If \code{TRUE}, any \code{NA} values will first be removed and later re-inserted into the discretised time series.}

\item{usePlateaus}{Treat consequative "same" values after "peak" or "trough" as a "peak"/"trough".}

\item{doPlot}{Create a plot of the symbolized series.}
}
\value{
A discretised version of \code{y}
}
\description{
Return a discrete representation of \code{y} by binning the observed values and returning the transfer probabilities.
}
\seealso{
Other Time series operations: 
\code{\link{ts_center}()},
\code{\link{ts_changeindex}()},
\code{\link{ts_checkfix}()},
\code{\link{ts_detrend}()},
\code{\link{ts_diff}()},
\code{\link{ts_discrete}()},
\code{\link{ts_duration}()},
\code{\link{ts_embed}()},
\code{\link{ts_integrate}()},
\code{\link{ts_levels}()},
\code{\link{ts_peaks}()},
\code{\link{ts_permtest_block}()},
\code{\link{ts_permtest_transmat}()},
\code{\link{ts_rasterize}()},
\code{\link{ts_sd}()},
\code{\link{ts_slice}()},
\code{\link{ts_standardise}()},
\code{\link{ts_sumorder}()},
\code{\link{ts_trimfill}()},
\code{\link{ts_windower}()}
}
\concept{Time series operations}
